+++ paper.bib
@@ -20,7 +20,7 @@
   volume   = {484},
   pages    = {29--41},
   issn     = {0168-583X},
-  doi      = {https://doi.org/10.1016/j.nimb.2020.10.005},
+  doi      = {10.1016/j.nimb.2020.10.005},
   url      = {https://www.sciencedirect.com/science/article/pii/S0168583X20304225},
   abstract = {Induced radioactivity is critical in a variety of fields including maintenance and decommissioning of accelerators and reactors, nuclear waste management, and radionuclide production for medical and industrial applications. Serving this purpose, DCHAIN-PHITS is the time-dependent radionuclide production, buildup, burnup, and decay code which is coupled to and distributed with the PHITS general purpose Monte Carlo particle transport code. While the core of the DCHAIN code was first written over forty years ago, the DCHAIN-PHITS incarnation of the code represents its most substantial update in nearly twenty years. This work provides an overview of the function of DCHAIN-PHITS and the new additions to the code. These additions include updated decay and cross section nuclear data libraries, statistical uncertainty propagation, support for tetrahedral and three-dimensional grid mesh geometries, reaction tracking capabilities, and general performance improvements.},
   keywords = {Activation, Decay, Burnup, Induced activity, PHITS},
