### 2025-01-30 f26751c
function to merge dumps (mainly for MPI dumps in dir mode)
