+++ PHITS_tools.py
@@ -46,6 +46,7 @@ functions return the data objects they produce for your own further analyses.
 - `rebinner`                        : rebin a set of y-data to a new x-binning structure (edges need not necessarily be preserved)
 - `fetch_MC_material`               : returns a string of a formatted material for PHITS or MCNP (mostly those in [PNNL-15870 Rev. 1](https://www.osti.gov/biblio/1023125))
 - `ICRP116_effective_dose_coeff`    : returns effective dose conversion coefficient of a mono-energetic particle of some species and some geometry; does coefficients are those in [ICRP 116](https://doi.org/10.1016/j.icrp.2011.10.001)
+- `merge_dump_file_pickles`         : merge multiple dump file outputs into a single file (useful for dumps in MPI runs)
 - `ZZZAAAM_to_nuclide_plain_str`    : returns a nuclide plaintext string for a given "ZZZAAAM" number (1000Z+10A+M)
 - `nuclide_plain_str_to_latex_str`  : convert a plaintext string for a nuclide to a LaTeX formatted raw string
 - `Element_Z_to_Sym`                : return an elemental symbol string given its proton number Z
@@ -662,10 +663,16 @@ def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_s
     else:
         compression_file_extension = ''
 
+    path_to_dump_file = Path(path_to_dump_file)
+    dump_file_suffixes = path_to_dump_file.suffixes
+    MPI_subdump_num_str = ''
+    if len(dump_file_suffixes) > 1 and is_number(dump_file_suffixes[-1][1:]) and '_dmp' in path_to_dump_file.stem:  # MPI dump found
+        MPI_subdump_num_str = dump_file_suffixes[-1]  #[1:]
+
     if save_namedtuple_list:
         import lzma
         path_to_dump_file = Path(path_to_dump_file)
-        pickle_path = Path(path_to_dump_file.parent, path_to_dump_file.stem + '_namedtuple_list.pickle' + compression_file_extension)
+        pickle_path = Path(path_to_dump_file.parent, path_to_dump_file.stem + MPI_subdump_num_str + '_namedtuple_list.pickle' + compression_file_extension)
         num_records = len(records_list)
         record_type = [(ordered_record_entries_list[i], dump_col_dtypes[i]) for i in range(len(ordered_record_entries_list))]
         records_np_array = np.array(records_list, dtype=record_type)
@@ -683,7 +690,7 @@ def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_s
         records_df = pd.DataFrame(records_list, columns=Record._fields)
         if save_Pandas_dataframe:
             path_to_dump_file= Path(path_to_dump_file)
-            pickle_path = Path(path_to_dump_file.parent, path_to_dump_file.stem + '_Pandas_df.pickle' + compression_file_extension)
+            pickle_path = Path(path_to_dump_file.parent, path_to_dump_file.stem + MPI_subdump_num_str + '_Pandas_df.pickle' + compression_file_extension)
             records_df.to_pickle(pickle_path)
             #with open(pickle_path, 'wb') as handle:
             #    pickle.dump(records_df, handle, protocol=pickle.HIGHEST_PROTOCOL)
@@ -709,7 +716,8 @@ def parse_all_tally_output_in_dir(tally_output_dirpath, output_file_suffix = '.o
                                   dump_data_number=None , dump_data_sequence=None,
                                   dump_return_directional_info=False, dump_use_degrees=False,
                                   dump_max_entries_read=None,
-                                  dump_save_namedtuple_list=True, dump_save_Pandas_dataframe=True
[…]
