### 2025-01-30 e542807
fixed CLI issue when passed directory instead of file
