### 2025-01-30 5b0ff0f
handling of Fortran write errors in rare PHITS outputs
