### 2025-01-30 0413ffa
added LZMA option for standard tally outputs
