+++ PHITS_tools.py
@@ -132,7 +132,7 @@ if in_debug_mode:
 
 
 def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calculate_absolute_errors = True,
-                            save_output_pickle = True, prefer_reading_existing_pickle = False):
+                            save_output_pickle=True, prefer_reading_existing_pickle=False, compress_pickle_with_lzma=False):
     '''
     Description:
         Parse any PHITS tally output file, returning tally metadata and an array of its values (and optionally
@@ -165,6 +165,16 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                       seeks to generate already exists.  If `False` (default behavior), this function will parse the PHITS
                       output files as usual and overwrite the existing pickle file.  If `True`, this function will instead
                       simply just read the existing found pickle file and return its stored `tally_output` contents. (D=`False`)
+       - `compress_pickle_with_lzma` = (optional, D=`False`; requires `save_output_pickle=True`) Boolean designating 
+                      whether the pickle file to be saved will be compressed with 
+                      [LZMA compression](https://docs.python.org/3/library/lzma.html) (included in
+                      the baseline [Python standard library](https://docs.python.org/3/library/index.html)); if so, the 
+                      pickle file's extension will be `'.pickle.xz'` instead of just `'.pickle'`.
+                      A *.pickle.xz file can then be opened (after importing `pickle` and `lzma`) as:  
+                      `with lzma.open(path_to_picklexz_file, 'rb') as file: tally_output = pickle.load(file)`.  
+                      For most "normal" tallies, the pickle file sizes are likely to be small enough to not warrant 
+                      compression, but LZMA compression can reduce the file size by several orders of magnitude, 
+                      great for output from massive tallies.
 
     Output:
         - `tally_output` = a dictionary object with the below keys and values:
@@ -315,11 +325,18 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
     '''
     tally_output_filepath = Path(tally_output_filepath)
     pickle_filepath = Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pickle')
-    if prefer_reading_existing_pickle and os.path.isfile(pickle_filepath):
+    picklexz_filepath = Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pickle.xz')
+    if prefer_reading_existing_pickle and (os.path.isfile(pickle_filepath) or os.path.isfile(picklexz_filepath)):
         import pickle
-        print('Reading found pickle file: ', pickle_filepath)
-        with open(pickle_filepath, 'rb') as handle:
-            tally_output = pickle.load(handle)
+        if os.path.isfile(pickle_filepath):
+            print('Reading found pickle file: ', pickle_filepath)
+            with open(pickle_filepath, 'rb') as handle:
+                tally_output = pickle.load(handle)
+        else:
+            print('Reading found pickle file: ', picklexz_filepath)
+            import lzma
+            with lzma.open(picklexz_filepath, 'rb') as handle:
+                tally_output = pickle.load(handle)
         return tally_output
 
     # main toggled settings
@@ -467,11 +484,20 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
 
     if save_output_pickle:
         import pickle
-        path_to_pickle_file = Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pickle')
+        if compress_pickle_with_lzma:
+            import lzma
+            compression_file_extension = '.xz'
+        else:
+            compression_file_extension = ''
[…]
