### 2025-02-03 236796b
added functionality to split large binary dump files
