+++ PHITS_tools.py
@@ -507,7 +507,8 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
 def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_sequence=None, return_directional_info=False,
                           use_degrees=False,max_entries_read=None,return_namedtuple_list=True,
                           return_Pandas_dataframe=True, save_namedtuple_list=False, save_Pandas_dataframe=False,
-                          compress_pickles_with_lzma=True):
+                          compress_pickles_with_lzma=True, prefer_reading_existing_pickle=False,
+                          split_binary_dumps_over_X_GB=20, merge_split_dump_handling=0):
     '''
     Description:
         Parses the dump file of a [T-Cross], [T-Product], or [T-Time] tally generated by PHITS, in ASCII or binary format.
@@ -545,6 +546,13 @@ def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_s
                  of the dump file to be read.  By default, all records in the dump file are read.
         - `return_namedtuple_list` = (optional, D=`True`) Boolean designating whether `dump_data_list` is returned.
         - `return_Pandas_dataframe` = (optional, D=`True`) Boolean designating whether `dump_data_frame` is returned.
+        - `prefer_reading_existing_pickle` = (optional, D=`False`) A Boolean determining what this function does if the 
+                 pickle file(s) this function seeks to generate already exist.  If `False` (default behavior), 
+                 this function will parse the PHITS dump files as usual and overwrite the existing pickle file(s). 
+                 If `True`, this function will instead simply just read the existing found pickle file(s) and return the stored contents. 
+                 Note that it can also be used for creating LZMA compressed/decompressed versions of the pickle file(s) 
+                 if the file is set to be saved and the found existing file's compression (or lack thereof) contrasts
+                 with the current setting of `compress_pickles_with_lzma`.
         - `compress_pickles_with_lzma` = (optional, D=`True`) Boolean designating whether the pickle files to be saved of
                  the namedtuple list (if `save_namedtuple_list = True`) and/or the Pandas DataFrame (if `save_Pandas_dataframe = True`)
                  will be compressed with [LZMA compression](https://docs.python.org/3/library/lzma.html) (included within
@@ -568,8 +576,17 @@ def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_s
                 save the list of namedtuples, also using the LZMA compression in both cases._
         - `save_Pandas_dataframe` = (optional, D=`False`) Boolean designating whether `dump_data_frame` is saved to a pickle
                 file (via the Pandas `.to_pickle()` method), which will by default
-                be compressed with LZMA (built-in with Python; see the `compress_pickles_with_lzma` option) .
-
+                be compressed with LZMA (built-in with Python; see the `compress_pickles_with_lzma` option). This pickle 
+                can then be accessed via the [`pandas.read_pickle(path_to_DF_pickle)`](https://pandas.pydata.org/docs/reference/api/pandas.read_pickle.html) function.
+        - `split_binary_dumps_over_X_GB` = (optional, D=`20` GB) an integer/float specifying the number of gigabytes over 
+                which a _binary_ dump file will be split into chunks of this size for processing&Dagger;.
+        - `merge_split_dump_handling` = (optional, D=`0`) For instances where binary dump files are to be split, 
+                described above for `split_binary_dumps_over_X_GB` and the the note below&Dagger;, input one of the following 
[…]
