### 2025-02-04 24d66e3
improved documentation formatting for main functions
