+++ PHITS_tools.py
@@ -149,23 +149,23 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
     Inputs:
        (required)
 
-        - `tally_output_filepath` = string or Path object denoting the path to the tally output file to be parsed
+        - **`tally_output_filepath`** = string or Path object denoting the path to the tally output file to be parsed
 
     Inputs:
        (optional)
 
-       - `make_PandasDF` = A Boolean determining whether a Pandas dataframe of the tally data array will be made (D=`True`)
-       - `calculate_absolute_errors` = A Boolean determining whether the absolute uncertainty of each tally output value
+       - **`make_PandasDF`** = A Boolean denoting whether a Pandas dataframe of the tally data array will be made (D=`True`)
+       - **`calculate_absolute_errors`** = A Boolean determining whether the absolute uncertainty of each tally output value
                       is to be calculated (simply as the product of the value and relative error); if `False`, the final
                       dimension of `tally_data`, `ierr`, will be of length-2 rather than length-3 (D=`True`)
-       - `save_output_pickle` = A Boolean determining whether the `tally_output` dictionary object is saved as a pickle file;
+       - **`save_output_pickle`** = A Boolean determining whether the `tally_output` dictionary object is saved as a pickle file;
                       if `True`, the file will be saved with the same path and name as the provided PHITS tally output file
                       but with the .pickle extension. (D=`True`)
-       - `prefer_reading_existing_pickle` = A Boolean determining what this function does if the pickle file this function
+       - **`prefer_reading_existing_pickle`** = A Boolean determining what this function does if the pickle file this function
                       seeks to generate already exists.  If `False` (default behavior), this function will parse the PHITS
                       output files as usual and overwrite the existing pickle file.  If `True`, this function will instead
                       simply just read the existing found pickle file and return its stored `tally_output` contents. (D=`False`)
-       - `compress_pickle_with_lzma` = (optional, D=`False`; requires `save_output_pickle=True`) Boolean designating 
+       - **`compress_pickle_with_lzma`** = (optional, D=`False`; requires `save_output_pickle=True`) Boolean designating 
                       whether the pickle file to be saved will be compressed with 
                       [LZMA compression](https://docs.python.org/3/library/lzma.html) (included in
                       the baseline [Python standard library](https://docs.python.org/3/library/index.html)); if so, the 
@@ -177,7 +177,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                       great for output from massive tallies.
 
     Output:
-        - `tally_output` = a dictionary object with the below keys and values:
+        - **`tally_output`** = a dictionary object with the below keys and values:
             - `'tally_data'` = a 10-dimensional NumPy array containing all tally results, explained in more detail below
             - `'tally_metadata'` = a dictionary/Munch&dagger; object with various data extracted from the tally output file, such as axis binning and units
             - `'tally_dataframe'` = (optionally included if setting `make_PandasDF = True`) a Pandas dataframe version of `tally_data`
@@ -514,6 +514,7 @@ def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_s
         Parses the dump file of a [T-Cross], [T-Product], or [T-Time] tally generated by PHITS, in ASCII or binary format.
 
     Dependencies:
+        - `import numpy as np`
         - `from collections import namedtuple`
         - `from scipy.io import FortranFile`
         - `import pandas as pd` (if `return_Pandas_dataframe = True`)
@@ -521,11 +522,11 @@ def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_s
     Inputs:
        (required)
 
-        - `path_to_dump_file` = string or Path object denoting the path to the dump tally output file to be parsed
[…]
