### 2025-02-05 971a89a
added more info on CLI usage
