### 2025-02-05 7566820
added CLI option info to docs
