+++ PHITS_tools.py
@@ -24,6 +24,7 @@ There are three main ways one can use this Python module:
           containing multiple files to be parsed) as the required argument.
           Execute `python PHITS_tools.py --help` to see all of the different options that can be used with this module
           to parse standard or dump PHITS output files (individually and directories containing them) via the CLI.
+          How the CLI options explicity translate to the functions documented here is covered further below.
  3. As a **graphical user interface (GUI)**
       - When the module is executed without any additional arguments, `python PHITS_tools.py`, (or with the `--GUI` or `-g` flag in the CLI)
           a GUI will be launched to step you through selecting what "mode" you would like to run PHITS Tools in (`STANDARD`, `DUMP`, or `DIRECTORY`),
@@ -70,6 +71,42 @@ functions return the data objects they produce for your own further analyses.
 - `data_row_to_num_list`            : extract numeric values from a line in the tally content section
 - `calculate_tally_absolute_errors` : calculate absolute uncertainties from read values and relative errors
 - `build_tally_Pandas_dataframe`    : make Pandas dataframe from the main results NumPy array and the metadata
+
+
+### **CLI options**
+
+Essentially, the CLI serves to interface with the core three functions of PHITS Tools: `parse_tally_output_file`,
+ `parse_tally_dump_file`, and `parse_all_tally_output_in_dir`. 
+ The required `file` argument is checked to see if it is a directory or a file, and, if the latter, whether the `-d` 
+ option is used denoting a dump output file, otherwise defaulting to assuming it is a PHITS standard tally output file; 
+ then `file` and the relevant settings are sent to the corresponding main function. 
+ Explicitly, inclusion of the various CLI options have the following effects on the main functions' arguments and settings:
+
+- Affecting all functions
+      - `-skip` sets `prefer_reading_existing_pickle = True` (`False` if excluded)
+- `parse_tally_output_file` (and passed to it via `parse_all_tally_output_in_dir`)
+      - `-np` sets `make_PandasDF = False` (`True` if excluded)
+      - `-na` sets `calculate_absolute_errors = False` (`True` if excluded)
+      - `-lzma` sets `compress_pickle_with_lzma = True` (`False` if excluded)
+- `parse_tally_dump_file` (and passed to it via `parse_all_tally_output_in_dir`)
+      - `-dvals` passes the provided sequence of values to `dump_data_sequence` (`None` if excluded)
+      - `-dbin` specifies that the file is binary (`dump_data_number=len(dump_data_sequence)` and *is positive*)
+      - `-dnmax` passes its value to `max_entries_read` (`None` if excluded)
+      - `-ddir` sets `return_directional_info = True` (`False` if excluded)
+      - `-ddeg` sets `use_degrees = True` (`False` if excluded)
+      - `-dnsl` sets `save_namedtuple_list = False` (`True` if excluded)
+      - `-dnsp` sets `save_Pandas_dataframe = False` (`True` if excluded)
+      - `-dmaxGB` passes its value to `split_binary_dumps_over_X_GB` (`20` GB if excluded)
+      - `-dsplit` passes its value to `merge_split_dump_handling` (`0` if excluded)
+- `parse_all_tally_output_in_dir` exclusively
+      - `-r` sets `include_subdirectories = True` (`False` if excluded)
+      - `-fpre` passes its value to `output_file_prefix` (`''` if excluded)
+      - `-fsuf` passes its value to `output_file_suffix` (`'.out'` if excluded)
+      - `-fstr` passes its value to `output_file_required_string` (`''` if excluded)
+      - `-d` sets `include_dump_files = True` (`False` if excluded)
[…]
