### 2025-02-05 9a3d459
update CLI info in docs
