### 2025-02-05 c6f34ea
improved CLI info in docs
