+++ PHITS_tools.py
@@ -83,12 +83,14 @@ Essentially, the CLI serves to interface with the core three functions of PHITS 
  Explicitly, inclusion of the various CLI options have the following effects on the main functions' arguments and settings:
 
 - Affecting all functions
+      - `file` is passed to `tally_output_filepath`, `path_to_dump_file`, or `tally_output_dirpath`
       - `-skip` sets `prefer_reading_existing_pickle = True` (`False` if excluded)
 - `parse_tally_output_file` (and passed to it via `parse_all_tally_output_in_dir`)
       - `-np` sets `make_PandasDF = False` (`True` if excluded)
       - `-na` sets `calculate_absolute_errors = False` (`True` if excluded)
       - `-lzma` sets `compress_pickle_with_lzma = True` (`False` if excluded)
 - `parse_tally_dump_file` (and passed to it via `parse_all_tally_output_in_dir`)
+      - `-d` tells the CLI that `file` should be processed as a dump file (if it's not a directory)
       - `-dvals` passes the provided sequence of values to `dump_data_sequence` (`None` if excluded)
       - `-dbin` specifies that the file is binary (that `dump_data_number=len(dump_data_sequence)` and *is positive*)
       - `-dnmax` passes its value to `max_entries_read` (`None` if excluded)
