### 2025-03-07 3edddc3
added Jupyter notebook version of the example
