### 2025-03-10 d5acd64
added more documentation on metadata dictionary
