+++ PHITS_tools.py
@@ -227,13 +227,20 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
        will just be a standard dictionary object.  If, for whatever reason, you have Munch installed but do not wish for 
        `tally_metadata` to be a Munch object, then in the first line of code for the `parse_tally_header()` function, 
        set `prefer_to_munch_meta_dict = False`._
-
+    
+       -----
+    
+       -----
+    
     Notes:
+    
+       **The `'tally_data'` 10-dimensional NumPy array**
 
        Many quantities can be scored across the various tallies in the PHITS code.  This function outputs a "universal"
-       array `tally_data` that can accomodate all of the different scoring geometry meshes, physical quantities with
-       assigned meshes, and output axes provided within PHITS.  This is achieved with a 10-dimensional array accessible as
-
+       array `tally_data` (`tally_data = tally_output['tally_metadata']`) that can accomodate all of the different 
+       scoring geometry meshes, physical quantities with assigned meshes, and output axes provided within PHITS. 
+       This is achieved with a 10-dimensional array accessible as
+        
        `tally_data[ ir, iy, iz, ie, it, ia, il, ip, ic, ierr ]`, with indices explained below:
 
        Tally data indices and corresponding mesh/axis:
@@ -250,16 +257,73 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
         - `9` | `ierr = 0/1/2`, Value / relative uncertainty / absolute uncertainty (expanded to `3/4/5`, or `2/3` if
         `calculate_absolute_errors = False`, for [T-Cross] `mesh=r-z` with `enclos=0` case; see notes further below)
 
-       -----
 
        By default, all array dimensions are length-1 (except `ierr`, which is length-3).  These dimensions are set/corrected
        automatically when parsing the tally output file.  Thus, for very simple tallies, most of these indices will be
        set to 0 when accessing tally results, e.g. `tally_data[2,0,0,:,0,0,0,:,0,:]` to access the full energy spectrum
        in the third region for all scored particles / particle groups with the values and uncertainties.
+       
+       -----
+       
+       **The `'tally_metadata'` dictionary**
 
        The output `tally_metadata` dictionary contains all information needed to identify every bin along every
        dimension: region numbers/groups, particle names/groups, bin edges and midpoints for all mesh types
        (x, y, z, r, energy, angle, time, and LET) used in the tally.
+       
+       It typically contains quite a bit of information. To conveniently view its contents, one can import the built-in
+       pprint library `import pprint` and then use `pprint.pp(dict(tally_output['tally_metadata']))`.
+       
+       At a basic level, the "header" of the PHITS output file (everything before the first instance of `newpage:`) is
+       parsed and, in effect, a key+value pair is created for every line containing an equals sign `=`.
+       
+       The following keys exist denoting the lengths of the first 9 axes/dimensions of the `tally_data` NumPy array: 
+       `'nreg'`/`'nx'`/`'nr'`/`'ntet'`, `'ny'`, `'nz'`, `'ne'`, `'nt'`, `'na'`, `'nl'`,  `'npart'`, and `'nc'`. 
+       These keys are all initialized as `None` and assigned a value if found in the PHITS output file. Thus, if one of
+       these retains its value of `None`, the corresponding array axis length should be 1.  The `'found_mesh_kinds'` key 
+       lists which of these meshes are identified.
[…]
