### 2025-03-10 ec005a3
fixed typo in expanded documentation
