### 2025-03-18 6e2c2c8
added to docs quick code for pickle loading
