### 2025-03-19 57e151b
addressed edge case for tallies with >999 output tables
