### 2025-04-01 5b6f78c
Update README.md with PyPI info
