### 2025-04-01 #7691 @Lindt8
As a note, I recall specifying a specific release number when submitting this to JOSS.  I just wanted to note that since submission there have been some updates to the PHITS Tools package (new [features](https://github.com/Lindt8/PHITS-Tools/releases), improved and expanded [documentation](https://lindt8.github.io/PHITS-Tools/), Jupyter notebook version of example, package now listed on [PyPI](https://pypi.org/project/PHITS-Tools/), etc.) now included in the most recent release (1.2.3).
