+++ README.md
@@ -4,6 +4,9 @@
 [![status](https://joss.theoj.org/papers/ef67acccadb883867ba60dc9e018ff70/status.svg)](https://joss.theoj.org/papers/ef67acccadb883867ba60dc9e018ff70)
 [![PyPI - Version](https://img.shields.io/pypi/v/PHITS-Tools)](https://pypi.org/project/PHITS-Tools/)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14262720.svg)](https://doi.org/10.5281/zenodo.14262720)
+[![PHITS forumn post on PHITS Tools](https://img.shields.io/badge/PHITS%20forum%20post%20-%20%2333a2d9)](https://meteor.nucl.kyushu-u.ac.jp/phitsforum/t/topic/3651/)
+
+
 
 ## Purpose
 
