### 2025-04-14 7a0a503
fixed missing Pandas df columns for [T-Deposit2] and [T-Interact] with axis=act
