### 2025-04-14 cf01f6f
Revert "fixed missing Pandas df columns for [T-Deposit2] and [T-Interact] with axis=act"
This reverts commit 7a0a503d5620ca896c2122d3c2cfd3ab4133aad0.
