+++ PHITS_tools.py
@@ -140,7 +140,7 @@ run_with_CLI_inputs = False
 in_debug_mode = False # toggles printing of debug messages throughout the code
 #in_debug_mode = True # toggles printing of debug messages throughout the code
 test_explicit_files_dirs = False # used for testing specific files at the bottom of this file
-test_explicit_files_dirs = True
+#test_explicit_files_dirs = True
 
 if __name__ == "__main__":
     #in_debug_mode = True
@@ -2903,7 +2903,7 @@ def parse_tally_content(tdata,meta,tally_blocks,is_err_in_separate_file,err_mode
                     hli = li
                     hli_found = True
                     continue
-                if hli_found and (line[:12] == '#   sum over' or line[:7] == '#   sum' or line[:5] == '#----' or line[:12] == '#   overflow' or (len(block[li-1]) == 0 and hli != 0 and li>hli+2) or "'" in line or '{' in line):
+                if hli_found and (line[:12] == '#   sum over' or line[:7] == '#   sum' or line[:5] == '#----' or (len(block[li-1]) == 0 and hli != 0 and li>hli+2) or "'" in line or '{' in line):
                     fli = li
                     if (len(block[li-1]) == 0 and hli != 0 and li>hli+2): fli = li - 1 # triggered by blank line after data
                     #if "'" in line or '{' in line:
@@ -3865,18 +3865,10 @@ def build_tally_Pandas_dataframe(tdata,meta):
     ecols, tcols, acols, lcols, pcols, ccols = False, False, False, False, False, False
     single_specified_bin_axes = [] # log axes which are provided by user but only contain 1 bin
     single_bin_ranges_or_values = []
-    if meta['ne'] != None or ie_max>1: #  
+    if meta['ne'] != None:
         if meta['ne']==1:
             single_specified_bin_axes.append('e')
             single_bin_ranges_or_values.append(['Energy',meta['e-mesh_bin_edges']])
-        elif meta['tally_type']=='[T-Deposit2]':
-            if meta['ne1']==1:
-                single_specified_bin_axes.append('e1')
-                single_bin_ranges_or_values.append(['Energy1', meta['e1-mesh_bin_edges']])
-            else:
-                ecols = True
-                ecol_names_list = ['ie', 'e1_mid']
-                col_names_list += ecol_names_list
         else:
             ecols = True
             ecol_names_list = ['ie','e_mid']
@@ -3914,8 +3906,8 @@ def build_tally_Pandas_dataframe(tdata,meta):
     else:
         single_bin_ranges_or_values.append(['LET','default/all'])
 
-    if meta['nc'] != None or ic_max>1:
-        if meta['nc'] == 1 or ic_max==1:
+    if meta['nc'] != None:
+        if meta['nc'] == 1:
             pass
         else:
             ccols = True
@@ -3929,17 +3921,8 @@ def build_tally_Pandas_dataframe(tdata,meta):
                 elif meta['axis'] == 'mass':
                     ccol_names_list = ['ic/A/mass']
                     col_names_list += ccol_names_list
-            elif meta['tally_type'] == '[T-Interact]':
-                if meta['axis'] == 'act':
-                    ccol_names_list = ['ic', '#Interactions']
-                    col_names_list += ccol_names_list
             elif meta['tally_type'] == '[T-Deposit2]':
-                if meta['ne2'] == 1:
-                    single_specified_bin_axes.append('e2')
-                    single_bin_ranges_or_values.append(['Energy2', meta['e2-mesh_bin_edges']])
-                else:
-                    ccol_names_list = ['ic', 'e2_mid']
-                    col_names_list += ccol_names_list
+                pass
 
[…]
