### 2025-04-14 2a70ce6
fixed missing Pandas df columns for [T-Deposit2] and [T-Interact] with axis=act
