### 2025-04-24 1c642ce
added automatic plotting functionality
