+++ PHITS_tools.py
@@ -71,6 +71,7 @@ functions return the data objects they produce for your own further analyses.
 - `data_row_to_num_list`            : extract numeric values from a line in the tally content section
 - `calculate_tally_absolute_errors` : calculate absolute uncertainties from read values and relative errors
 - `build_tally_Pandas_dataframe`    : make Pandas dataframe from the main results NumPy array and the metadata
+- `autoplot_tally_results`          : make plot(s), saved as PDFs, of tally results from tally output Pandas DataFrame(s)
 
 
 ### **CLI options**
@@ -89,6 +90,7 @@ Essentially, the CLI serves to interface with the core three functions of PHITS 
       - `-np` sets `make_PandasDF = False` (`True` if excluded)
       - `-na` sets `calculate_absolute_errors = False` (`True` if excluded)
       - `-lzma` sets `compress_pickle_with_lzma = True` (`False` if excluded)
+      - `-p` sets `autoplot_tally_output = True` (`False` if excluded)
 - `parse_tally_dump_file` (and passed to it via `parse_all_tally_output_in_dir`)
       - `-d` tells the CLI that `file` should be processed as a dump file (if it's not a directory)
       - `-dvals` passes the provided sequence of values to `dump_data_sequence` (`None` if excluded)
@@ -108,6 +110,7 @@ Essentially, the CLI serves to interface with the core three functions of PHITS 
       - `-d` sets `include_dump_files = True` (`False` if excluded)
       - `-dnmmpi` sets `dump_merge_MPI_subdumps = False` (`True` if excluded)
       - `-dndmpi` sets `dump_delete_MPI_subdumps_post_merge = False` (`True` if excluded)
+      - `-pa` sets `autoplot_all_tally_output_in_dir = True` (`False` if excluded)
 
 
 '''
@@ -171,7 +174,8 @@ if in_debug_mode:
 
 
 def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calculate_absolute_errors = True,
-                            save_output_pickle=True, prefer_reading_existing_pickle=False, compress_pickle_with_lzma=False):
+                            save_output_pickle=True, prefer_reading_existing_pickle=False, compress_pickle_with_lzma=False, 
+                            autoplot_tally_output=False):
     '''
     Description:
         Parse any PHITS tally output file, returning tally metadata and an array of its values (and optionally
@@ -214,6 +218,9 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                       For most "normal" tallies, the pickle file sizes are likely to be small enough to not warrant 
                       compression, but LZMA compression can reduce the file size by several orders of magnitude, 
                       great for output from massive tallies.
+       - **`autoplot_tally_output`** = (optional, D=`False`; requires `make_PandasDF=True`) Boolean denoting whether 
+                      the tally's output will be automatically plotted and saved to a PDF (of the same name/path as 
+                      `tally_output_filepath` but ending in `.pdf`) using the `autoplot_tally_results()` function.
 
     Output:
         - **`tally_output`** = a dictionary object with the below keys and values:
@@ -543,9 +550,9 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
             return dchain_output
         else:
             return None
-    if tally_metadata['tally_type'] == '[T-Yield]' and tally_metadata['axis'] == 'dchain':
+    if (tally_metadata['tally_type'] == '[T-Yield]' or tally_metadata['tally_type'] == '[T-Track]') and tally_metadata['axis'] == 'dchain':
         dchain_tools_url = 'github.com/Lindt8/DCHAIN-Tools'
-        print('This function does not support [T-Yield] with setting "axis = dchain".')
[…]
