+++ requirements.txt
@@ -3,6 +3,7 @@ numpy>=2.2.1
 pandas>=2.2.3
 scipy>=1.14.1
 matplotlib>=3.10.0
+seaborn>=0.13.2
 
 ## Optional Dependencies (uncomment to use)
 # Munch is used to make the metadata dictionary into a Munch object
