### 2025-04-25 5b23a31
autoplot function can now return FacetGrid objects
