+++ PHITS_tools.py
@@ -187,6 +187,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
     Dependencies:
         - `import numpy as np`
         - `import pandas as pd` (if `make_PandasDF = True`)
+        - `import seaborn as sns` (if `autoplot_tally_output = True`)
         - `from munch import Munch` (will still run if package not found)
 
     Inputs:
@@ -1085,6 +1086,7 @@ def parse_all_tally_output_in_dir(tally_output_dirpath, output_file_suffix = '.o
     Dependencies:
         - `import numpy as np`
         - `import pandas as pd` (if `make_PandasDF = True`)
+        - `import seaborn as sns` (if `autoplot_tally_output=True` or `autoplot_all_tally_output_in_dir=True`)
         - `from munch import Munch` (will still run if package not found)
         - `from collections import namedtuple` (if processing tally dump files)
         - `from scipy.io import FortranFile` (if processing tally dump files)
@@ -4181,7 +4183,7 @@ def build_tally_Pandas_dataframe(tdata,meta):
     return tally_df
 
 
-def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename='results.pdf',show_plots=False):
+def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename='results.pdf',show_plots=False,return_fg_list=False):
     '''
     Description:
         Generates visualizations/plots of the data in the output Pandas DataFrames from the `parse_tally_output_file()` 
@@ -4194,7 +4196,10 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
         This function seeks to compile plots of results from one or multiple tallies into a single PDF file. 
         It is primarily intended to be called by `parse_tally_output_file()` and `parse_all_tally_output_in_dir()`.
         The [seaborn](https://seaborn.pydata.org/) package's [relplot](https://seaborn.pydata.org/generated/seaborn.relplot.html) function is used for generating these plots.
-
+        If you wish to make modifications to the automatically generated figures, use the `return_fg_list=True` setting 
+        and apply your desired modifications to the returned FacetGrid objects.  (This is demonstrated in the 
+        [example](https://github.com/Lindt8/PHITS-Tools/tree/main/example) distributed with PHITS Tools.)
+        
     Dependencies:
         - `import seaborn as sns`
         - `import pandas as pd`
@@ -4214,9 +4219,14 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                 PDF of plots will be saved.
         - `show_plots` = (optional, D=`False`) Boolean denoting whether this function will make a `plt.show()` call
                 immediately before the `return` statement.
+        - `return_fg_list` = (optional, D=`False`) Boolean denoting whether a list of the generated seaborn FacetGrid 
+                objects returned by the sns.relplot() calls should be returned by this function, allowing modification 
+                of the automatically generated plots.  (Strictly speaking, the objects are copies of the FacetGrid objects 
+                made using the built-in [copy.deepcopy()](https://docs.python.org/3/library/copy.html#copy.deepcopy) function.)  If `False`, this function returns `None`.
 
     Outputs:
-        - `None` (and the saved file of plot(s) specified by `output_filename`)
+        - `None` if `return_fg_list=False` (default) or `fg_list` if `return_fg_list=True`; see description of the `return_fg_list` input argument
+        - (and the saved file of plot(s) specified by `output_filename`)
 
     '''
     '''
[…]
