### 2025-04-28 b857bd1
improved autoplot function documentation
