+++ PHITS_tools.py
@@ -45,6 +45,7 @@ functions return the data objects they produce for your own further analyses.
 
 - `tally`                           : tally/histogram values (and their indices) falling within a desired binning structure (useful with "dump" files)
 - `rebinner`                        : rebin a set of y-data to a new x-binning structure (edges need not necessarily be preserved)
+- `autoplot_tally_results`          : make plot(s), saved as PDFs, of tally results from tally output Pandas DataFrame(s)
 - `fetch_MC_material`               : returns a string of a formatted material for PHITS or MCNP (mostly those in [PNNL-15870 Rev. 1](https://www.osti.gov/biblio/1023125))
 - `ICRP116_effective_dose_coeff`    : returns effective dose conversion coefficient of a mono-energetic particle of some species and some geometry; does coefficients are those in [ICRP 116](https://doi.org/10.1016/j.icrp.2011.10.001)
 - `merge_dump_file_pickles`         : merge multiple dump file outputs into a single file (useful for dumps in MPI runs)
@@ -71,7 +72,6 @@ functions return the data objects they produce for your own further analyses.
 - `data_row_to_num_list`            : extract numeric values from a line in the tally content section
 - `calculate_tally_absolute_errors` : calculate absolute uncertainties from read values and relative errors
 - `build_tally_Pandas_dataframe`    : make Pandas dataframe from the main results NumPy array and the metadata
-- `autoplot_tally_results`          : make plot(s), saved as PDFs, of tally results from tally output Pandas DataFrame(s)
 
 
 ### **CLI options**
@@ -1496,6 +1496,806 @@ def rebinner(output_xbins,input_xbins,input_ybins):
                 output_ybins[i] = output_ybins[i] + f_in_dxo*input_ybins[j]
 
     return output_ybins
+
+
+def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename='results.pdf',show_plots=False,return_fg_list=False):
+    '''
+    Description:
+        Generates visualizations/plots of the data in the output Pandas DataFrames from the `parse_tally_output_file()` 
+        function in an automated fashion.  Note that this function only seeks to accomplish exactly this. 
+        It is not a function for generating customized plots; it exists to automate creating visualizations of PHITS 
+        output using sets of predetermined rules and settings, principally for initial checking of results. 
+        Generally, it does not respect plotting-relevant settings provided to PHITS tallies (e.g., `samepage`, `axis`, `angel`, etc.), 
+        though it will use the `title` parameter and the plot axis titles for ANGEL included in the .out files, which are
+        influenced by some tally parameters, such as `unit` and `y-txt`.
+        This function seeks to compile plots of results from one or multiple tallies into a single PDF file. 
+        It is primarily intended to be called by `parse_tally_output_file()` and `parse_all_tally_output_in_dir()`.
+        The [seaborn](https://seaborn.pydata.org/) package's [relplot](https://seaborn.pydata.org/generated/seaborn.relplot.html) function is used for generating these plots.
+        If you wish to make modifications to the automatically generated figures, use the `return_fg_list=True` setting 
+        and apply your desired modifications to the returned FacetGrid objects.  (This is demonstrated in the 
+        [example](https://github.com/Lindt8/PHITS-Tools/tree/main/example) distributed with PHITS Tools.)
+        
+    Dependencies:
+        - `import seaborn as sns`
+        - `import pandas as pd`
+        - `import matplotlib.pyplot as plt`
+        - `from matplotlib.colors import LogNorm, SymLogNorm, Normalize`
+        - `from matplotlib.backends.backend_pdf import PdfPages`
+
+    Inputs:
+        - `tally_output_list` = the `tally_output` output from the `parse_tally_output_file()` function, a string/Path
[…]
