### 2025-04-28 4d80b3d
added extra recognized TeX characters
