### 2025-04-29 d27043d
autoplot extra input options, can save other image formats (+PNG by default)
