+++ PHITS_tools.py
@@ -220,8 +220,8 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                       compression, but LZMA compression can reduce the file size by several orders of magnitude, 
                       great for output from massive tallies.
        - **`autoplot_tally_output`** = (optional, D=`False`; requires `make_PandasDF=True`) Boolean denoting whether 
-                      the tally's output will be automatically plotted and saved to a PDF (of the same name/path as 
-                      `tally_output_filepath` but ending in `.pdf`) using the `autoplot_tally_results()` function.
+                      the tally's output will be automatically plotted and saved to a PDF and PNG (of the same name/path as 
+                      `tally_output_filepath` but ending in `.pdf`/`.png`) using the `autoplot_tally_results()` function.
 
     Output:
         - **`tally_output`** = a dictionary object with the below keys and values:
@@ -632,7 +632,9 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
             print('Plotting via "autoplot_tally_output=True" requires also setting "make_PandasDF=True".')
         else:
             plot_filepath = Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pdf')
-            autoplot_tally_results(tally_output, plot_errorbars=calculate_absolute_errors, output_filename=plot_filepath)
+            autoplot_tally_results(tally_output, output_filename=plot_filepath,
+                                   plot_errorbars=calculate_absolute_errors, 
+                                   additional_save_extensions=['.png'])
     
     return tally_output
 
@@ -1156,8 +1158,8 @@ def parse_all_tally_output_in_dir(tally_output_dirpath, output_file_suffix = '.o
                       compression, but LZMA compression can reduce the file size by several orders of magnitude, 
                       great for output from massive tallies.
        - **`autoplot_tally_output`** = (optional, D=`False`; requires `make_PandasDF=True`) Boolean denoting whether, for each tally, 
-                      the tally's output will be automatically plotted and saved to a PDF (of the same name/path as 
-                      the tally's output file but ending in `.pdf`) using the `autoplot_tally_results()` function. 
+                      the tally's output will be automatically plotted and saved to a PDF and PNG (of the same name/path as 
+                      `tally_output_filepath` but ending in `.pdf`/`.png`) using the `autoplot_tally_results()` function. 
 
     Inputs:
        (optional, the same as in and directly passed to the `parse_tally_dump_file()` function)
@@ -1498,7 +1500,9 @@ def rebinner(output_xbins,input_xbins,input_ybins):
     return output_ybins
 
 
-def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename='results.pdf',show_plots=False,return_fg_list=False):
+def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename='results.pdf',
+                           additional_save_extensions=[],show_plots=False,return_fg_list=False,
+                           max_num_values_to_plot=1e7,rasterizesize_threshold=5e4,rasterize_dpi=300):
     '''
     Description:
         Generates visualizations/plots of the data in the output Pandas DataFrames from the `parse_tally_output_file()` 
@@ -1508,10 +1512,10 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
         Generally, it does not respect plotting-relevant settings provided to PHITS tallies (e.g., `samepage`, `axis`, `angel`, etc.), 
         though it will use the `title` parameter and the plot axis titles for ANGEL included in the .out files, which are
         influenced by some tally parameters, such as `unit` and `y-txt`.
-        This function seeks to compile plots of results from one or multiple tallies into a single PDF file. 
[…]
