### 2025-05-14 9a6692b
added phits.out parsing, "input file mode", and output merging
