+++ PHITS_tools.py
@@ -40,6 +40,7 @@ functions return the data objects they produce for your own further analyses.
 - `parse_tally_output_file`         : general parser for standard output files for all PHITS tallies
 - `parse_tally_dump_file`           : parser for dump files from "dump" flag in PHITS [T-Cross], [T-Time], and [T-Track] tallies
 - `parse_all_tally_output_in_dir`   : run `parse_tally_output_file()` over all standard output files in a directory (and, optionally, `parse_tally_dump_file()` over all dump files too)
+- `parse_phitsout_file`             : creates a metadata dictionary of a PHITS run from its "phits.out" file
 
 ### General Purpose Functions
 
@@ -72,7 +73,7 @@ functions return the data objects they produce for your own further analyses.
 - `data_row_to_num_list`            : extract numeric values from a line in the tally content section
 - `calculate_tally_absolute_errors` : calculate absolute uncertainties from read values and relative errors
 - `build_tally_Pandas_dataframe`    : make Pandas dataframe from the main results NumPy array and the metadata
-
+- `extract_tally_outputs_from_phits_input` : extract a dictionary of files produced by a PHITS run
 
 ### **CLI options**
 
@@ -90,6 +91,7 @@ Essentially, the CLI serves to interface with the core three functions of PHITS 
       - `-np` sets `make_PandasDF = False` (`True` if excluded)
       - `-na` sets `calculate_absolute_errors = False` (`True` if excluded)
       - `-lzma` sets `compress_pickle_with_lzma = True` (`False` if excluded)
+      - `-po` sets `include_phitsout_in_metadata = True` (`False` if excluded)
       - `-p` sets `autoplot_tally_output = True` (`False` if excluded)
 - `parse_tally_dump_file` (and passed to it via `parse_all_tally_output_in_dir`)
       - `-d` tells the CLI that `file` should be processed as a dump file (if it's not a directory)
@@ -110,8 +112,74 @@ Essentially, the CLI serves to interface with the core three functions of PHITS 
       - `-d` sets `include_dump_files = True` (`False` if excluded)
       - `-dnmmpi` sets `dump_merge_MPI_subdumps = False` (`True` if excluded)
       - `-dndmpi` sets `dump_delete_MPI_subdumps_post_merge = False` (`True` if excluded)
+      - `-m` sets `merge_tally_outputs = True` (`False` if excluded)
+      - `-smo` sets `merge_tally_outputs = True` (`False` if excluded), `save_output_pickle = False` (`True` if excluded), and `save_pickle_of_merged_tally_outputs = True` (`None` if excluded)
       - `-pa` sets `autoplot_all_tally_output_in_dir = True` (`False` if excluded)
 
+Below is a picture of all of these options available for use within the CLI and their descriptions.  
+
+![](https://github.com/Lindt8/PHITS-Tools/blob/main/docs/PHITS_tools_CLI.png?raw=true "PHITS Tools CLI options")
+
+### **Automatic processing at PHITS runtime**
+
+PHITS Tools can be used to automatically process the output of every PHITS run executed with the "phits.bat" and "phits.sh" 
+scripts found in the "phits/bin/" directory of your PHITS distribution.  To do this, first you must identify the location 
+of your "PHITS_tools.py" file.  If using the file directly downloaded from GitHub, this should be in a location of your choosing.
+If you installed PHITS Tools via `pip install PHITS-Tools`, you can find its location with `pip show PHITS-Tools -f`. 
+Once you have identified the location of PHITS_tools.py, for example "/path/locating/PHITS_Tools/PHITS_tools.py", you can
+add the following line to your PHITS execution script:
+
+On Windows, using "phits/bin/phits.bat":
+
[…]
