+++ test/test_PHITS_tools.py
@@ -7,6 +7,7 @@ path_to_phits_base_folder = Path('C:\phits')
 
 phits_sample_dir = Path(path_to_phits_base_folder,'sample')
 phits_recommendation_dir = Path(path_to_phits_base_folder,'recommendation')
+test_input_mode_parsing = True  # Determines if [INPUT_FILE mode] of parse_all_tally_output_in_dir() is tested for all input files
 test_autoplotting = False  # Determines if autoplot_tally_results() is tested too for each tally output (notably slows testing)
 plot_paths = []
 
@@ -29,18 +30,16 @@ for f in recommendation_files:
     if keep_file:
         files_to_parse.append(Path(f))
 
-
 log_file_str = ''
 num_tests = len(files_to_parse)
 i = 0
-num_passed = 0
-num_failed = 0
-num_warn = 0
-known_issue_files = [r'phits\sample\source\Cosmicray\GCR-ground\cross.out',
-                     r'phits\sample\source\Cosmicray\GCR-LEO\cross.out',
-                     r'phits\sample\source\Cosmicray\GCR-space\cross.out',
-                     r'phits\sample\source\Cosmicray\SEP-space\cross.out',
-                     r'phits\sample\source\Cosmicray\TP-LEO\cross.out']
+num_passed, num_failed, num_warn = 0, 0, 0
+known_issue_files = [r'phits\sample\source\Cosmicray\GCR-ground\cross.out', r'phits\sample\source\Cosmicray\GCR-ground\GCR-ground.inp',
+                     r'phits\sample\source\Cosmicray\GCR-LEO\cross.out', r'phits\sample\source\Cosmicray\GCR-LEO\GCR-LEO.inp',
+                     r'phits\sample\source\Cosmicray\GCR-space\cross.out', r'phits\sample\source\Cosmicray\GCR-space\GCR-space.inp',
+                     r'phits\sample\source\Cosmicray\SEP-space\cross.out', r'phits\sample\source\Cosmicray\SEP-space\SEP-space.inp',
+                     r'phits\sample\source\Cosmicray\TP-LEO\cross.out', r'phits\sample\source\Cosmicray\TP-LEO\TP-LEO.inp',
+                     r'phits\sample\misc\snowman\rotate3dshow.inp']
 for f in files_to_parse:
     i += 1
     test_num_str = '{:3d}/{:3d}'.format(i,num_tests)
@@ -48,7 +47,7 @@ for f in files_to_parse:
         if '_dmp.out' in str(f):
             x = PHITS_tools.parse_tally_dump_file(f, save_namedtuple_list=False, save_Pandas_dataframe=False)
         else:
-            x = PHITS_tools.parse_tally_output_file(f,save_output_pickle=False,autoplot_tally_output=test_autoplotting)
+            x = PHITS_tools.parse_tally_output_file(f, save_output_pickle=False, autoplot_tally_output=test_autoplotting)
             if test_autoplotting and Path(f.parent, f.stem + '.pdf').is_file(): plot_paths.append(Path(f.parent, f.stem + '.pdf'))
         log_str = test_num_str + '     pass  ' + str(f) + '\n'
         num_passed += 1
@@ -70,6 +69,57 @@ log_str += '{:3d} of {:3d} tests failed (including "WARN")\n'.format(num_failed,
 log_str += '{:3d} of {:3d} the failed tests are from old distributed files and should succeed if the corresponding PHITS input is reran (labeled with "WARN").\n'.format(num_warn,num_failed)
 print(log_str)
 log_file_str += log_str
+
+
+if test_input_mode_parsing:
[…]
