+++ README.md
@@ -66,6 +66,7 @@ The CLI principally serves to interface with the core three functions of PHITS T
   - `-np` sets `make_PandasDF = False` (`True` if excluded)
   - `-na` sets `calculate_absolute_errors = False` (`True` if excluded)
   - `-lzma` sets `compress_pickle_with_lzma = True` (`False` if excluded)
+  - `-po` sets `include_phitsout_in_metadata = True` (`False` if excluded)
   - `-p` sets `autoplot_tally_output = True` (`False` if excluded)
 - [**`parse_tally_dump_file()`**](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.parse_tally_dump_file) (and passed to it via [**`parse_all_tally_output_in_dir()`**](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.parse_all_tally_output_in_dir))
   - `-d` tells the CLI that `file` should be processed as a dump file (if it's not a directory)
@@ -86,11 +87,74 @@ The CLI principally serves to interface with the core three functions of PHITS T
   - `-d` sets `include_dump_files = True` (`False` if excluded)
   - `-dnmmpi` sets `dump_merge_MPI_subdumps = False` (`True` if excluded)
   - `-dndmpi` sets `dump_delete_MPI_subdumps_post_merge = False` (`True` if excluded)
+  - `-m` sets `merge_tally_outputs = True` (`False` if excluded)
+  - `-smo` sets `merge_tally_outputs = True` (`False` if excluded), `save_output_pickle = False` (`True` if excluded), and `save_pickle_of_merged_tally_outputs = True` (`None` if excluded)
   - `-pa` sets `autoplot_all_tally_output_in_dir = True` (`False` if excluded)
 
 Below is a picture of all of these options available for use within the CLI.  
 
 ![](https://github.com/Lindt8/PHITS-Tools/blob/main/docs/PHITS_tools_CLI.png?raw=true "PHITS Tools CLI options")
+
+## **Automatic processing at PHITS runtime**
+
+PHITS Tools can be used to automatically process the output of every PHITS run executed with the "phits.bat" and "phits.sh" 
+scripts found in the "phits/bin/" directory of your PHITS distribution.  To do this, first you must identify the location 
+of your "PHITS_tools.py" file.  If using the file directly downloaded from GitHub, this should be in a location of your choosing.
+If you installed PHITS Tools via `pip install PHITS-Tools`, you can find its location with `pip show PHITS-Tools -f`. 
+Once you have identified the location of PHITS_tools.py, for example "/path/locating/PHITS_Tools/PHITS_tools.py", you can
+add the following line to your PHITS execution script:
+
+On Windows, using "phits/bin/phits.bat":
+
+- Scroll down toward the bottom of the script, to the section with the line `rem - Your file processing starts here.`
+- After the if statement (right before the `rem - Your file processing ends here` line), insert a new line with the following command:
+- `python "C:\path\locating\PHITS_Tools\PHITS_tools.py" "%%~nxF" -po -m -d -ddir -ddeg -lzma -p -pa`
+
+On Linux/Mac, using "phits/bin/phits.sh":
+
+- Scroll down toward the bottom of the script, to the section titled `# Run PHITS`
+- On the line after the end of the if statement `fi`, add the following command:
+- `python "/path/locating/PHITS_Tools/PHITS_tools.py" $1 -po -m -d -ddir -ddeg -lzma -p -pa`
+
+(Of course, if necessary, replace "`python`" with however you typically call python in your environment, e.g. `py`, `python3`, etc.)
+
+Adding this line causes the following to happen:
+
+- After PHITS finishes running normally, the PHITS input file is passed to PHITS Tools.
[…]
