### 2025-05-15 #7691 @Lindt8
PHITS Tools has been updated to [version 1.4.0](https://github.com/Lindt8/PHITS-Tools/releases).
