### 2025-05-19 #7691 @Lindt8
Sure!  I have searched through the reviewer database and found some people I think would be suited to review this, having familiarity with radiation transport and similar codes to PHITS, such as MCNP, who should also be able to relate to the challenges PHITS Tools is addressing.

[Micah Gale](https://reviewers.joss.theoj.org/reviewers/4294)
[Alberto Previti](https://reviewers.joss.theoj.org/reviewers/4184)
[Joanna Piper Morgan](https://reviewers.joss.theoj.org/reviewers/3743)
[Samuel Ingram](https://reviewers.joss.theoj.org/reviewers/8919)
[Damar Wicaksono](https://reviewers.joss.theoj.org/reviewers/2720)
[Frédéric Tessier](https://reviewers.joss.theoj.org/reviewers/2952)
[Madicken Munk](https://reviewers.joss.theoj.org/reviewers/1175)
