### 2025-05-21 130b4b5
fixed edge case where LZMA compression was skipped for phits.out processing
