+++ PHITS_tools.py
@@ -184,6 +184,27 @@ automatic processing by omitting `-d`.  As an example, a more "minimal" automati
 This would only create the ".pickle.xz" file of the merged standard tally outputs and the PDF containing all of their 
 plots together, skipping any processing of dump files.
 
+**Automatic processing at DCHAIN runtime**
+
+Similarly, PHITS Tools can be used to automatically process output from the DCHAIN code, utilizing an import of the [DCHAIN Tools module](https://github.com/Lindt8/DCHAIN-Tools).
+
+On Windows, using "phits/dchain-sp/bin/dchain.bat":
+
+- Scroll down toward the bottom of the script, to the section with the line `rem - Your file processing ends here.`
+- Right above that line (before the `goto :continue`), insert a new line with the following command:
+- `python "C:\path\locating\PHITS_Tools\PHITS_tools.py" "%%~nxF" -po -lzma`
+
+On Linux/Mac, using "phits/dchain-sp/bin/dchain.sh":
+
+- Scroll down toward the bottom of the script, right before the line with `echo ' end of dchain '`
+- On the line after the end of the if statement `fi`, add the following command:
+- `python "/path/locating/PHITS_Tools/PHITS_tools.py" ${jnam} -po -lzma`
+
+This will create a ".pickle.xz" file of the processed DCHAIN outputs, as a dictionary object, with contents as 
+described in the documentation for `parse_tally_output_file()` under the "[T-Dchain] special case" section, 
+also including an entry with a dictionary of information on the corresponding PHITS run via its "phits.out" file, if found.
+
+
 '''
 '''
 Each function beings with a comment block containing the following sections:
@@ -210,7 +231,7 @@ from pathlib import Path
 # default program settings
 launch_GUI = False
 run_with_CLI_inputs = False
-in_debug_mode = False # toggles printing of debug messages throughout the code
+in_debug_mode = False # True # toggles printing of debug messages throughout the code
 #in_debug_mode = True # toggles printing of debug messages throughout the code
 test_explicit_files_dirs = False # used for testing specific files at the bottom of this file
 #test_explicit_files_dirs = True
@@ -452,11 +473,11 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
         - [T-Userdefined] (due to having no standard format)
         - [T-Dchain]&Dagger;
 
-       &Dagger;If provided with the output file of [T-Dchain] (the input file for the DCHAIN code) or the `*.act` main
-       output file produced by the DCHAIN code, this function will attempt to import the [DCHAIN Tools module](https://github.com/Lindt8/DCHAIN-Tools)
+       &Dagger;If provided with the output file of [T-Dchain] (the input file for the DCHAIN code), the `*.act` main
+       output file produced by the DCHAIN code, or the `*.dtrk`/`*.dyld` [T-Track]/[T-Yield] tally outputs spawned by 
+       [T-Dchain], this function will attempt to import the [DCHAIN Tools module](https://github.com/Lindt8/DCHAIN-Tools)
        and process the found DCHAIN output files, returning the output dictionary object and (optionally) saving it to a pickle file.
-       Otherwise, for [T-Dchain] or [T-Yield] with `axis = dchain`, you can manually use the separate suite of parsing
-       functions included in the [DCHAIN Tools](https://github.com/Lindt8/DCHAIN-Tools) module.
+       See the **[T-Dchain] special case** section further below for more details.
 
        -----
 
@@ -500,7 +521,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
[…]
