### 2025-05-29 f1ad52b
added "table of isotope" plot type for axis=chart/dchain [T-Yield] tallies
