+++ PHITS_tools.py
@@ -832,7 +832,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
         else:
             print('\tERROR! tally type',tally_metadata['tally_type'],'is not supported by this function!')
             return None
-    if tally_metadata['tally_type'] == '[T-Dchain]' or tally_output_filepath.suffix == '.act':
+    if tally_metadata['tally_type'] == '[T-Dchain]' or (tally_metadata['tally_type'] == 'UNKNOWN' and tally_output_filepath.suffix == '.act'):
         print('\tNOTE: The DCHAIN Tools module is used to process the DCHAIN output files with same basename of provided file.')
         dchain_tools_url = 'github.com/Lindt8/DCHAIN-Tools'
         dchain_tools_go_to_github_str = 'The DCHAIN Tools module ( '+dchain_tools_url+' ) is capable of parsing all DCHAIN-related output.'
@@ -1827,7 +1827,10 @@ def parse_all_tally_output_in_dir(tally_output_dirpath, output_file_suffix = Non
                     else:
                         tally_include_in_plotting_list.append(True)
             else:
-                tally_output_pickle_path_list.append(path_to_pickle_file)
+                if 'path_to_pickle_file' in tally_output:
+                    tally_output_pickle_path_list.append(tally_output['path_to_pickle_file']) 
+                else:
+                    tally_output_pickle_path_list.append(path_to_pickle_file)
                 tot_num_values = np.prod(np.shape(tally_output['tally_data'])[:-1])
                 if tot_num_values > max_num_values_to_plot:
                     tally_include_in_plotting_list.append(False)
@@ -1997,7 +2000,7 @@ def parse_phitsout_file(phitsout_filepath, include_input_echo=True, save_phitsou
     import io
     import pandas as pd
     phitsout_filepath = Path(phitsout_filepath)
-    print('Processing file:', phitsout_filepath)
+    print('\tProcessing file:', phitsout_filepath)
     starting_dict = {'job':{},'summary':{},'memory':{},'batch':{}}
     if prefer_to_munch_meta_dict:
         try:
@@ -2208,7 +2211,7 @@ def parse_phitsout_file(phitsout_filepath, include_input_echo=True, save_phitsou
         else:
             with open(pickle_path, 'wb') as handle:
                 pickle.dump(phitsout_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)
-        print('\tPickle file written:', pickle_path)
+        print('\t\tPickle file written:', pickle_path)
     return phitsout_dict
 
 
@@ -2391,12 +2394,15 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
         Generally, it does not respect plotting-relevant settings provided to PHITS tallies (e.g., `samepage`, `axis`, `angel`, etc.), 
         though it will use the `title` parameter and the plot axis titles for ANGEL included in the .out files, which are
         influenced by some tally parameters, such as `unit` and `y-txt`.
+        
         This function seeks to compile plots of results from one or multiple tallies into a single PDF file (and individual files in other image formats). 
         The [seaborn](https://seaborn.pydata.org/) package's [relplot](https://seaborn.pydata.org/generated/seaborn.relplot.html) function is used for generating these plots.
         This function is primarily intended to be called by `parse_tally_output_file()` and `parse_all_tally_output_in_dir()`. 
         However, if you wish to make modifications to the automatically generated figures, you can use the `return_fg_list=True` setting 
         and apply your desired modifications to the returned FacetGrid objects.  (This is demonstrated in the 
         [example](https://github.com/Lindt8/PHITS-Tools/tree/main/example) distributed with PHITS Tools.)
+        
[…]
