### 2025-05-29 #8311 @editorialbot
[REVIEW]: The PHITS Tools Python package for parsing, organizing, and analyzing results from the PHITS radiation transport and DCHAIN activation codes
**Submitting author:** @Lindt8 (<a href="http://orcid.org/0000-0003-3761-5415">Hunter Ratliff</a>)
**Repository:** https://github.com/Lindt8/PHITS-Tools
**Branch with paper.md** (empty if default branch):
**Version:** v1.6.0
