### 2025-05-30 053cc60
added link to publicly distributed "recommendation" files available for testing
