### 2025-05-30 9d298c4
fixed dir mode edge-case issue for handling unsupported outputs when merging enabled, pickle saving disabled, and combined plotting enabled
