### 2025-05-30 eb1740e
use raw str for path in test script
