### 2025-05-30 73c596e
progress counter for input mode parsing tests is now separate from the main progress counter
