+++ README.md
@@ -1,13 +1,13 @@
 #### Online PHITS Tools documentation: [lindt8.github.io/PHITS-Tools/](https://lindt8.github.io/PHITS-Tools/)
 
 # PHITS Tools
+[![Documentation](https://img.shields.io/badge/Documentation-brightgreen)](https://lindt8.github.io/PHITS-Tools/)
 [![status](https://joss.theoj.org/papers/ef67acccadb883867ba60dc9e018ff70/status.svg)](https://joss.theoj.org/papers/ef67acccadb883867ba60dc9e018ff70)
 [![PyPI - Version](https://img.shields.io/pypi/v/PHITS-Tools?logo=pypi&logoColor=fff&label=PyPI)](https://pypi.org/project/PHITS-Tools/)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14262720.svg)](https://doi.org/10.5281/zenodo.14262720)
-[![Documentation](https://img.shields.io/badge/Documentation-brightgreen)](https://lindt8.github.io/PHITS-Tools/)
 [![PHITS forumn discussion on PHITS Tools](https://img.shields.io/badge/PHITS%20forum%20discussion%20-%20%2333a2d9)](https://meteor.nucl.kyushu-u.ac.jp/phitsforum/t/topic/3651/)
 
-
+[Purpose](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#purpose) | [Installation](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#installation) | [Usage](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#primary-usageinterfaces) | [CLI options](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#cli-options) | [Automatic processing](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#automatic-processing-at-phits-runtime) | [Testing/Issues/Contribution](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#testing-reporting-issues-and-contributing) 
 
 ## Purpose
 
