### 2025-06-27 Lindt8/PHITS-Tools#1 @MicahGale
Improve packaging and long term maintainability
This is part of the [PHITS tools JOSS review](https://github.com/openjournals/joss-reviews/issues/8311).

I see that you have added this package to [PyPI](https://pypi.org/project/PHITS-Tools/), which is great. However, looking at the repo it is unclear to me how you did this, and I imagine it's a very manually process that only you can do right now.

This is not great for long-term maintainability. To fix this I think you should make this into a true python package. [PyOpenSci has some good tutorials on this](https://www.pyopensci.org/python-package-guide/index.html), also I have done this for [MontePy](https://github.com/idaholab/MontePy), so let me know what questions you have.

In general I would recommend:

1. [x] restructure the python files to be a true package
2. [x] Create a [`pyproject.toml`](https://packaging.python.org/en/latest/specifications/pyproject-toml/) file.
3. [x] Consider if [`requirements.txt` is still needed in addition to `pyproject.toml`](https://stackoverflow.com/questions/74508024/is-requirements-txt-still-needed-when-using-pyproject-toml)
4. [x] Create a single source of truth for the version number. Also you should probably document the versioning scheme used. I haven't seen 4 version numbers used before. I generally recommend [Semantic versioning](https://en.wikipedia.org/wiki/Software_versioning#Semantic_versioning).
5. [x] [Specify a command line script entry point](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#creating-executable-scripts) so the package is executable from anywhere on the computer.
6. [x] Consider adding CI testing of building and installing this package
In the end you should be able to:

1. Generate a source tarball and a wheel using `python -m build .`
2. Install the repository locally just using: `python -m pip .`
3. execute the package with the preferred command: e.g., `python -m PHITS_Tools` or `PHITS_tools` (or other name).

Note: I would recommend renaming the package, but that will be a separate issue.
