### 2025-06-27 Lindt8/PHITS-Tools#2 @MicahGale
Consider alternatives to git submodules for DCHAIN tools
This is part of the [PHITS tools JOSS review](https://github.com/openjournals/joss-reviews/issues/8311).

I see that DCHAIN tools is included as a git submodule. In my experience git submodules can be a bit tedious, especially for new users. I am wondering if this style of dependency is necessary. Is DCHAIN_Tools self-sufficient enough that it could be distributed as it's own package, and just added as a dependency to this package?

I'm not saying you must do this, just offering an alternative to consider.
