### 2025-06-27 Lindt8/PHITS-Tools#5 @MicahGale
Improve testing
This is related to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8311).

I see that you are started creating a test-suite, which is a good start. However, this isn't using any industry standard tools, like `pytest` and is rather adhoc. I would recommend [reviewing the PyOpenSci testing guide](https://www.pyopensci.org/python-package-guide/tests/index.html). Right now it seems like you have around 1% test coverage? (see [attached coverage reports](https://github.com/user-attachments/files/20951891/htmlcov.zip)).

It seems like your test is trying to do a large integration test and handle logging and reporting. A testing framework, like `pytest`, can handle the logging and reporting for you. In general [unit tests are much more effective than integration testing](https://dl.acm.org/doi/10.1145/2896941.2896944)

I would recommend:

* [x] implement `pytest` as a testing framework.
* [x] implement #1, so that you can ensure that you are testing the local code, and not the installed code
* [x] create unit tests to achieve 80+% coverage
* [x] Focus testing a lot on "physics related methods", i.e., those that are calculating new tallies, combining tallies, etc.; really the functions where the math has an impact on the end user.
