### 2025-06-27 Lindt8/PHITS-Tools#6 @MicahGale
Create Community guidelines
This is related to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8311).

JOSS requires community guidelines to be included for:

1. Contributing to the software (I think a `CONTRIBUTING.md`)
2. reporting issues (Similarly `CONTRIBUTING.md` and possibly some GH issue templates)
3. Seeking support

I don't believe this explicitly requires a code of conduct, but could be considered as well. [Github has some resources for creating these](https://docs.github.com/en/site-policy/github-terms/github-community-guidelines).
