### 2025-07-01 a07aaa8
Create CONTRIBUTING.md
