+++ .github/ISSUE_TEMPLATE/problematic-phits-output-file-found.md
@@ -1,0 +1,37 @@
+---
+name: Problematic PHITS output file found
+about: Create a report to help us improve support for all PHITS outputs
+title: "[PROBLEM_INPUT]"
+labels: bug
+assignees: ''
+
+---
+
+**Describe the problem**
+A clear and concise description of what the problem is.
+
+**Problematic PHITS output file**
+Upload the problematic PHITS output file here. You may need to change the file extension to `.txt` to upload it here to a GitHub Issue.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Executed `PHITS_tools.py` via CLI with the following settings: '...' / via GUI with settings options shown in the screenshot below / as an imported module with function '....'
+2. See error
+
+**Error message**
+Include the error message returned when attempting to process the above PHITS output file with the settings specified above.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Your environment:**
+ - OS: [e.g. Windows 11, MacOS 15.5]
+ - PHITS version: [e.g. 3.34]
+ - PHITS parallelization used: [e.g. no / OMP / MPI / MPI+OMP]
+ - Python version [e.g. 3.12]
+
+**Additional context**
+Add any other context about the problem here.
