+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,36 @@
+---
+name: Bug report
+about: Create a report to help us improve code functionality
+title: "[BUG]"
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**Problematic PHITS output file or other data**
+Upload the problematic PHITS output file or other text data here, if required to reproduce the issue. You may need to change the file extension to `.txt` to upload it here to a GitHub Issue.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Your environment:**
+ - OS: [e.g. Windows 11, MacOS 15.5]
+ - PHITS version: [e.g. 3.34]
+ - PHITS parallelization used: [e.g. no / OMP / MPI / MPI+OMP]
+ - Python version [e.g. 3.12]
+
+**Additional context**
+Add any other context about the problem here.
