### 2025-07-01 Lindt8/PHITS-Tools#6 @Lindt8
Addressed in a07aaa86ee66fa73750bdac7c4347819bd476460 and 4d808fd325bde9971851b70b3394723d7d0368fb, added `CONTRIBUTING.md` and issue templates, respectively.

Thank you for the suggestion here.  I originally just had a couple of paragraphs on this at the bottom of `README.md`, but it is definitely better to have this information in its own `CONTRIBUTING.md` file.

I was also not aware of the GitHub Issue templates feature; I have set up some issue templates now.  Thanks for introducing that feature to me!

Let me know if you have any other concerns regarding this issue.  If not, I will close it.
