### 2025-07-01 Lindt8/PHITS-Tools#6 @MicahGale
Great, I think those are sufficient for closing this issue.

I noticed that you pushed these commits to `main` though. That's a good workflow for early development, but gets riskier as a project grows, and the project has a higher expectation of stability. I think it might be time for moving to a [branching workflow](https://www.geeksforgeeks.org/git/branching-strategies-in-git/), with some [protected branches](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-protected-branches/managing-a-branch-protection-rule). I personally prefer the "gitlab flow" where:

* `main` (or `master`) is only for deployed code and is protected
* `develop` is for completed features that haven't been deployed yet, and is also protected.
* "feature" branches that develop new features, bug fixes, etc.

This workflow makes it less likely buggy code ships, or that you break the code that people use while trying to add a new feature.
