### 2025-07-03 6f9cb6f
added functions for generating JSON and text files from list of materials dictionaries
