+++ MC_materials/manage_mc_materials.py
@@ -8,11 +8,11 @@ import re
 import os
 import csv
 import sys
-import pickle
+import json
 from pathlib import Path
+from PHITS_tools import Element_Z_to_Sym,fetch_MC_material
 
-
-def pnnl_lib_csv_to_dict(path_to_materials_compendium_csv=Path(Path.cwd(), 'materials_compendium.csv')):
+def pnnl_lib_csv_to_dict_list(path_to_materials_compendium_csv=Path(Path.cwd(), 'materials_compendium.csv')):
     '''
     Description:
         This function converts a CSV file of the PNNL
@@ -28,8 +28,7 @@ def pnnl_lib_csv_to_dict(path_to_materials_compendium_csv=Path(Path.cwd(), 'mate
         - `path_to_materials_compendium_csv` = string or Path object denoting the path to the "materials_compendium.csv" file
 
     Outputs:
-        - `mat` = a dictionary object of the extracted materials information
-    
+        - `mat_list` = a list of dictionary objects of the extracted materials information
     '''
 
     def is_line_not_empty(line):
@@ -43,7 +42,7 @@ def pnnl_lib_csv_to_dict(path_to_materials_compendium_csv=Path(Path.cwd(), 'mate
     f.close()
 
     # isolate just the lines we care about
-    mat_list = []
+    mat_dict_list = []
     matno = 1
     in_core_lines = False
     in_element_section = False
@@ -113,7 +112,7 @@ def pnnl_lib_csv_to_dict(path_to_materials_compendium_csv=Path(Path.cwd(), 'mate
                 in_mcnp_photons = False
                 in_mcnp_section = False
                 # wrap up entry
-                mat_list.append(mat)
+                mat_dict_list.append(mat)
             else:
                 mat['photons']['weight fraction']['ZA'].append(line[1])
                 mat['photons']['weight fraction']['value'].append(line[2])
@@ -122,6 +121,325 @@ def pnnl_lib_csv_to_dict(path_to_materials_compendium_csv=Path(Path.cwd(), 'mate
                 mat['photons']['atom density']['ZA'].append(line[5])
                 mat['photons']['atom density']['value'].append(line[6])
     
-    return mat
+    return mat_dict_list
 
-#def material_lib_dict_to_text(path_to_materials_compendium_csv):
+def materials_dict_list_to_json(mat_list,json_filepath=Path(Path.cwd(), 'materials_compendium.json')):
+    '''
+    Description:
+        This function converts the materials list of dictionaries into a JSON file.
+    
+    Inputs:
+        - `mat_dict_list` = a list of dictionary objects of the extracted materials information
+
+    Outputs:
+        - `None`; the materials data will be saved to `json_filepath`.
+    '''
+    with open(json_filepath, 'w') as f:
+        json.dump(mat_list, f)
+    return None
+
+def materials_json_to_dict_list(json_filepath=Path(Path.cwd(), 'materials_compendium.json')):
+    '''
+    Description:
+        This function converts the materials list of dictionaries into a JSON file.
+    
+    Inputs:
+        - `json_filepath` = string or Path object denoting the path to the JSON materials file
+
+    Outputs:
+        - `mat_list` = a list of dictionary objects of the extracted materials information
+    '''
+    with open(json_filepath, "r") as f:
+        mat_list = json.load(f)
+    return mat_list
+
+
+def write_descripive_material_entry(mat,mati):
+    '''
+    Description:
+        Generate a descriptive text block for a material dictionary object.
+
+    Inputs:
+        - `mat` = a dictionary object formatted like all entries in the library
+        - `mati` = integer specifying material number within the database
+
+    Outputs:
+        - `entry_text` = a string of formatted text with all information about the material
+    '''
+    banner_width = 80
+
+    if 'PNNL_number' in mat:  # entries are strings already
[…]
