+++ MC_materials/PNNL_materials_compendium.txt
@@ -7,6 +7,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   1   : A-150 Tissue-Equivalent Plastic (A150TEP) 
+  Source = PNNL-15870 Rev. 1 
   Formula = - 
   Molecular weight (g/mole) = - 
   Density (g/cm3) = 1.127000 
@@ -42,6 +43,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   2   : Acetone 
+  Source = PNNL-15870 Rev. 1 
   Formula = C3H6O 
   Molecular weight (g/mole) = 58.07914 
   Density (g/cm3) = 0.789900 
@@ -68,6 +70,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   3   : Acetylene 
+  Source = PNNL-15870 Rev. 1 
   Formula = C2H2 
   Molecular weight (g/mole) = 26.03728 
   Density (g/cm3) = 0.001097 
@@ -91,6 +94,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   4   : Air (dry, near sea level) 
+  Source = PNNL-15870 Rev. 1 
   Formula = - 
   Molecular weight (g/mole) = - 
   Density (g/cm3) = 0.001205 
@@ -120,6 +124,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   5   : Alanine 
+  Source = PNNL-15870 Rev. 1 
   Formula = - 
   Molecular weight (g/mole) = - 
   Density (g/cm3) = 1.420000 
@@ -149,6 +154,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   6   : Aluminum 
+  Source = PNNL-15870 Rev. 1 
   Formula = Al 
   Molecular weight (g/mole) = 26.981538 
   Density (g/cm3) = 2.698900 
@@ -169,6 +175,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   7   : Aluminum Oxide 
+  Source = PNNL-15870 Rev. 1 
   Formula = Al2O3 
   Molecular weight (g/mole) = 101.961276 
   Density (g/cm3) = 3.970000 
@@ -192,6 +199,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   8   : Aluminum, alloy 2024-O 
+  Source = PNNL-15870 Rev. 1 
   Formula = - 
   Molecular weight (g/mole) = - 
   Density (g/cm3) = 2.780000 
@@ -236,6 +244,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   9   : Aluminum, alloy 2090-T83 
+  Source = PNNL-15870 Rev. 1 
   Formula = - 
   Molecular weight (g/mole) = - 
   Density (g/cm3) = 2.590000 
@@ -286,6 +295,7 @@ https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1
 
 ********************************************************************************
   10  : Aluminum, alloy 3003 
+  Source = PNNL-15870 Rev. 1 
[…]
