### 2025-07-07 3067cbe
continued implementation of MC materials management
