+++ PHITS_tools.py
@@ -54,6 +54,7 @@ functions return the data objects they produce for your own further analyses.
 - `ICRP116_effective_dose_coeff`    : returns effective dose conversion coefficient of a mono-energetic particle of some species and some geometry; does coefficients are those in [ICRP 116](https://doi.org/10.1016/j.icrp.2011.10.001)
 - `merge_dump_file_pickles`         : merge multiple dump file outputs into a single file (useful for dumps in MPI runs)
 - `ZZZAAAM_to_nuclide_plain_str`    : returns a nuclide plaintext string for a given "ZZZAAAM" number (1000Z+10A+M)
+- `nuclide_plain_str_to_ZZZAAAM`    : returns a "ZZZAAAM" number (1000Z+10A+M) for a given nuclide plaintext string 
 - `nuclide_plain_str_to_latex_str`  : convert a plaintext string for a nuclide to a LaTeX formatted raw string
 - `Element_Z_to_Sym`                : return an elemental symbol string given its proton number Z
 - `Element_Sym_to_Z`                : returns an atomic number Z provided the elemental symbol
@@ -3296,14 +3297,16 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
 
 
 
-def fetch_MC_material(matid=None,matname=None,matsource=None,concentration_type=None,particle=None):
+def fetch_MC_material(matid=None,matname=None,matsource=None,concentration_type=None,particle=None,matdict=None,
+                      database_filename='Compiled_MC_materials'):
     '''
     Description:
         Returns a materials definition string formatted for use in PHITS or MCNP (including a density estimate);
         most available materials are those found in [PNNL-15870 Rev. 1](https://www.osti.gov/biblio/1023125).
 
     Dependencies:
-        - PYTHONPATH environmental variable must be set and one entry must contain the directory
+        - Either PHITS Tools must be installed via `pip` (which automatically handles this) or your
+                PYTHONPATH environmental variable must be set and one entry must contain the directory
                 which contains the vital "MC_materials/Compiled_MC_materials.pkl" file.
 
     Inputs:
@@ -3317,34 +3320,62 @@ def fetch_MC_material(matid=None,matname=None,matsource=None,concentration_type=
                 `'atom fraction'` (default if a chemical formula is present; e.g. "Ethane" (Formula = C2H6)) to be returned
        - `particle` = selection of whether natural (`'photons'`, default) or isotopic (`'neutrons'`) elements are used
                 Note that if "enriched" or "depleted" appears in the material's name, particle=`'neutrons'` is set automatically.
-
+       - `matdict` = dictionary object of same format as entries in materials database; if this is provided, 
+                `matname` and `database_filename` are ignored, using the provided material dictionary entry instead.
+       - `database_filename` = (D=`'Compiled_MC_materials'`) File basename of the database file to be pulled from.
+       
     Outputs:
        - `mat_str` = string containing the material's information, ready to be inserted directly into a PHITS/MCNP input file
     '''
     import os
+    import json
+    import pkgutil
     import pickle
     if not matid and not matname:
         print('Either "matid" or "matname" MUST be defined')
         return None
-
-    # First, locate and open materials library
-    try:
-        user_paths = os.environ['PYTHONPATH'].split(os.pathsep)
-        lib_file = None
-        for i in user_paths:
-            if 'phits_tools' in i.lower() or 'phits-tools' in i.lower():
-                lib_file = i + r"\MC_materials\Compiled_MC_materials"
-        if not lib_file:
[…]
