+++ MC_materials/manage_mc_materials.py
@@ -10,7 +10,7 @@ import csv
 import sys
 import json
 from pathlib import Path
-from PHITS_tools import Element_Z_to_Sym,fetch_MC_material
+from PHITS_tools import Element_Z_to_Sym, fetch_MC_material, nuclide_plain_str_to_ZZZAAAM
 
 def pnnl_lib_csv_to_dict_list(path_to_materials_compendium_csv=Path(Path.cwd(), 'materials_compendium.csv')):
     '''
@@ -130,12 +130,14 @@ def materials_dict_list_to_json(mat_list,json_filepath=Path(Path.cwd(), 'materia
     
     Inputs:
         - `mat_dict_list` = a list of dictionary objects of the extracted materials information
+        - `json_filepath` = string or Path object denoting the path to the JSON materials file to be written
 
     Outputs:
         - `None`; the materials data will be saved to `json_filepath`.
     '''
     with open(json_filepath, 'w') as f:
         json.dump(mat_list, f)
+    print('Materials library file written:', json_filepath)
     return None
 
 def materials_json_to_dict_list(json_filepath=Path(Path.cwd(), 'materials_compendium.json')):
@@ -316,6 +318,7 @@ def write_descriptive_file(mat_list,lib_filepath=Path(Path.cwd(),'MC_materials.t
     f = open(lib_filepath, 'w+')
     f.write(lib_text)
     f.close()
+    print('Materials library file written:', lib_filepath)
 
     # Make an index
     if write_index_file:
@@ -375,6 +378,7 @@ def write_MC_formated_files(mat_list,lib_filepath=Path(Path.cwd(),'MC_materials.
             f = open(fpath, 'w+')
             f.write(lib_text)
             f.close()
+            print('Materials library file written:', fpath)
     return None
 
 
@@ -396,7 +400,7 @@ def write_general_MC_file(mat_list,lib_filepath=Path(Path.cwd(),'MC_materials_ge
             If left blank (`''`), a default header string as defined at the start of this function will be used.
         
     Outputs:
-        - - `None`; the materials text data will be saved to `lib_filepath`.
+        - `None`; the materials text data will be saved to `lib_filepath`.
     '''
     cc = comment_char 
     if header_text=='':
@@ -410,16 +414,81 @@ def write_general_MC_file(mat_list,lib_filepath=Path(Path.cwd(),'MC_materials_ge
         header_text += cc + '  ' + 'The sources for other entries are specified.' + '\n'
     lib_text = header_text
     for i in range(len(mat_list)):
-        print('TODO: UPDATE CALLED FUNCTION WITH JSON DATABASE')
-        lib_text += fetch_MC_material(i + 1)
+        lib_text += fetch_MC_material(i + 1, matdict=mat_list[i])
     # save file
     f = open(lib_filepath, 'w+')
     f.write(lib_text)
     f.close()
+    print('Materials library file written:', lib_filepath)
     return None
 
 
+def update_materials_database_files(json_filepath,name,mat_str,matid=None,density=None,source=None,
+                                    source_short=None,formula=None,molecular_weight=None,total_atom_density=None,
+                                    new_database_base_name=None,update_descriptive_file=True,
+                                    update_MC_formated_files=True,update_general_MC_file=True):
+    '''
+    Description:
+    
+        Add or modify a material in a MC materials database JSON file (optionally updating text files too).
 
+    Inputs:
+        (required)
+        
+        - `json_filepath` = string or Path object denoting the path to the JSON materials file
+        - `name` = string of the name of the material to be added / updated.  See the "Notes" section further below for 
+                more information on how entry identification is handled.
+        - `mat_str` = string designating the composition of the material.  This should be provided as a series of 
+                alternating nuclide IDs and concentrations. 
[…]
