### 2025-07-08 b6bdbd5
added main function for updating materials database files
