+++ MC_materials/manage_mc_materials.py
@@ -1,5 +1,30 @@
 '''
-This module is used for managing the MC materials database.
+
+This submodule is used for managing the MC materials database found in the "MC_materials" directory, which is located 
+in the same directory as PHITS_tools.py.  The MC materials database consists of a list of materials with the necessary
+compositional information needed for particle transport simulations, and it uses the PNNL 
+Compendium of Material Composition Data for Radiation Transport Modeling (Rev. 1), PNNL-15870 Rev. 1, document as
+its foundational source, on top of which additional materials can be added freely.
+
+The MC materials database files created and managed by this module's functions are also those retrieved by the 
+[`fetch_MC_material()`](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.fetch_MC_material) function in PHITS Tools.
+
+
+
+### Main function for updating MC materials database files
+
+- `update_materials_database_files` : add/modify materials entries, create new database files
+
+### Additional helper functions
+
+- `pnnl_lib_csv_to_dict_list`       : converts the PNNL materials_compendium.csv file to a list of dictionaries
+- `materials_dict_list_to_json`     : save a list of dictionaries to a JSON file
+- `materials_json_to_dict_list`     : read a JSON file into a list of dictionaries
+- `write_descripive_material_entry` : write a plaintext material entry for a material's dictionary object
+- `write_mc_material_entry`         : write a PHITS/MCNP-formatted text entry for a material's dictionary object
+- `write_descriptive_file`          : write a file of plaintext material entries for all material dictionaries in a list
+- `write_MC_formated_files`         : write 4 files of PHITS/MCNP-formatted material entries for all material dictionaries in a list
+- `write_general_MC_file`           : write a file of PHITS/MCNP-formatted material entries (choosing "most sensible" of 4 for each)
 
 '''
 
@@ -10,7 +35,429 @@ import csv
 import sys
 import json
 from pathlib import Path
-from PHITS_tools import Element_Z_to_Sym, fetch_MC_material, nuclide_plain_str_to_ZZZAAAM
+from PHITS_tools import Element_Z_to_Sym, Element_Sym_to_Z, fetch_MC_material, nuclide_plain_str_to_ZZZAAAM
+
+
+
+
+def update_materials_database_files(json_filepath,name,mat_str,matid=None,density=None,source=None,
+                                    source_short=None,formula=None,molecular_weight=None,total_atom_density=None,
+                                    new_database_base_name=None,update_descriptive_file=True,
+                                    update_MC_formated_files=True,update_general_MC_file=True,save_backup_list=True):
+    '''
+    Description:
+    
+        Add or modify a material in a MC materials database JSON file (optionally updating text files too).
+
+    Inputs:
+        (required)
+        
+        - `json_filepath` = string or Path object denoting the path to the JSON materials file. If a string is provided 
+                that does not end in `'.json'`, this function will search to see if a JSON file of the same basename exists
+                in a directory called "MC_materials" inside the same directory as the PHITS_tools.py module (which requires
+                either PHITS Tools to be installed via `pip` or locatable within your PYTHONPATH system variable).
+                The JSON libraries that are shipped with PHITS Tools by default are titled `'Compiled_MC_materials'` 
+                and `'PNNL_materials_compendium'`.
+        - `name` = string of the name of the material to be added / updated.  See the "Notes" section further below for 
+                more information on how entry identification is handled.
+        - `mat_str` = string designating the composition of the material.  This should be provided as a series of 
[…]
