### 2025-07-10 ce57e8a
added lots of documentation and function for local MC materials directory
