+++ MC_materials/PNNL_materials_compendium_by_weight_fraction_for_photons.txt
@@ -1,0 +1,5200 @@
+This file was assembled from the Compendium of Material Composition Data for
+Radiation Transport Modeling (Rev. 1), PNNL-15870 Rev. 1, published by the
+Pacific Northwest National Laboratory.
+This file seeks to just compile the core information of the compendium in an
+easily accessible plain-text format.  The full document can be found at: 
+https://www.pnnl.gov/main/publications/external/technical_reports/PNNL-15870Rev1.pdf
+
+$************************************************************
+$  1   : A-150 Tissue-Equivalent Plastic (A150TEP) 
+$  Source = PNNL-15870 Rev. 1 
+$  Density (g/cm3) = 1.127000 
+$  Total atom density (atoms/b-cm) = 1.169E-01 
+$  Composition by weight fraction 
+M1         1000  -0.101327       $  H
+           6000  -0.775501       $  C
+           7000  -0.035057       $  N
+           8000  -0.052316       $  O
+           9000  -0.017422       $  F
+          20000  -0.018378       $  Ca
+$************************************************************
+
+$************************************************************
+$  2   : Acetone 
+$  Source = PNNL-15870 Rev. 1 
+$  Formula = C3H6O 
+$  Molecular weight (g/mole) = 58.07914 
+$  Density (g/cm3) = 0.789900 
+$  Total atom density (atoms/b-cm) = 8.190E-02 
+$  Composition by weight fraction 
+M2         1000  -0.104122       $  H
+           6000  -0.620405       $  C
+           8000  -0.275473       $  O
+$************************************************************
+
+$************************************************************
+$  3   : Acetylene 
+$  Source = PNNL-15870 Rev. 1 
+$  Formula = C2H2 
+$  Molecular weight (g/mole) = 26.03728 
+$  Density (g/cm3) = 0.001097 
+$  Total atom density (atoms/b-cm) = 1.015E-04 
+$  Composition by weight fraction 
+M3         1000  -0.077418       $  H
+           6000  -0.922582       $  C
+$************************************************************
+
+$************************************************************
+$  4   : Air (dry, near sea level) 
+$  Source = PNNL-15870 Rev. 1 
+$  Density (g/cm3) = 0.001205 
+$  Total atom density (atoms/b-cm) = 4.988E-05 
+$  Composition by weight fraction 
+M4         6000  -0.000124       $  C
+           7000  -0.755268       $  N
+           8000  -0.231781       $  O
+          18000  -0.012827       $  Ar
+$************************************************************
+
+$************************************************************
+$  5   : Alanine 
+$  Source = PNNL-15870 Rev. 1 
+$  Density (g/cm3) = 1.420000 
+$  Total atom density (atoms/b-cm) = 1.248E-01 
+$  Composition by weight fraction 
+M5         1000  -0.079190       $  H
+           6000  -0.404439       $  C
+           7000  -0.157213       $  N
+           8000  -0.359159       $  O
+$************************************************************
+
+$************************************************************
+$  6   : Aluminum 
+$  Source = PNNL-15870 Rev. 1 
+$  Formula = Al 
+$  Molecular weight (g/mole) = 26.981538 
+$  Density (g/cm3) = 2.698900 
+$  Total atom density (atoms/b-cm) = 6.024E-02 
+$  Composition by weight fraction 
+M6        13000  -1.000000       $  Al
+$************************************************************
+
+$************************************************************
+$  7   : Aluminum Oxide 
+$  Source = PNNL-15870 Rev. 1 
+$  Formula = Al2O3 
+$  Molecular weight (g/mole) = 101.961276 
+$  Density (g/cm3) = 3.970000 
+$  Total atom density (atoms/b-cm) = 1.172E-01 
[…]
