### 2025-07-11 fa2e03e
more documentation, updated relevant materials paths and missing file handling
