+++ MC_materials/manage_mc_materials.py
@@ -52,7 +52,7 @@ Executing `update_materials_database_files` with information with a new material
         of being overwritten/deleted when updating PHITS Tools to a new version via `pip install PHITS-Tools --upgrade`.
 2. The JSON database file will be updated with the new/updated materials information provided to the function.
     - If `new_database_base_name` is not left as `None`, instead of updating the existing database, a new databased using 
-        `new_database_base_name` as its `SET_NAME` is created from the existing database, applying the specified 
+        `new_database_base_name` as its "`SET_NAME`" is created from the existing database, applying the specified 
         materials addition/updates to it too.  This new database is then used for the remaining operations.
     - If `save_backup_list=True`, a timestamped copy of the updated/new JSON file is placed in the `MC_materials/backup_lists/` directory.
 3. (if `update_MC_formated_files=True`) `write_descriptive_file()` is called, updating/creating `SET_NAME.txt` and `SET_NAME_index.txt`
@@ -105,7 +105,8 @@ update_materials_database_files(json_filepath,
                                 update_descriptive_file=True,
                                 update_MC_formated_files=True,
                                 update_general_MC_file=True,
-                                save_backup_list=True)
+                                save_backup_list=True,
+                                prefer_user_data_folder=True)
 ```
 
 To just add a new material to an existing database, in this case adding Martian regolith to 
@@ -132,7 +133,8 @@ update_materials_database_files(json_filepath,
                                 update_descriptive_file=True,
                                 update_MC_formated_files=True,
                                 update_general_MC_file=True,
-                                save_backup_list=True)
+                                save_backup_list=True,
+                                prefer_user_data_folder=True)
 ```
 
 And, finally, if you wish to recreate the JSON file and text files for the initial PNNL database `'PNNL_materials_compendium'`, 
@@ -228,6 +230,12 @@ def update_materials_database_files(json_filepath,name,mat_str,matid=None,densit
         - `save_backup_list` = (D=`True`) Boolean denoting whether an extra (timestamped) copy of the produced JSON 
                 materials list database file should also be saved separately in a "backup_lists" directory located in 
                 the same directory as `json_filepath`.
+        - `prefer_user_data_folder` = (D=`True`) Boolean denoting whether this function should prioritize the local 
+                MC materials databases in your local [`"$HOME`](https://docs.python.org/3/library/pathlib.html#pathlib.Path.home)`/.PHITS-Tools/"` 
+                directory over those in the PHITS Tools distribution. If the local user directory does not yet exist, 
+                setting this to `True` will cause it to be created with a call of `setup_local_mc_materials_directory`.
+                This local directory is created to allow creation and curation of user MC materials databases locally without fear
+                of being overwritten/deleted when updating PHITS Tools to a new version via `pip install PHITS-Tools --upgrade`.
 
     Outputs:
     
@@ -239,7 +247,10 @@ def update_materials_database_files(json_filepath,name,mat_str,matid=None,densit
         
         Entries in the database are uniquely identified by their `matid` or the combination of `name` and `source`.
         If provided with a `matid` or `name` and `source` combination already present within the database, a prompt
-        will appear asking whether the existing entry should be overwritten.
+        window, as pictured below (in this case, with `matid` specified), will appear showing both the old and new versions 
+        of the entry and asking for confirmation on whether the existing entry should be overwritten with the new one.
+        
+        ![](https://github.com/Lindt8/PHITS-Tools/blob/main/docs/update_MC_material_window.png?raw=true "Overwrite MC materials confirmation window")
[…]
