+++ PHITS_tools.py
@@ -3298,7 +3298,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
 
 
 def fetch_MC_material(matid=None,matname=None,matsource=None,concentration_type=None,particle=None,matdict=None,
-                      database_filename='Compiled_MC_materials'):
+                      database_filename='Compiled_MC_materials',prefer_user_data_folder=True):
     '''
     Description:
         Returns a materials definition string formatted for use in PHITS or MCNP (including a density estimate);
@@ -3310,14 +3310,14 @@ def fetch_MC_material(matid=None,matname=None,matsource=None,concentration_type=
     Dependencies:
         - Either PHITS Tools must be installed via `pip` (which automatically handles this) or your
                 PYTHONPATH environmental variable must be set and one entry must contain the directory
-                which contains the vital "MC_materials/Compiled_MC_materials.json" file.
+                which contains PHITS Tools and the vital "MC_materials/Compiled_MC_materials.json" file.
 
     Inputs:
        (required to enter `matid` OR `matname`, with `matid` taking priority if conflicting)
 
-       - `matid` = ID number in the "Compiled_MC_materials" file
-       - `matname` = exact name of material in "Compiled_MC_materials" file
-       - `matsource` = exact source of material in "Compiled_MC_materials" file, only used when multiple
+       - `matid` = ID number in the database specified by `database_filename`
+       - `matname` = exact name of material in the database specified by `database_filename`
+       - `matsource` = exact source of material in the database specified by `database_filename`, only used when multiple
                 materials have identical names
        - `concentration_type` = selection between `'weight fraction'` (default if no chemical formula is present in database; e.g., "Incoloy-800") and
                 `'atom fraction'` (default if a chemical formula is present; e.g. "Ethane" (Formula = C2H6)) to be returned
@@ -3326,7 +3326,15 @@ def fetch_MC_material(matid=None,matname=None,matsource=None,concentration_type=
        - `matdict` = dictionary object of same format as entries in materials database; if this is provided, 
                 `matname` and `database_filename` are ignored, using the provided material dictionary entry instead.
        - `database_filename` = (D=`'Compiled_MC_materials'`) File basename of the database file to be pulled from.
-       
+       - `prefer_user_data_folder` = (D=`True`) Boolean denoting whether this function should prioritize the local 
+                MC materials databases in your local [`"$HOME`](https://docs.python.org/3/library/pathlib.html#pathlib.Path.home)`/.PHITS-Tools/"` 
+                directory over those in the PHITS Tools distribution, if the local user directory exists. For information 
+                on creation and modification of this local MC materials directory, see the [**`PHITS_tools.manage_mc_materials` submodule documentation**](https://lindt8.github.io/PHITS-Tools/docs/manage_mc_materials.html).
+                If `False` (or if `True` but no local user MC_materials directory exists), the MC_materials directory 
+                distributed with PHITS Tools will be used instead, searching for the location first using 
+                [`pkgutil.get_loader`](https://docs.python.org/3/library/pkgutil.html#pkgutil.get_loader)`("PHITS_tools").get_filename()`
+                then, failing that, the PYTHONPATH environmental variable.
+                
     Outputs:
        - `mat_str` = string containing the material's information, ready to be inserted directly into a PHITS/MCNP input file
     '''
@@ -3339,26 +3347,40 @@ def fetch_MC_material(matid=None,matname=None,matsource=None,concentration_type=
         return None
     
     if matdict is None:
-        # First, locate and open materials library
-        try:
-            lib_file = None
-            try: # First, check MC_materials folder distributed with PHITS Tools
[…]
