### 2025-07-11 8b7f7fa
fixed docs bullet point formatting
