### 2025-07-11 0eaa269
added build files (1.6.0-beta.1)
