### 2025-07-15 1627e94
renamed beta release to match PyPA specifications
