### 2025-07-15 81c3dad
remove "generated" files from dist via .gitattributes and pyproject.toml (MC materials .txt files now generated locally)
