+++ pyproject.toml
@@ -24,3 +24,17 @@ dependencies = ["numpy>=2.2.1", "pandas>=2.2.3", "scipy>=1.14.1", "matplotlib>=3
 Documentation = "https://github.com/Lindt8/PHITS-Tools#readme"
 Issues = "https://github.com/Lindt8/PHITS-Tools/issues"
 Source = "https://github.com/Lindt8/PHITS-Tools"
+
+[tool.hatch.build]
+exclude = [
+  "/.github",
+  "/.idea",
+  "/prepare_release.py",
+  "/test/*.pdf",
+  "/example/*.pdf",
+  "/example/*.pickle*",
+  "/example/*.png",
+  "/example/*.out",
+  "/example/*.eps",
+  "/MC_materials/*.txt"
+]
