### 2025-07-15 a1cddf7
remove "generated" files from dist via .gitattributes and pyproject.toml (MC materials .txt files now generated locally)
