### 2025-07-16 Lindt8/PHITS-Tools#2 @Lindt8
Sorry for the delay in getting back to this. And actually, this is not really something I had considered, but yes, DCHAIN Tools could be distributed as a standalone package, though, for now, I will keep it as it is as a submodule.

My rationale behind structuring it this way was for a few reasons:

1. Structuring DCHAIN-Tools as a submodule of PHITS-Tools mirrors how the actual DCHAIN code is structured like a submodule within the PHITS distribution; users are already accustomed to viewing DCHAIN more like a submodule belonging to the greater PHITS package rather than as a separate code.  (While we refer to it as "DCHAIN" for brevity, the code's official title is "DCHAIN-PHITS".)
2. This somewhat mirrors how PHITS and DCHAIN themselves interact. While DCHAIN technically _can_ be used standalone, I would wager 99.9% of its use is with its coupling to PHITS.
3. Furthermore, DCHAIN is not distributed standalone; it can only be obtained along with a PHITS distribution.
4. So, given that, in theory, 100% of DCHAIN users are also PHITS users and that they are almost exclusively using the two codes in tandem, it made sense to me to just distribute both PHITS-Tools and DCHAIN-Tools within the same package and only need to worry about maintaining a single package.

I was originally debating retiring the DCHAIN-Tools repository and just moving it entirely to PHITS-Tools, but I decided against that as it felt like a step backwards in organization, not to mention that the DCHAIN-Tools repository is already linked in a number of places and wanting to avoid extra link hopping to visit the up-to-date version of it.  DCHAIN-Tools does not see frequent updates anymore, so it is not much of an inconvenience to just `git submodule update --remote` in PHITS-Tools on the occasion I have pushed an update to DCHAIN-Tools.

To alleviate potential usage confusion though, I should update the documentation for DCHAIN-Tools to reflect its "modern" intended use as a submodule of PHITS-Tools rather than just a standalone module as it had been for some years prior, instructing users who have installed PHITS-Tools to just access it via `import PHITS_tools.dchain_tools as dchain_tools`. This should also be more prominent in the PHITS Tools documentation... I will make those changes.

But yes, there are a few downsides with git submodules.  For one, while the PyPI wheel and tarball build files do contain all of the DCHAIN-Tools repo files, the source ZIP and tarball files with PHITS-Tools GitHub releases do not.  And when cloning the repo, there is the extra complication of needing to add the `--recurse-submodules` flag to get the DCHAIN-Tools submodule files included too.
`gh repo clone https://github.com/Lindt8/PHITS-Tools -- --recurse-submodules`
