### 2025-07-16 Lindt8/PHITS-Tools#2 @MicahGale
Sounds like a good plan.

Just some notes:

1. `from PHITS_tools import dchain_tools` will have the same effect.
2. I think just documenting somewhere for new developers they need that flag would be a good idea. Maybe in `CONTRIBUTING.md` (I forgot the exact file name) of `PHITS_tools`.
