### 2025-07-16 904bbda
improved documentation for submodule usage
