### 2025-07-17 82880bf
improved documentation for submodule usage
