### 2025-07-17 b84dca9
add command line script entry points, resolves #1
