+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "PHITS-Tools"
-version = "1.6.0b2"
+version = "1.6.0b3"
 description = 'Python module for parsing and post-processing output from the PHITS code'
 readme = "README.md"
 license = "MPL-2.0"
@@ -19,6 +19,14 @@ classifiers = [
   "Operating System :: OS Independent",
 ]
 dependencies = ["numpy>=2.2.1", "pandas>=2.2.3", "scipy>=1.14.1", "matplotlib>=3.10.0", "seaborn>=0.13.2"]
+
+[project.scripts]
+PHITS-Tools = "PHITS_tools:run_PHITS_tools_CLI_or_GUI"
+PHITS_tools = "PHITS_tools:run_PHITS_tools_CLI_or_GUI"
+phits-tools = "PHITS_tools:run_PHITS_tools_CLI_or_GUI"
+
+[project.gui-scripts]
+PHITS-Tools-GUI = "PHITS_tools:run_PHITS_tools_GUI"
 
 [project.urls]
 Documentation = "https://github.com/Lindt8/PHITS-Tools#readme"
