+++ tests/test_nuclide_formatting.py
@@ -1,0 +1,122 @@
+import pytest 
+from PHITS_tools import element_Z_to_symbol, element_symbol_to_Z, kfcode_to_common_name
+from PHITS_tools import ZZZAAAM_to_nuclide_plain_str, nuclide_plain_str_to_ZZZAAAM, nuclide_plain_str_to_latex_str
+from PHITS_tools import Element_Z_to_Sym, Element_Sym_to_Z
+
+def test_element_Z_to_symbol():
+    assert element_Z_to_symbol(1) == 'H'
+    assert element_Z_to_symbol(92) == 'U'
+    assert element_Z_to_symbol(7) == 'N'
+    assert element_Z_to_symbol(52) == 'Te'
+    assert element_Z_to_symbol(45) == 'Rh'
+    assert element_Z_to_symbol(118) == 'Og'
+    assert element_Z_to_symbol(0) == 'n'
+    assert element_Z_to_symbol(-1) is None
+    assert element_Z_to_symbol(119) is None
+
+def test_element_symbol_to_Z():
+    assert element_symbol_to_Z('H') == 1
+    assert element_symbol_to_Z('U') == 92
+    assert element_symbol_to_Z('N') == 7
+    assert element_symbol_to_Z('Te') == 52
+    assert element_symbol_to_Z('Rh') == 45
+    assert element_symbol_to_Z('Og') == 118
+    assert element_symbol_to_Z('XX') == 0
+    assert element_symbol_to_Z('test') == -1
+    assert element_symbol_to_Z('AA') == -1
+    assert element_symbol_to_Z('oG') == 118
+
+def test_kfcode_to_common_name():
+    assert kfcode_to_common_name(2112) == 'neutron'
+    assert kfcode_to_common_name(2212) == 'proton'
+    assert kfcode_to_common_name(-13) == 'muon+'
+    assert kfcode_to_common_name(8000016) == 'O-16'
+    assert kfcode_to_common_name(118000294) == 'Og-294'
+    assert kfcode_to_common_name(1000001) == 'H-1'
+    assert kfcode_to_common_name(1000000) == '1000000'
+    assert kfcode_to_common_name(123456) == '123456'
+
+def test_ZZZAAAM_to_nuclide_plain_str():
+    assert ZZZAAAM_to_nuclide_plain_str(80160) == 'O-16'
+    assert ZZZAAAM_to_nuclide_plain_str(8016, ZZZAAA=True) == 'O-16'
+    assert ZZZAAAM_to_nuclide_plain_str(80160, include_Z=True) == '8-O-16'
+    assert ZZZAAAM_to_nuclide_plain_str(430991) == 'Tc-99m1'
+    assert ZZZAAAM_to_nuclide_plain_str(430991, include_Z=True) == '43-Tc-99m1'
+
+def test_nuclide_plain_str_to_ZZZAAAM():
+    assert nuclide_plain_str_to_ZZZAAAM('O-16') == 80160
+    assert nuclide_plain_str_to_ZZZAAAM('O-nat') == 80000
+    assert nuclide_plain_str_to_ZZZAAAM('O 16') == 80160
+    assert nuclide_plain_str_to_ZZZAAAM('O_16') == 80160
+    assert nuclide_plain_str_to_ZZZAAAM('O16') == 80160
+    assert nuclide_plain_str_to_ZZZAAAM('16O') == 80160
+    assert nuclide_plain_str_to_ZZZAAAM('16-O') == 80160
+    assert nuclide_plain_str_to_ZZZAAAM('natO') == 80000
[…]
