+++ PHITS_tools.py
@@ -3945,13 +3945,17 @@ def nuclide_plain_str_to_ZZZAAAM(nuc_str):
 
     Inputs:
        - `nuc_str` = string to be converted; a huge variety of formats are supported, but they all must follow the following rules:
+           + `nuc_str` must begin with either the atomic mass number or the elemental symbol.
+           + `nuc_str` should NOT contain the atomic/proton number (Z).
            + Isomeric/metastable state characters must always immediately follow the atomic mass characters.
-               Isomeric state labels MUST either:
-               - (1) be a single lower-case character OR
-               - (2) begin with any non-numeric character and end with a number
-           + Atomic mass numbers must be nonnegative integers OR the string "nat" (in which case no metastable states can be written)
-           + Elemental symbols MUST begin with an upper-case character
-
+               Isomeric state labels must either:
+               - (1) be a single lower-case character in `['g','m','n','o','p','q']` OR
+               - (2) be `'m'` followed by a number from 1 to 5, in `['m1','m2','m3','m4','m5']`
+           + Atomic mass numbers must be nonnegative integers OR the string `"nat"` (in which case no metastable states 
+             can be written and A=0); if omitted, `"nat"` is assumed.
+           + Elemental symbols must begin with an upper-case character
+               - `'n'`, `'p'`, `'d'`, and `'t'` can also be specified for neutron, proton, deuteron, and triton, respectively.
+           + Space `' '`, hyphen `'-'`, and underscore `'_'` can be used anywhere in `nuc_str`; they will be ignored. 
 
     Outputs:
         - ZZZAAAM integer
@@ -3962,20 +3966,29 @@ def nuclide_plain_str_to_ZZZAAAM(nuc_str):
     for dc in delete_characters_list:
         nuc_str = nuc_str.replace(dc,'')
 
+    if 'nat' in nuc_str:
+        print('WARNING: specifying natural abundances via "nat" sets A=0 in the ZZZAAAM integer')
+        nuc_str = nuc_str.replace('nat','0')
+        # print('Must specify a specific nuclide, not natural abundances')
+        # return None
+
     # determine which characters are letters versus numbers
     isalpha_list = []
     isdigit_list = []
     for c in nuc_str:
         isalpha_list.append(c.isalpha())
         isdigit_list.append(c.isdigit())
-
+    
+    if not any(isdigit_list):
+        print('WARNING: No mass value is present in the provided string, assuming natural abundances.')
+        nuc_str += '0'
+        isalpha_list.append(False)
+        isdigit_list.append(True)
+    
     symbol = ''
     mass = ''
     isost = ''
 
-    if 'nat' in nuc_str:
-        print('Must specify a specific nuclide, not natural abundances')
-        return None
 
     # string MUST begin with either mass number or elemental symbol
     if isdigit_list[0]: # mass first
@@ -4033,15 +4046,21 @@ def nuclide_plain_str_to_ZZZAAAM(nuc_str):
         symbol = isost
         isost = ''
 
-
-    if symbol == 'n':
-        Z = 0
-    elif symbol == 'p' or symbol == 'd' or symbol == 't':
-        Z = 1
+    if symbol in ['n', 'p', 'd', 't']:
+        if symbol == 'n':
+            Z, A = 0, 1
+        elif symbol == 'p':
+            Z, A = 1, 1
+        elif symbol == 'd': 
+            Z, A = 1, 2
+        elif symbol == 't':
+            Z, A = 1, 3
     else:
         Z = element_symbol_to_Z(symbol)
-
-    A = int(mass)
+        if Z == -1:
[…]
