### 2025-07-21 9b5d53a
unit tests for tally function
