### 2025-07-21 c0c7b5c
unit tests for rebinner function
