### 2025-07-21 bf2b83f
unit tests for find and ICRP116 dose coeff functions
