### 2025-07-21 947961b
unit tests for is_number function
