### 2025-07-22 419454b
restructured repo for import from root and removed duplicate *.py files, resolves #1
