+++ src/PHITS_tools/__init__.py
@@ -1,1 +1,0 @@
-from .PHITS_tools import *
