### 2025-07-22 Lindt8/PHITS-Tools#1 @Lindt8
I agree, and I also took this as an opportunity to fix my earlier clunky packaging solution involving having duplicates of all of the `*.py` module files.  Changes are being made in [`feature/restructure_repo`](https://github.com/Lindt8/PHITS-Tools/tree/feature/restructure_repo) (also linked to this issue's right sidebar).

In its current structure, you should be able to `import PHITS_tools`, `from PHITS_tools import dchain_tools`, etc. from the repository's root directory. I moved the main `PHITS_tools.py` module file from the root directory to inside the `PHITS_tools/` directory.  The `src/PHITS_tools/` directory and its redundant `*.py` files have been removed, with the `PHITS_tools/__init__.py` and `pyproject.toml` files now ensuring submodules are properly packaged and locatable/importable regardless of whether you've just got a copy/clone/mirror of the repo or if you've installed it via `pip`/its wheel file.

I also made a [little beta release](https://github.com/Lindt8/PHITS-Tools/releases/tag/v1.6.0b4) out of the current [`feature/restructure_repo`](https://github.com/Lindt8/PHITS-Tools/tree/feature/restructure_repo) branch, [available on PyPI](https://pypi.org/project/PHITS-Tools/1.6.0b4/#files) / via `pip install PHITS-Tools==1.6.0b4` .
