### 2025-07-23 a015c48
updates MC_materials paths used with pkgutil
