+++ MC_materials/manage_mc_materials.py
@@ -40,8 +40,8 @@ Executing `update_materials_database_files` with information with a new material
 1. (if `prefer_user_data_folder=True`) `setup_local_mc_materials_directory` is executed, doing the following:
     - Confirmation of existence of a local `MC_materials` directory; if not found:
     - Creation of local `MC_materials` directory:
-        - In the same directory as `PHITS_tools.py` is a `MC_materials` directory. If it does not exist already, 
-            a new local directory, outside the PHITS Tools installation, will be created at 
+        - In the same directory as `PHITS_tools.py` (or one directory higher) is a `MC_materials` directory. If it does 
+            not exist already, a new local directory, outside the PHITS Tools installation, will be created at 
             [`pathlib.Path.home()`](https://docs.python.org/3/library/pathlib.html#pathlib.Path.home)` / '.PHITS-Tools' / 'MC_materials'`.
             - On Windows, this defaults to `C:\Users\USERNAME\.PHITS-Tools\MC_materials`
             - On MacOS and Linux, this defaults to `/home/USERNAME/.PHITS-Tools/MC_materials`
@@ -170,7 +170,7 @@ import csv
 import sys
 import json
 from pathlib import Path
-from PHITS_tools import Element_Z_to_Sym, Element_Sym_to_Z, fetch_MC_material, nuclide_plain_str_to_ZZZAAAM
+#from PHITS_tools import Element_Z_to_Sym, Element_Sym_to_Z, fetch_MC_material, nuclide_plain_str_to_ZZZAAAM
 
 
 def update_materials_database_files(json_filepath,name,mat_str,matid=None,density=None,source=None,
@@ -192,8 +192,8 @@ def update_materials_database_files(json_filepath,name,mat_str,matid=None,densit
             - if `prefer_user_data_folder=True`: your local [`"$HOME`](https://docs.python.org/3/library/pathlib.html#pathlib.Path.home)`/.PHITS-Tools/"` 
                     directory, creating it and copying the distributed MC_materials data to it if not yet existing via 
                     the `setup_local_mc_materials_directory` function.
-            - if `prefer_user_data_folder=False`: the same directory as the PHITS_tools.py module (which requires
-                    either PHITS Tools to be installed via `pip` or locatable within your PYTHONPATH system variable).
+            - if `prefer_user_data_folder=False`: the same directory as the PHITS_tools.py module (or one directory up), 
+                    which requires either PHITS Tools to be installed via `pip` or locatable within your PYTHONPATH system variable).
                 The JSON libraries that are shipped with PHITS Tools by default are titled `'Compiled_MC_materials'` 
                 and `'PNNL_materials_compendium'`.
         - `name` = string of the name of the material to be added / updated.  See the "Notes" section further below for 
@@ -256,6 +256,7 @@ def update_materials_database_files(json_filepath,name,mat_str,matid=None,densit
     import datetime
     import tkinter
     from tkinter import messagebox
+    from PHITS_tools import Element_Z_to_Sym, Element_Sym_to_Z, nuclide_plain_str_to_ZZZAAAM
     
     add_new_material = True
     rewrite_existing_material = False
@@ -277,21 +278,31 @@ def update_materials_database_files(json_filepath,name,mat_str,matid=None,densit
             try: # First, check MC_materials folder distributed with PHITS Tools
                 phits_tools_module_path = pkgutil.get_loader("PHITS_tools").get_filename()
                 mc_materials_dir_path = Path(Path(phits_tools_module_path).parent, 'MC_materials/')
+                mc_materials_updir_path = Path(Path(phits_tools_module_path).parent, '..', 'MC_materials/')
                 if mc_materials_dir_path.exists():
                     lib_file = Path(mc_materials_dir_path,database_filename)
[…]
