### 2025-07-23 70c682a
update README with more install instruction details
