+++ README.md
@@ -7,7 +7,7 @@
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14262720.svg)](https://doi.org/10.5281/zenodo.14262720)
 [![PHITS forumn discussion on PHITS Tools](https://img.shields.io/badge/PHITS%20forum%20discussion%20-%20%2333a2d9)](https://meteor.nucl.kyushu-u.ac.jp/phitsforum/t/topic/3651/)
 
-[Purpose](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#purpose) | [Installation](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#installation) | [Usage](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#primary-usageinterfaces) | [CLI options](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#cli-options) | [Automatic processing](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#automatic-processing-at-phits-runtime) | [Testing/Issues/Contribution](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#testing-reporting-issues-and-contributing) 
+[Purpose](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#purpose) | [Installation](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#installation) | [Usage](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#primary-usageinterfaces) | [CLI options](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#cli-options) | [Automatic processing](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#automatic-processing-at-phits-runtime) | [Submodules](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#submodules) | [Testing/Issues/Contribution](https://github.com/Lindt8/PHITS-Tools?tab=readme-ov-file#testing-reporting-issues-and-contributing) 
 
 ## Purpose
 
@@ -25,13 +25,15 @@ Install PHITS Tools:
 Import PHITS Tools like any other Python module:
 `import PHITS_tools` / `from PHITS_tools import *`
 
-*Note:* To use the CLI/GUI, you must execute the `PHITS_tools.py` module file with `python`.  To find the installed location of the module file, execute: `pip show PHITS-Tools -f`
+*Note:* To use the CLI/GUI, you must either use one of the commands `PHITS-Tools`/`PHITS_tools`/`phits-tools`/`PHITS-Tools-GUI` (more details further below) or execute the `PHITS_tools.py` module file with `python`.  To find the installed location of the module file, execute: `pip show PHITS-Tools -f`
 
 ### Manually 
 
-One may use the functions by first placing the PHITS_tools.py Python script into a folder in their PYTHONPATH system variable or in the active directory and then just importing them normally (`import PHITS_tools` / `from PHITS_tools import *`) or by executing the script `python PHITS_tools.py` with the PHITS output file to be parsed as the required argument (see `python PHITS_tools.py --help` for all CLI options) / without a file argument to be guided through with a GUI.
+One may use the functions by first placing the `PHITS_tools.py` Python script into a folder in their `PYTHONPATH` system variable or in the active directory and then just importing them normally (`import PHITS_tools` / `from PHITS_tools import *`) or by executing the script `python PHITS_tools.py` with the PHITS output file to be parsed as the required argument (see `python PHITS_tools.py --help` for all CLI options) / without a file argument to be guided through with a GUI.
 
[…]
