### 2025-07-23 7a3de6a
add extra clarification to manual install instructions
