### 2025-07-24 f0c0ea4
add GHA to build docs, resolves #11
