### 2025-07-24 c51b6b1
add GHA to build docs, resolves #11
