### 2025-07-28 ba4cb01
simplify some Path and pdoc code, clarify submodules in CONTRIBUTING.md, resolves #10 , #1
