### 2025-07-29 b58f84c
update module paths in doc building script
