### 2025-07-29 4313035
define version number exclusively in PHITS_tools.py, resolves #8
