+++ PHITS_tools/PHITS_tools.py
@@ -256,6 +256,8 @@ import os
 import numpy as np
 from pathlib import Path
 
+__version__ = '1.6.0b5'
+
 # default program settings
 launch_GUI = False
 run_with_CLI_inputs = False
@@ -2039,6 +2041,7 @@ def parse_phitsout_file(phitsout_filepath, include_input_echo=True, save_phitsou
             meta = starting_dict
     else:
         meta = starting_dict
+    meta['PHITS-Tools_version'] = __version__
     in_header = True 
     in_input_echo = False 
     in_footer = False
@@ -2662,7 +2665,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                     plt.tight_layout()
                     # add PHITS Tools info
                     fontdict = {'color': '#666666', 'weight': 'normal', 'size': 8, 'style': 'italic'}
-                    fig.text(0.005, 0.005, 'Figure generated by PHITS Tools, github.com/Lindt8/PHITS-Tools',
+                    fig.text(0.005, 0.005, r'Figure generated by PHITS Tools $\cdot$ github.com/Lindt8/PHITS-Tools $\cdot$'+' v{:}'.format(__version__),
                              fontdict=fontdict, ha='left', va='bottom', url='https://github.com/Lindt8/PHITS-Tools')
                     pdf.savefig()
                     if return_fg_list: fg_list.append(copy.deepcopy(fig))
@@ -3306,7 +3309,8 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                         
                     # add PHITS Tools info
                     fontdict = {'color':'#666666', 'weight':'normal', 'size': 8, 'style':'italic'}
-                    fig.text(0.005,0.0005,'Figure generated by PHITS Tools, github.com/Lindt8/PHITS-Tools',fontdict=fontdict,ha='left', va='bottom', url='https://github.com/Lindt8/PHITS-Tools')
+                    fig.text(0.005,0.0005,r'Figure generated by PHITS Tools $\cdot$ github.com/Lindt8/PHITS-Tools $\cdot$'+' v{:}'.format(__version__),
+                             fontdict=fontdict, ha='left', va='bottom', url='https://github.com/Lindt8/PHITS-Tools')
     
                     if use_rasterization:
                         pdf.savefig(bbox_extra_artists=[st], dpi=rasterize_dpi)
@@ -4601,6 +4605,7 @@ def parse_tally_header(tally_header,tally_content):
     else:
         meta = {}
     meta['tally_type'] = tally_type
+    meta['PHITS-Tools_version'] = __version__
     unsupported_tally_types = ['[T-WWG]', '[T-WWBG]', '[T-Volume]', '[T-Userdefined]', '[T-Gshow]', '[T-Rshow]',
                                '[T-3Dshow]', '[T-4Dtrack]', 'UNKNOWN'] # '[T-Dchain]',
     if tally_type in unsupported_tally_types or is_a_dchain_input_file:
