### 2025-07-29 ed8efdf
add version number to GUI and CLI, resolves #8
