### 2025-07-29 6ad464f
add optional dependencies to pyproject.toml, resolves #11
