### 2025-07-29 16394eb
FutureWarnings added for deprecated functions, resolves #3
