### 2025-07-29 9e9f4cb
ported and renamed useful functions from DCHAIN Tools (now deprecated there), resolves #3
