+++ PHITS_tools/PHITS_tools.py
@@ -75,14 +75,20 @@ functions return the data objects they produce for your own further analyses.
 - `fetch_MC_material`               : returns a string of a formatted material for PHITS or MCNP (mostly those in [PNNL-15870 Rev. 1](https://www.osti.gov/biblio/1023125)); see [**`PHITS_tools.manage_mc_materials` submodule documentation**](https://lindt8.github.io/PHITS-Tools/docs/manage_mc_materials.html) for details on managing the materials database
 - `ICRP116_effective_dose_coeff`    : returns effective dose conversion coefficient of a mono-energetic particle of some species and some geometry; does coefficients are those in [ICRP 116](https://doi.org/10.1016/j.icrp.2011.10.001)
 - `merge_dump_file_pickles`         : merge multiple dump file outputs into a single file (useful for dumps in MPI runs)
+- `is_number`                       : returns Boolean denoting whether provided string is that of a number
+- `find`                            : return index of the first instance of a value in a list
+
+### Nuclide/Particle Information Functions
+
 - `ZZZAAAM_to_nuclide_plain_str`    : returns a nuclide plaintext string for a given "ZZZAAAM" number (1000Z+10A+M)
 - `nuclide_plain_str_to_ZZZAAAM`    : returns a "ZZZAAAM" number (1000Z+10A+M) for a given nuclide plaintext string 
 - `nuclide_plain_str_to_latex_str`  : convert a plaintext string for a nuclide to a LaTeX formatted raw string
+- `nuclide_Z_and_A_to_latex_str`    : form a LaTeX-formatted string of a nuclide provided its Z/A/m information
 - `element_Z_to_symbol`             : return an elemental symbol string given its proton number Z
 - `element_symbol_to_Z`             : returns an atomic number Z provided the elemental symbol
+- `element_Z_or_symbol_to_name`     : returns a string of the name of an element provided its atomic number Z or symbol
+- `element_Z_or_symbol_to_mass`     : returns an element's average atomic mass provided its atomic number Z or symbol
 - `kfcode_to_common_name`           : converts a particle kf-code to a plaintext string
-- `is_number`                       : returns Boolean denoting whether provided string is that of a number
-- `find`                            : return index of the first instance of a value in a list
 
 ### Subfunctions for PHITS output parsing
 (These are meant as dependencies more so than for standalone usage.)
@@ -3950,6 +3956,41 @@ def merge_dump_file_pickles(dump_filepath_list, merged_dump_base_filepath='merge
     return merge_success
 
 
+def is_number(n):
+    '''
+    Description:
+        Determine if a string is that of a number or not.
+
+    Inputs:
+        - `n` = string to be tested
+
+    Outputs:
+        - `True` if value is a number (can be converted to float() without an error)
+        - `False` otherwise
+    '''
+    try:
+        float(n)
+    except ValueError:
+        return False
+    return True
+
+def find(target, myList):
+    '''
+    Description:
+        Search for and return the index of the first occurance of a value in a list.
+
+    Inputs:
+        - `target` = value to be searched for
+        - `myList` = list of values
+
+    Output:
+        - index of first instance of `target` in `myList`
+    '''
+    for i in range(len(myList)):
+        if myList[i] == target:
+            return i
+
+
 
 def ZZZAAAM_to_nuclide_plain_str(ZZZAAAM,include_Z=False,ZZZAAA=False,delimiter='-'):
     '''
[…]
