### 2025-07-30 719f0d6
add tests for more nuclide functions ported from DCHAIN-Tools
