### 2025-07-30 c09d96d
coverage testing configuration
