+++ tests/test_short_string_parsers.py
@@ -18,3 +18,5 @@ from PHITS_tools import is_number, split_str_of_equalities, parse_group_string, 
 ])
 def test_is_number(test_item,expected):
     assert is_number(test_item) == expected
+
+
