### 2025-07-30 3d30c39
unit tests for string/line parsing functions
