### 2025-07-30 08ddfda
coverage testing implementation
