+++ PHITS_tools/PHITS_tools.py
@@ -293,7 +293,7 @@ if __name__ == "__main__":
             import argparse
 
 
-if in_debug_mode:
+if in_debug_mode:   # pragma: no cover
     import pprint
     import time
     # Timer start
@@ -2616,7 +2616,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
         tally_output_list = [tally_output]
     else:  # list of tally output objects and/or pickles
         for i, to in enumerate(tally_output_list): # open any pickle files provided
-            if not isinstance(to, dict):
+            if not isinstance(to, dict):  # pragma: no cover
                 if to is None:
                     print("WARNING: 'None' tally output encountered in tally_output_list in autoplot_tally_results()!")
                     continue
@@ -2676,7 +2676,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                 tot_plot_axes = sum(1 for i in array_axes_lens if i > 1)
                 tot_num_values = np.prod(array_axes_lens)
                 
-                if tot_num_values > max_num_values_to_plot:
+                if tot_num_values > max_num_values_to_plot:  # pragma: no cover
                     print('\tWARNING: Tally output for ',tally_metadata['file'],' is VERY LARGE (',tot_num_values,' elements), deemed too large for automatic plotting.')
                     if return_fg_list: fg_list.append(None)
                     continue
@@ -3022,7 +3022,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                         hue_var = 'value'
                         row_var = plot_axes_sorted[2]
                     '''
-                elif tot_plot_axes==4:
+                elif tot_plot_axes==4:  # pragma: no cover
                     if plot_axes_sorted[0] in num:
                         x_var = plot_axes_sorted[0]
                     else:
@@ -3048,7 +3048,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                         hue_var = 'value'
                         row_var = plot_axes_sorted[2]
                         col_var = plot_axes_sorted[3]
-                elif tot_plot_axes==5:
+                elif tot_plot_axes==5:  # pragma: no cover
                     plot_kind = 'line'
                     if plot_axes_sorted[0] in num:
                         x_var = plot_axes_sorted[0]
@@ -3060,7 +3060,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                     row_var = plot_axes_sorted[2]
                     col_var = plot_axes_sorted[3]
                     style_var = plot_axes_sorted[4]
-                elif tot_plot_axes==6:
+                elif tot_plot_axes==6:  # pragma: no cover
                     plot_kind = 'line'
                     if plot_axes_sorted[0] in num:
                         x_var = plot_axes_sorted[0]
@@ -3073,7 +3073,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                     col_var = plot_axes_sorted[3]
                     style_var = plot_axes_sorted[4]
                     size_var = plot_axes_sorted[5]
-                else:
+                else:  # pragma: no cover
                     print('\tCannot create plot with 7+ variables...')
                     continue
                 '''
@@ -3199,7 +3199,7 @@ def autoplot_tally_results(tally_output_list,plot_errorbars=True,output_filename
                             row_var_renamed = row_var
                     if col_var != None: col_var_renamed = ax_labels[col_var]
                     
-                    if in_debug_mode:
+                    if in_debug_mode:  # pragma: no cover
                         print('plot_kind=', plot_kind, 
                               '\ny_var=', y_var, y_var_renamed, 
                               '\nx_var=', x_var, x_var_renamed, 
[…]
