+++ test/test_options_in_processing.py
@@ -1,0 +1,99 @@
+import pytest, PHITS_tools, os, pickle
+import numpy as np
+from pathlib import Path
+from traceback import format_exc
+
+path_to_phits_base_folder = Path(r'C:\phits')
+
+phits_sample_dir = Path(path_to_phits_base_folder,'sample')
+phits_recommendation_dir = Path(path_to_phits_base_folder,'recommendation')
+
+def test_extract_tally_outputs_from_phits_input_options():
+    phits_input = phits_recommendation_dir / 'NeutronSource' / 'NeutronSource.inp'
+    x = PHITS_tools.extract_tally_outputs_from_phits_input(phits_input, only_seek_phitsout=True)
+    assert x['phitsout'] != []
+
+@pytest.mark.slow
+@pytest.mark.integration
+def test_parse_tally_output_file_options():
+    tally_output_filepath = phits_recommendation_dir / 'muon' / 'product.out'
+    tod = PHITS_tools.parse_tally_output_file(tally_output_filepath, make_PandasDF=True, 
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=False, include_phitsout_in_metadata=False, 
+                                              prefer_reading_existing_pickle=False, compress_pickle_with_lzma=False, 
+                                              autoplot_tally_output=False
+                                              )
+    assert tod['tally_metadata'] != []
+    tod = PHITS_tools.parse_tally_output_file(tally_output_filepath, make_PandasDF=True,
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=True, include_phitsout_in_metadata=False,
+                                              prefer_reading_existing_pickle=False, compress_pickle_with_lzma=False,
+                                              autoplot_tally_output=False
+                                              )
+    output_pickle = Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pickle')
+    assert output_pickle.exists()
+    tod2 = PHITS_tools.parse_tally_output_file(tally_output_filepath, make_PandasDF=True,
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=False, include_phitsout_in_metadata=False,
+                                              prefer_reading_existing_pickle=True, compress_pickle_with_lzma=False,
+                                              autoplot_tally_output=False
+                                              )
+    assert (tod['tally_data'] == tod2['tally_data']).all()
+    os.remove(output_pickle)
+    tod = PHITS_tools.parse_tally_output_file(tally_output_filepath, make_PandasDF=True,
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=True, include_phitsout_in_metadata=True,
+                                              prefer_reading_existing_pickle=False, compress_pickle_with_lzma=True,
+                                              autoplot_tally_output=False
+                                              )
+    output_pickle = Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pickle.xz')
+    assert output_pickle.exists()
+    tod = PHITS_tools.parse_tally_output_file(tally_output_filepath, make_PandasDF=True,
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=True, include_phitsout_in_metadata=True,
+                                              prefer_reading_existing_pickle=True, compress_pickle_with_lzma=True,
+                                              autoplot_tally_output=True
+                                              )
+    os.remove(output_pickle)
+    assert Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pdf').exists()
+    assert Path(tally_output_filepath.parent, tally_output_filepath.stem + '.png').exists()
+    assert np.shape(tod['tally_data'])[-1] == 3
+    tod = PHITS_tools.parse_tally_output_file(tally_output_filepath, make_PandasDF=False,
[…]
