+++ pyproject.toml
@@ -153,6 +153,7 @@ exclude_lines = [
     "pragma: no cover",
     "if __name__ == '__main__':",
     "if False:",
+    "if in_debug_mode:",
 ]
 # Set coverage thresholds
 fail_under = 80

[…]
