### 2025-08-01 9495b29
functional testing now using pytest properly, more coverage testing
