+++ test/test_overall_functionality.py
@@ -1,4 +1,4 @@
-import PHITS_tools
+import pytest, PHITS_tools
 from pathlib import Path
 from traceback import format_exc
 import re
@@ -7,7 +7,6 @@ path_to_phits_base_folder = Path(r'C:\phits')
 
 phits_sample_dir = Path(path_to_phits_base_folder,'sample')
 phits_recommendation_dir = Path(path_to_phits_base_folder,'recommendation')
-test_input_mode_parsing = True  # Determines if [INPUT_FILE mode] of parse_all_tally_output_in_dir() is tested for all input files
 test_autoplotting = True  # Determines if autoplot_tally_results() is tested too for each tally output (notably slows testing)
 plot_paths = []
 
@@ -30,7 +29,9 @@ for f in recommendation_files:
     if keep_file:
         files_to_parse.append(Path(f))
 
+log_file_path = Path(__file__).parent / 'test_overall_functionality.log'
 log_file_str = ''
+log_str = ''
 num_tests = len(files_to_parse)
 i = 0
 num_passed, num_failed, num_warn = 0, 0, 0
@@ -40,95 +41,74 @@ known_issue_files = [r'phits\sample\source\Cosmicray\GCR-ground\cross.out', r'ph
                      r'phits\sample\source\Cosmicray\SEP-space\cross.out', r'phits\sample\source\Cosmicray\SEP-space\SEP-space.inp',
                      r'phits\sample\source\Cosmicray\TP-LEO\cross.out', r'phits\sample\source\Cosmicray\TP-LEO\TP-LEO.inp',
                      r'phits\sample\misc\snowman\rotate3dshow.inp']
-for f in files_to_parse:
+
+# save log file, deleting any old contents
+with open(log_file_path, "w") as f:
+    f.write(log_file_str)
+
+def process_and_log_output_file(f):
+    global log_str, log_file_str, i, num_passed, num_failed, num_warn
     i += 1
-    test_num_str = '{:3d}/{:3d}'.format(i,num_tests)
+    test_num_str = '{:3d}/{:3d}'.format(i, num_tests)
     try:
         if '_dmp.out' in str(f):
             x = PHITS_tools.parse_tally_dump_file(f, save_namedtuple_list=False, save_Pandas_dataframe=False)
         else:
-            x = PHITS_tools.parse_tally_output_file(f, save_output_pickle=False, autoplot_tally_output=test_autoplotting)
-            if test_autoplotting and Path(f.parent, f.stem + '.pdf').is_file(): plot_paths.append(Path(f.parent, f.stem + '.pdf'))
+            x = PHITS_tools.parse_tally_output_file(f, save_output_pickle=False,
+                                                    autoplot_tally_output=test_autoplotting)
+            if test_autoplotting and Path(f.parent, f.stem + '.pdf').is_file(): plot_paths.append(
+                Path(f.parent, f.stem + '.pdf'))
         log_str = test_num_str + '     pass  ' + str(f) + '\n'
         num_passed += 1
+        status = 'pass'
     except Exception as e:
         if re.sub(r'^.*?phits', 'phits', str(f)) in known_issue_files:
             log_str = test_num_str + '  !  WARN  ' + str(f) + '\n'
             num_warn += 1
+            status = 'WARN'
         else:
             log_str = test_num_str + '  x  FAIL  ' + str(f) + '\n'
+            status = 'FAIL'
         log_str += '\t\t' + repr(e) + '\n'
-        log_str += '\t\t' + format_exc().replace('\n','\n\t\t')
+        log_str += '\t\t' + format_exc().replace('\n', '\n\t\t')
         log_str = log_str[:-2]
         num_failed += 1
[…]
