+++ pyproject.toml
@@ -154,6 +154,7 @@ exclude_lines = [
     "if __name__ == '__main__':",
     "if False:",
     "if in_debug_mode:",
+    "raise ValueError",
 ]
 # Set coverage thresholds
 fail_under = 80

[…]
