### 2025-08-01 955f91a
expanded functional testing and README info
