+++ test/test_options_in_processing.py
@@ -54,6 +54,14 @@ def test_parse_tally_output_file_options():
                                               prefer_reading_existing_pickle=True, compress_pickle_with_lzma=True,
                                               autoplot_tally_output=True
                                               )
+    os.remove(Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pdf'))
+    os.remove(Path(tally_output_filepath.parent, tally_output_filepath.stem + '.png'))
+    tod = PHITS_tools.parse_tally_output_file(tally_output_filepath, make_PandasDF=True,
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=True, include_phitsout_in_metadata=True,
+                                              prefer_reading_existing_pickle=True, compress_pickle_with_lzma=True,
+                                              autoplot_tally_output=True
+                                              )
     os.remove(output_pickle)
     assert Path(tally_output_filepath.parent, tally_output_filepath.stem + '.pdf').exists()
     assert Path(tally_output_filepath.parent, tally_output_filepath.stem + '.png').exists()
@@ -128,6 +136,24 @@ def test_dchain_sample_3_step_dose():
     the `W_reg_target.out` and `W_xyz_target.out` outputs through DCHAIN
     '''
     tally_output_directory = path_to_phits_base_folder / 'dchain-sp' / 'sample' / '3-step_dose_xyz' 
+    if not (tally_output_directory / 'W_reg_target.act').exists() or not (tally_output_directory / 'W_xyz_target.act').exists():
+        print('Please follow the instructions in the `test_dchain_sample_3_step_dose()` docstring to run this test.')
+        pytest.skip("Missing required files. Please follow the instructions in the `test_dchain_sample_3_step_dose()` docstring to run this test.")
+        # pytest.fail('message', pytrace=False) is an option too.
+    tod = PHITS_tools.parse_tally_output_file(tally_output_directory / 'W_reg_target.dyld', make_PandasDF=True,
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=True, include_phitsout_in_metadata=True,
+                                              prefer_reading_existing_pickle=False, compress_pickle_with_lzma=True,
+                                              autoplot_tally_output=False
+                                              )
+    assert tod['tally_metadata'] != []
+    tod = PHITS_tools.parse_tally_output_file(tally_output_directory / 'W_reg_target.dtrk', make_PandasDF=True,
+                                              calculate_absolute_errors=True,
+                                              save_output_pickle=True, include_phitsout_in_metadata=True,
+                                              prefer_reading_existing_pickle=False, compress_pickle_with_lzma=True,
+                                              autoplot_tally_output=False
+                                              )
+    assert tod['tally_metadata'] != []
     x = PHITS_tools.parse_all_tally_output_in_dir(tally_output_directory,
                                                   return_tally_output=False, make_PandasDF=True,
                                                   calculate_absolute_errors=True, save_output_pickle=True,
@@ -168,6 +194,10 @@ def test_dump_file_options():
     '''
     tally_output_dir = phits_sample_dir / 'source' / 'Cosmicray' / 'GCR-blackhole'
     tally_output_file = tally_output_dir / 'cross_dmp.out'
+    if not (tally_output_file).exists():
+        print('Please follow the instructions in the `test_dump_file_options()` docstring to run this test.')
+        pytest.skip("Missing required files. Please follow the instructions in the `test_dump_file_options()` docstring to run this test.")
+        # pytest.fail('message', pytrace=False) is an option too.
[…]
