+++ PHITS_tools/PHITS_tools.py
@@ -684,7 +684,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                     from inspect import signature
                     max_num_values_to_plot = signature(autoplot_tally_results).parameters['max_num_values_to_plot'].default  # 1e7
                     tot_num_values = np.prod(np.shape(tally_output['tally_data'])[:-1])
-                    if tot_num_values > max_num_values_to_plot:
+                    if tot_num_values > max_num_values_to_plot:  # pragma: no cover
                         print('\tWARNING: Tally output for ', tally_output['tally_metadata']['file'], ' is VERY LARGE (', tot_num_values,
                               ' elements), deemed too large for automatic plotting (default max of', max_num_values_to_plot, 'elements).')
                     else:
@@ -712,7 +712,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
     else:
         is_val_file = True
 
-    if is_dmp_file:
+    if is_dmp_file:  # pragma: no cover
         print('\tERROR: The provided file is a "dump" output file. Use the function titled "parse_tally_dump_file" to process it instead.')
         return None
 
@@ -725,7 +725,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
             tally_output_filepath = potential_val_file
             is_val_file = True
             is_err_file = False
-        else:
+        else:  # pragma: no cover
             print('\t\t The corresponding file with tally values could not be found, so only these uncertainties will be parsed.')
 
     # Split content of output file into header and content
@@ -766,7 +766,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
         '''
         try:
             from dchain_tools import parse_dtrk_file, parse_dyld_files, Dname_to_ZAM
-        except:
+        except:  # pragma: no cover
             try:
                 from PHITS_tools.dchain_tools import parse_dtrk_file, parse_dyld_files, Dname_to_ZAM
             except:
@@ -872,7 +872,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                 from inspect import signature
                 max_num_values_to_plot = signature(autoplot_tally_results).parameters['max_num_values_to_plot'].default  # 1e7
                 tot_num_values = np.prod(np.shape(tally_output['tally_data'])[:-1])
-                if tot_num_values > max_num_values_to_plot:
+                if tot_num_values > max_num_values_to_plot:  # pragma: no cover
                     print('\t\tWARNING: Tally output for ', tally_output['tally_metadata']['file'], ' is VERY LARGE (', tot_num_values,
                           ' elements), deemed too large for automatic plotting (default max of',max_num_values_to_plot,'elements).')
                 else:
@@ -905,12 +905,12 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                 print('\tFailed to find the main DCHAIN *.act output file:',act_filepath)
                 if dtrk_filepath.is_file() or dyld_filepath.is_file():
                     print("\tJust processing found .dtrk/.dyld files instead...")
-                else:
+                else:  # pragma: no cover
                     print('\tNor were the *.dyld or *.dtrk files found. Aborting this process...')
                     return None
         try:
             from dchain_tools import process_dchain_simulation_output
-        except:
+        except:  # pragma: no cover
             try:
                 from PHITS_tools.dchain_tools import process_dchain_simulation_output
             except:
[…]
