### 2025-08-01 fd5e66b
remove lazy testing from early development stages
