+++ PHITS_tools/PHITS_tools.py
@@ -7248,287 +7248,10 @@ if run_with_CLI_inputs:
 elif launch_GUI:
     run_PHITS_tools_GUI()
 
-
-
-
-
 elif test_explicit_files_dirs:  # pragma: no cover
-    # pragma: no cover start
-    base_path = r'G:\Cloud\OneDrive\work\PHITS\test_tallies\tally\\'
-    #output_file_path = Path(base_path + 't-deposit\deposit_reg.out')
-    #output_file_path = Path(base_path + 't-deposit\deposit_eng_sp-reg.out')
-    output_file_path = Path(base_path + 't-track\\track_reg.out')
-    #output_file_path = Path(base_path + 't-track\\track_r-z.out')
-    #output_file_path = Path(base_path + 't-track\\track_xyz-xy.out')
-    #output_file_path = Path(base_path + r't-track\track_r-z_axis-rad.out')
-    #output_file_path = Path(base_path + r't-track\track_r-z_axis-deg.out')
-    #output_file_path = Path(base_path + 't-deposit\deposit_r-z.out')
-    #output_file_path = Path(base_path + 't-deposit\deposit_r-z_2dtype4.out')
-    #output_file_path = Path(base_path + 't-deposit\deposit_r-z_2dtype5.out')
-    #output_file_path = Path(base_path + 't-deposit\deposit_xyz_2dtype5.out')
-    #output_file_path = Path(base_path + 'tet_test\deposit-tet_axis-tet.out')
-    #output_file_path = Path(base_path + 'tet_test\deposit-tet_axis-eng.out')
-    #output_file_path = Path(base_path + 't-cross\cross_reg_axis-eng.out')
-    #output_file_path = Path(base_path + 't-cross\cross_reg_axis-reg.out')
-    #output_file_path = Path(base_path + 't-cross\cross_xyz_axis-eng.out')
-    #output_file_path = Path(base_path + 't-cross\cross_xyz_axis-eng_enclosed.out')
-    ###output_file_path = Path(base_path + 't-cross\cross_xyz_axis-reg.out')
-    #output_file_path = Path(base_path + 't-cross\cross_xyz_axis-xy.out')
-    #output_file_path = Path(base_path + 't-cross\cross-r-z_axis-eng.out')
-    #output_file_path = Path(base_path + 't-cross\cross-r-z_axis-eng_0r.out')
-    #output_file_path = Path(base_path + 't-cross\cross-r-z_axis-eng_enclosed.out')
-    #output_file_path = Path(base_path + 't-cross\complex\proton_in_hist_rz.out')
-    #output_file_path = Path(base_path + 't-cross\complex\\neutron_yield_rz-e-a-mesh.out')
-    #output_file_path = Path(base_path + 't-cross\complex\\neutron_yield.out')
-    #output_file_path = Path(base_path + 't-cross\complex\\xtra_neutron_yield_EvsTheta_whole-target.out')
-    #output_file_path = Path(base_path + 't-dpa\dpa_reg.out')
-    #output_file_path = Path(base_path + 't-dpa\dpa_xyz.out')
-    #output_file_path = Path(base_path + 't-dpa\dpa_r-z.out')
-    #output_file_path = Path(base_path + 'samepage\\proton_in_hist_rz_axis-eng_samepage-z.out')
-    #output_file_path = Path(base_path + 'samepage\\proton_in_hist_rz_reduced.out') # has NULL characters in it
-    #output_file_path = Path(base_path + 'samepage\\proton_in_hist_rz_sp-eng.out')
-    #output_file_path = Path(base_path + 't-deposit2\deposit2_reg.out')
[…]
