+++ .github/workflows/ci-tests.yml
@@ -26,7 +26,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -e ".[dev]"
+        pip install -e ".[develop]"
 
     - name: Check for critical syntax errors
       run: |
