### 2025-08-04 3fb5350
added note on updating submodule with git
