+++ .github/workflows/ci-tests.yml
@@ -61,12 +61,14 @@ jobs:
           --verbose
 
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v3
+      uses: codecov/codecov-action@v5
       with:
-        file: ./coverage.xml
+        files: ./coverage.xml
         flags: unittests
         name: codecov-umbrella
         fail_ci_if_error: false  # Don't fail CI if codecov upload fails
+        verbose: true
+        root_dir: .
 
     - name: Coverage comment (on PR only)
       if: github.event_name == 'pull_request'
