### 2025-08-04 1c37aa1
add ability to manually upload coverage reports
