+++ Makefile
@@ -1,5 +1,5 @@
 # Common Python project Makefile tasks
-.PHONY: test test-cov test-all test-unit clean-cov install clean docs build publish lint
+.PHONY: test test-cov test-all test-unit clean-cov install clean docs build publish lint upload-unit upload-full
 
 # Just run the tests without coverage reporting
 test:
@@ -7,11 +7,11 @@ test:
 
 # Testing with coverage report
 test-cov:
-	pytest tests/ --cov=PHITS_tools --cov-report=html:htmlcov-unit --cov-report=term-missing
+	pytest tests/ --cov=PHITS_tools --cov-report=html:htmlcov-unit --cov-report=xml:coverage-unit.xml --cov-report=term-missing
 
 # Test everything, including integration tests in test/
 test-all:
-	pytest tests/ test/ --cov=PHITS_tools --cov-report=html:htmlcov-full --cov-report=term-missing
+	pytest tests/ test/ --cov=PHITS_tools --cov-report=html:htmlcov-full --cov-report=xml:coverage-full.xml --cov-report=term-missing
 
 # Just unit tests
 test-unit:
@@ -48,3 +48,10 @@ lint:
 	@echo "Checking for critical syntax errors..."
 	flake8 PHITS_tools/ tests/ --select=E9,F63,F7,F82,F83 --show-source --statistics
 	@echo "No critical errors found"
+
+upload-unit:
+	codecovcli upload-coverage --coverage-files-search-direct-file ./coverage-unit.xml --flag ci-unittests --name codecov-unit --slug Lindt8/PHITS-Tools -t %CODECOV_TOKEN% --disable-search
+
+
+upload-full:
+	codecovcli upload-coverage --coverage-files-search-direct-file ./coverage-full.xml --flag full-suite --name codecov-full --slug Lindt8/PHITS-Tools -t %CODECOV_TOKEN% --disable-search
