+++ .github/workflows/ci-tests.yml
@@ -63,9 +63,9 @@ jobs:
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v5
       with:
-        files: ./coverage.xml
-        flags: unittests
-        name: codecov-umbrella
+        files: ./coverage-unit.xml
+        flags: ci-unittests
+        name: codecov-unit
         fail_ci_if_error: false  # Don't fail CI if codecov upload fails
         verbose: true
         root_dir: .
