+++ codecov.yml
@@ -1,0 +1,55 @@
+# codecov.yml
+coverage:
+  status:
+    project:
+      default:
+        target: auto
+        threshold: 1%
+        base: auto
+      unit-tests:
+        flags:
+          - ci-unittests
+        target: 15%  # Realistic target for unit tests alone
+        threshold: 5%
+      full-suite:
+        flags:
+          - full-suite
+        target: 80%  # Full test suite target, with all conditions met in test/README.md
+        threshold: 2%
+
+flags:
+  ci-unittests:
+    paths:
+      - PHITS_tools/
+    carryforward: false
+    
+  full-suite:
+    paths:
+      - PHITS_tools/
+    carryforward: false
+
+comment:
+  layout: "header, diff, flags, components, footer"
+  behavior: default
+  require_changes: false
+  require_base: no
+  require_head: yes
+  
+  # Custom comment template
+  template: |
+    ## Coverage Report
+    
+    | Test Suite | Coverage | Change |
+    |------------|----------|--------|
+    | Unit Tests | {{ flags.ci-unittests.coverage }}% | {{ flags.ci-unittests.diff }}% |
+    | Full Suite | {{ flags.full-suite.coverage }}% | {{ flags.full-suite.diff }}% |
+    
+    **Note**: Unit tests cover the standalone functions in PHITS Tools that end-users are likely to utilize themselves.
+    Integration/functional tests require externally distributed PHITS output files not available in CI, see test/README.md.
+
+ignore:
+  - "tests/*"
+  - "test/*"
+  - "scripts/*"
+  - "docs/*"
+  - "build/*"
