### 2025-08-04 4e52656
fixed badges and codecov config syntax
