+++ README.md
@@ -203,9 +203,9 @@ these should work and be accessed as shown without any additional configuration 
 
 ## Testing, reporting issues, and contributing
 
-[![CI Tests](https://github.com/Lindt8/PHITS-Tools/actions/workflows/ci.yml/badge.svg)](https://github.com/Lindt8/PHITS-Tools/actions/workflows/ci.yml)
-[![Unit Test Coverage](https://codecov.io/gh/Lindt8/PHITS-Tools/flag/ci-unittests/graph/badge.svg)](https://codecov.io/gh/Lindt8/PHITS-Tools)
-[![Full Suite Coverage](https://codecov.io/gh/Lindt8/PHITS-Tools/flag/full-suite/graph/badge.svg)](https://codecov.io/gh/Lindt8/PHITS-Tools)
+[![CI Tests](https://github.com/Lindt8/PHITS-Tools/actions/workflows/ci-tests.yml/badge.svg)](https://github.com/Lindt8/PHITS-Tools/actions/workflows/ci-tests.yml)
+[![Unit Tests](https://codecov.io/gh/Lindt8/PHITS-Tools/graph/badge.svg?flag=ci-unittests&label=Unit%20Tests)](https://app.codecov.io/gh/Lindt8/PHITS-Tools/flags?flag=ci-unittests)
+[![Full Suite](https://codecov.io/gh/Lindt8/PHITS-Tools/graph/badge.svg?flag=full-suite&label=Full%20Suite)](https://app.codecov.io/gh/Lindt8/PHITS-Tools/flags?flag=full-suite)
 
 I have extensively tested this module with a rather large number of PHITS output files with all sorts of different geometry settings, combinations of meshes, output options, and other settings to try to capture as a wide array of output files as I could (including the ~300 output files within the `phits/sample/` and `phits/recommendation/` directories included in the distributed PHITS release, which can be tested in an automated way with `test/test_PHITS_tools.py` in this repository, along with a large number of supplemental variations to really test every option I could think of), but there still may be some usage/combinations of different settings I had not considered that may cause PHITS Tools to crash when attempting to parse a particular output file.  If you come across such an edge case&mdash;a standard PHITS tally output file that causes PHITS Tools to crash when attempting to parse it&mdash;please submit it as an issue and include the output file in question and I'll do my best to update the code to work with it!  Over time, hopefully all the possible edge cases can get stamped out this way. :)
 
