### 2025-08-04 ee81eba
fix vertical alignment with badges and text
