### 2025-08-04 7dc5c0f
make text and codecov badges share URL
