### 2025-08-04 13ce733
fix internal repo link references
