+++ test/README.md
@@ -9,7 +9,7 @@ Owing to the nature of these tests, it is difficult to have a "right answer" to 
 
 ## Testing instructions
 
-Testing was developed with [`pytest`](https://pypi.org/project/pytest/) and [`pytest-cov`](https://pypi.org/project/pytest-cov/) (for coverage testing).  Configuration details can be found in [`pyproject.toml`](pyproject.toml) and common test execution commands in the [`Makefile`](Makefile).
+Testing was developed with [`pytest`](https://pypi.org/project/pytest/) and [`pytest-cov`](https://pypi.org/project/pytest-cov/) (for coverage testing).  Configuration details can be found in [`pyproject.toml`](../pyproject.toml) and common test execution commands in the [`Makefile`](../Makefile).
 
 Owing to PHITS being export controlled and requring a state-issued license from the Japan Atomic Energy Agency or the OECD Nuclear Energy Agency (NEA), no files distributed with PHITS are redistributed here.  That said, if you already have a PHITS installation, you will find an extensive set of sample PHITS inputs and outputs provided within the `phits/sample/` and `phits/recommendation/` directories and their nested subdirectories, totalling approximately 300 test output files to be parsed by this script.  Also note that the `phits/recommendation/` directory, complete with (approx. 50) PHITS output files, is publicly available for download on the PHITS website at the bottom of [this page under "Recommendation Settings"](https://phits.jaea.go.jp/rireki-manuale.html) ([direct download of ZIP file here](https://phits.jaea.go.jp/lec/recommendation.zip)).
 
@@ -17,13 +17,13 @@ The testing scripts in this directory use these distributed sample and recommend
 
 ### Mass functional testing
 
-The [`test_overall_functionality.py`](test/test_overall_functionality.py) script tests the correct functioning of the `PHITS Tools` module in parsing a variety of PHITS output files.  When the script is ran, `PHITS Tools` will attempt parsing all of the outputs in `phits/sample/` and `phits/recommendation/` and print its test results to the terminal and to `test_overall_functionality.log`, saved in this directory.
+The [`test_overall_functionality.py`](test_overall_functionality.py) script tests the correct functioning of the `PHITS Tools` module in parsing a variety of PHITS output files.  When the script is ran, `PHITS Tools` will attempt parsing all of the outputs in `phits/sample/` and `phits/recommendation/` and print its test results to the terminal and to `test_overall_functionality.log`, saved in this directory.
 
-The [`test_input_mode_parsing.py`](test/test_input_mode_parsing.py) script is very similar to `test_overall_functionality.py` but instead will attempt parsing all of the PHITS _input files_ in `phits/sample/` and `phits/recommendation/` and then use `parse_all_tally_output_in_dir()` in _[INPUT_FILE mode]_ to process all of the outputs produced by each input.  It prints its test results to the terminal and to `test_input_mode_parsing.log`, saved in this directory.
+The [`test_input_mode_parsing.py`](test_input_mode_parsing.py) script is very similar to `test_overall_functionality.py` but instead will attempt parsing all of the PHITS _input files_ in `phits/sample/` and `phits/recommendation/` and then use `parse_all_tally_output_in_dir()` in _[INPUT_FILE mode]_ to process all of the outputs produced by each input.  It prints its test results to the terminal and to `test_input_mode_parsing.log`, saved in this directory.
 
 ### Options testing
 
-The [`test_options_in_processing.py`](test/test_options_in_processing.py) script, rather than surveying many PHITS outputs all using the same function options as in [`test_overall_functionality.py`](test/test_overall_functionality.py), focuses on testing a small subset of the PHITS outputs but testing a variety of options/settings combinations not covered by [`test_overall_functionality.py`](test/test_overall_functionality.py).  
[…]
