### 2025-08-04 8eafbb0
add testing info to CONTRIBUTING.md
