+++ CONTRIBUTING.md
@@ -17,6 +17,17 @@ If you would like to contribute a new function or changes to any existing functi
 
 Owing to the slightly complicated/unorthodox approach used for integrating/structuring the `dchain_tools` and `manage_mc_materials` submodules (see [`pyproject.toml`](https://github.com/Lindt8/PHITS-Tools/blob/main/pyproject.toml) and [`PHITS_tools/__init__.py`](https://github.com/Lindt8/PHITS-Tools/blob/main/PHITS_tools/__init__.py)), any additions/changes to them should be made to their respective module files (with those to [DCHAIN Tools](https://github.com/Lindt8/DCHAIN-Tools) made in its repository).  Otherwise, source development should occur within the `PHITS_tools/` directory.
 
+Please submit any pull requests to the `develop` branch.
+
+## Testing
+
+The testing for PHITS Tools is divided into two directories:
+
+- [`tests/`](tests/) for unit tests: These are used for [CI testing]((https://github.com/Lindt8/PHITS-Tools/actions/workflows/ci-tests.yml)) and are automatically ran with pull requests to the `develop` and `main` branches.  This set of unit tests primarily is for the standalone functions in PHITS Tools that end-users are likely to use for additional results post-processing or independent use, rather than the functions which are complex internal cogs in output file processing.
+- [`test/`](test/) for integration/functional tests: These require externally distributed PHITS output files (see [`test/README.md`](test/README.md) for full details) and test the start-to-finish function of PHITS Tools. These tests exist to verify if the code runs without error and, where reasonable, checks if the output has properties/formatting as expected.
+
+Testing was developed with [`pytest`](https://pypi.org/project/pytest/) and [`pytest-cov`](https://pypi.org/project/pytest-cov/) (for coverage testing).  Configuration details can be found in [`pyproject.toml`](pyproject.toml) and common test execution commands in the [`Makefile`](Makefile).  Coverage reports are also uploaded to [Codecov](https://app.codecov.io/github/lindt8/phits-tools); automatically [via CI](https://github.com/Lindt8/PHITS-Tools/actions/workflows/ci-tests.yml) for the unit tests (see flag [`ci-unittests`](https://app.codecov.io/github/lindt8/phits-tools?flags%5B0%5D=ci-unittests)) and manually via the `make upload-full` command (which makes use of [`codecovcli`](https://pypi.org/project/codecov-cli/)) for the integration/functional tests (see flag [`full-suite`](https://app.codecov.io/github/lindt8/phits-tools?flags%5B0%5D=full-suite)), of course after running those tests with `make test-all`.
+
 ## Note about DCHAIN Tools and Git Submodules
 Note that [DCHAIN Tools](https://github.com/Lindt8/DCHAIN-Tools) is included in PHITS Tools as a [Git Submodule](https://gist.github.com/gitaarik/8735255), and this has some implications.  First, any development for DCHAIN Tools should be pointed at the [DCHAIN Tools repository](https://github.com/Lindt8/DCHAIN-Tools) instead of here.  (Feel free to make suggestions for it in Issues either here or there, though with a preference for the DCHAIN Tools repo.)
 
