### 2025-08-05 d52cddc
update JOSS draft PDF workflow
