+++ .github/workflows/draft-pdf.yml
@@ -1,10 +1,20 @@
 name: Draft PDF
 on:
   push:
+    branches:
+      - main
+      - update-paper
     paths:
-      - paper.md
-      - paper.bib
-      - .github/workflows/draft-pdf.yml
+      - 'paper.md'
+      - 'paper.bib'
+      - '.github/workflows/draft-pdf.yml'
+  pull_request:
+    branches:
+      - main
+    paths:
+      - 'paper.md'
+      - 'paper.bib'
+      - '.github/workflows/draft-pdf.yml'
 
 jobs:
   paper:
@@ -19,15 +29,17 @@ jobs:
           journal: joss
           # This should be the path to the paper within your repo.
           paper-path: paper.md
-      - name: Upload
+      - name: Upload artifact
         uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
+          name: paper-${{ github.sha }}  # Make artifact name unique
           path: paper.pdf
       - name: Commit PDF to repository
+        if: github.ref == 'refs/heads/main'
         uses: EndBug/add-and-commit@v9
         with:
           message: '(auto) Paper PDF Draft'
