+++ .github/workflows/draft-pdf.yml
@@ -32,11 +32,10 @@ jobs:
       - name: Upload artifact
         uses: actions/upload-artifact@v4
         with:
-          name: paper
+          name: paper-${{ github.sha }}  # Make artifact name unique
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          name: paper-${{ github.sha }}  # Make artifact name unique
           path: paper.pdf
       - name: Commit PDF to repository
         if: github.ref == 'refs/heads/main'
