### 2025-08-05 5bbddbc
Update paper.md
