+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'PHITS & DCHAIN Tools: Python modules for parsing, organizing, and analyzing results from the PHITS radiation transport and DCHAIN activation codes'
+title: 'The PHITS Tools Python package for parsing, organizing, and analyzing results from the PHITS radiation transport and DCHAIN activation codes'
 tags:
   - PHITS
   - DCHAIN
@@ -67,17 +67,18 @@ The DCHAIN code coupled to PHITS specializes in calculating nuclide inventories 
 (such as activity, decay heat, decay gamma-ray emission spectra, and more) as a function of time for 
 any arbitrary irradiation schedule from any radiation source.
 
-The modules presented here automate the time-consuming task of extracting the 
+The package presented here automates the time-consuming task of extracting the 
 numerical results and metadata from PHITS/DCHAIN simulations and organizes them into 
 a standardized format, easing and expediting further practical real-world analyses. 
-They also provide functions for some of the most common analyses one 
+It also can automatically produce images of plotted tally outputs and 
+provides functions for some of the most common analyses one 
 may wish to perform on simulation outputs.
 
 
 
 # Statement of need
 
-`PHITS Tools` and `DCHAIN Tools` serve as an interface between the plaintext (and binary) outputs
+`PHITS Tools` (and its `DCHAIN Tools` submodule) serve as an interface between the plaintext (and binary) outputs
 of the PHITS and DCHAIN codes and Python&mdash;greatly expediting further programmatic analyses, 
 comparisons, and visualization&mdash;and provide some extra analysis tools. 
 The outputs of the PHITS code are, aside from the special binary "dump" files, plaintext files formatted 
@@ -87,7 +88,7 @@ code are formatted in a variety of tabular, human-readable structures.  Historic
 extraction and organization of numerical results and metadata from both codes often required 
 writing a bespoke processing script for most individual simulations, 
 possibly preceded by manual data extraction/isolation too. 
-`PHITS Tools` and `DCHAIN Tools` provide universal output parsers for the PHITS and DCHAIN codes, 
+`PHITS Tools` provides universal output parsers for the PHITS and DCHAIN codes, 
 capable of processing all of the relevant output files produced by each code and
 outputting the numerical results and metadata in a consistent, standardized output format.
 
@@ -130,9 +131,9 @@ may be more user-friendly to those accustomed to working in Pandas.
 
 `PHITS Tools` is also capable of parsing the "dump" output files (both binary and ASCII formats) 
 that are available for some tallies, and it can also automatically detect, parse, and process all PHITS 
-output files within a provided directory, very convenient for PHITS simulations employing 
-multiple tallies, each with its own output file, whose output are to be further studied, 
-e.g., compared to experimental data or other simulations. 
+output files within a provided directory or produced by a specified PHITS input file, very convenient 
+for PHITS simulations employing multiple tallies, each with its own output file, 
+whose output are to be further studied, e.g., compared to experimental data or other simulations. 
 The `PHITS Tools` module can be used by 
 (1) importing it as a Python module in a script and calling its functions, 
 (2) running it in the command line via its CLI with a provided PHITS output
@@ -141,23 +142,23 @@ The `PHITS Tools` module can be used by
 the various output processing options and settings within `PHITS Tools`.
 
 When used as an imported module, `PHITS Tools` provides a number of supplemental
-functions aiding with further analyses, such as tools for constructing one's own 
-tally over the history-by-history output of the "dump" files, rebinning histogrammed 
-results to a different desired binning structure, applying effective dose 
-conversion coefficients from ICRP 116 [@ICRP116_ref_withauthors] to tallied particle 
[…]
