### 2025-08-05 49048e7
Update paper.md
