### 2025-08-05 526c251
Update paper.md
