### 2025-08-05 1b21ea1
Update paper.md
