### 2025-08-05 c08d102
Update paper.md
