### 2025-08-05 b73347e
Update paper.md
