+++ paper.md
@@ -136,7 +136,7 @@ whose output are to be further studied, e.g., compared to experimental data or o
 `PHITS Tools` can be used by 
 (1) importing it as a Python package in a script and calling its functions, 
 (2) running it in the command line via its CLI with a provided PHITS output
- file (or directory/input) and settings flags/options, or 
+ file (or directory/input file) path and settings flags/options, or 
 (3) running it without any arguments to launch a GUI stepping the user through 
 the various output processing options and settings in `PHITS Tools`.
 
