### 2025-08-05 #8311 @Lindt8
Ah, for context, I was just updating/regenerating the draft paper PDF to reflect changes to the PHITS Tools package that have been made since the initial submission (see https://github.com/Lindt8/PHITS-Tools/pull/16, perhaps most notably changing the paper's title to: _The PHITS Tools Python package for parsing, organizing, and analyzing results from the PHITS radiation transport and DCHAIN activation codes_).

I have only just today finished addressing the issues @MicahGale submitted to the repository in his review (see https://github.com/Lindt8/PHITS-Tools/pull/15 and https://github.com/Lindt8/PHITS-Tools/issues).  He's provided a lot of guidance in helping me improve the project, and all of these changes made in this review process (sans today's pending PR https://github.com/Lindt8/PHITS-Tools/pull/15) have already been pulled into the [`develop`](https://github.com/Lindt8/PHITS-Tools/tree/develop) branch but not yet into `main`.
