### 2025-08-06 Lindt8/PHITS-Tools#5 @MicahGale
> [test/](https://github.com/Lindt8/PHITS-Tools/tree/feature/improve-testing/test/) for integration/functional tests: These require externally distributed PHITS output files (see [test/README.md](https://github.com/Lindt8/PHITS-Tools/tree/feature/improve-testing/test/README.md) for full details) and test the start-to-finish function of PHITS Tools. These tests exist to verify if the code runs without error and, where reasonable, checks if the output has properties/formatting as expected. These tests combined with the unit tests should achieve ~89% total coverage, with over 80% coverage in every function in the main PHITS Tools module (except the CLI and GUI launching functions, which are excluded), as shown in this coverage report: [htmlcov-full.zip](https://github.com/user-attachments/files/21596166/htmlcov-full.zip).

Is it necessary that `PHITS tools` be ran? If not (at least for most tests), then you could test against the output files, which would allow you to run these tests on actions without any export control concerns.
