+++ .github/workflows/ci-tests.yml
@@ -7,6 +7,47 @@ on:
     branches: [ feature/improve-testing ]
 
 jobs:
+  lint:
+    runs-on: ubuntu-latest
+    name: Code Quality Check
+    
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          submodules: recursive  # Initialize git submodules
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.11"  # Use single version for linting
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install flake8
+
+      - name: Check for critical syntax errors
+        run: |
+          # Only check for critical errors that would cause crashes
+          # E9xx: Runtime errors (syntax errors, indentation errors)
+          # F63: Invalid escape sequence  
+          # F7x: Logic errors (break/continue outside loop)
+          # F82: Undefined name in __all__
+          # F83: Undefined name in except handler
+          flake8 PHITS_tools/ tests/ \
+            --select=E9,F63,F7,F82,F83 \
+            --show-source \
+            --statistics \
+            --count
+
+          # Show other issues as warnings only (don't fail CI)
+          echo "=== Style warnings (non-blocking) ==="
+          flake8 PHITS_tools/ tests/ \
+            --ignore=E9,F63,F7,F82,F83 \
+            --exit-zero \
+            --statistics \
+            --count || true
+  
   test:
     runs-on: ubuntu-latest
     strategy:
@@ -27,28 +68,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -e ".[develop]"
-
-    - name: Check for critical syntax errors
-      run: |
-        # Only check for critical errors that would cause crashes
-        # E9xx: Runtime errors (syntax errors, indentation errors)
-        # F63: Invalid escape sequence  
-        # F7x: Logic errors (break/continue outside loop)
-        # F82: Undefined name in __all__
-        # F83: Undefined name in except handler
-        flake8 PHITS_tools/ tests/ \
-          --select=E9,F63,F7,F82,F83 \
-          --show-source \
-          --statistics \
-          --count
-        
-        # Show other issues as warnings only (don't fail CI)
-        echo "=== Style warnings (non-blocking) ==="
-        flake8 PHITS_tools/ tests/ \
-          --ignore=E9,F63,F7,F82,F83 \
-          --exit-zero \
-          --statistics \
-          --count || true
 
     - name: Run unit tests with coverage
       run: |
