### 2025-08-06 4552519
list relevant make commands explicitly
