### 2025-08-06 daad8b6
clarify git submodule update instruction
