### 2025-08-06 40c89d0
improve testing logging, globals, and default test data path specification
