+++ test/test_input_mode_parsing.py
@@ -1,12 +1,11 @@
 import pytest, PHITS_tools
 from pathlib import Path
 from traceback import format_exc
-import re
+import re, io
+from conftest import path_to_base_phits_test_dir
 
-path_to_phits_base_folder = Path(r'C:\phits')
-
-phits_sample_dir = Path(path_to_phits_base_folder,'sample')
-phits_recommendation_dir = Path(path_to_phits_base_folder,'recommendation')
+phits_sample_dir = Path(path_to_base_phits_test_dir,'sample')
+phits_recommendation_dir = Path(path_to_base_phits_test_dir,'recommendation')
 test_autoplotting = True  # Determines if autoplot_tally_results() is tested too for each tally output (notably slows testing)
 known_issue_files = [r'phits\sample\source\Cosmicray\GCR-ground\cross.out', r'phits\sample\source\Cosmicray\GCR-ground\GCR-ground.inp',
                      r'phits\sample\source\Cosmicray\GCR-LEO\cross.out', r'phits\sample\source\Cosmicray\GCR-LEO\GCR-LEO.inp',
@@ -19,10 +18,6 @@ sample_files = phits_sample_dir.rglob('*.out')
 recommendation_files = phits_recommendation_dir.rglob('*.out')
 
 log_file_path = Path(__file__).parent / 'test_input_mode_parsing.log'
-log_file_str = ''
-log_str = ''
-i = 0
-numi_passed, numi_failed, numi_warn = 0, 0, 0
 
 potential_inputs_to_parse = [Path(f) for f in phits_sample_dir.rglob('*.inp')] + [Path(f) for f in phits_recommendation_dir.rglob('*.inp')]
 inputs_to_parse = []
@@ -31,32 +26,37 @@ for f in potential_inputs_to_parse:
     if len([i for i in f.parent.glob('*.out')]) == 0: continue  # skip inputs in directories with no output files
     skip_this_input = False
     inputs_to_parse.append(f)
-numi_tests = len(inputs_to_parse)
-#log_str = '\n\n--------------------------------------------------------------------\n'
-#log_str += 'INPUT FILE PARSING TESTS\n\n'
-#log_file_str += log_str
 
-# save log file, deleting any old contents
-with open(log_file_path, "w") as f:
-    f.write(log_file_str)
+class TestState:
+    def __init__(self, num_tests):
+        self.log_buffer = io.StringIO()
+        self.num_tests = num_tests
+        self.i = 0
+        self.num_passed = 0
+        self.num_failed = 0
+        self.num_warn = 0
+
+@pytest.fixture(scope="session")
+def test_state():
+    state = TestState(len(inputs_to_parse))
+    return state
 
 
-def process_and_log_output_from_input_file(f):
-    global log_str, log_file_str, i, numi_passed, numi_failed, numi_warn
-    i += 1
-    test_num_str = '{:3d}/{:3d}'.format(i, numi_tests)
+def process_and_log_output_from_input_file(f, test_state):
+    test_state.i += 1
+    test_num_str = '{:3d}/{:3d}'.format(test_state.i, test_state.num_tests)
     try:
         x = PHITS_tools.parse_all_tally_output_in_dir(f, save_output_pickle=False, merge_tally_outputs=True,
                                                       save_pickle_of_merged_tally_outputs=True,
                                                       compress_pickle_with_lzma=True,
                                                       autoplot_all_tally_output_in_dir=test_autoplotting)
         log_str = test_num_str + '     pass  ' + str(f) + '\n'
-        numi_passed += 1
+        test_state.num_passed += 1
         status = 'pass'
     except Exception as e:
         if re.sub(r'^.*?phits', 'phits', str(f)) in known_issue_files:
[…]
