### 2025-08-06 11f682f
relax dependency version requirements, resolves #1
