+++ pyproject.toml
@@ -19,17 +19,17 @@ classifiers = [
   "Operating System :: OS Independent",
 ]
 dependencies = [
-  "numpy>=2.2.1", 
-  "pandas>=2.2.3", 
-  "scipy>=1.14.1", 
-  "matplotlib>=3.10.0", 
-  "seaborn>=0.13.2"
+  "numpy>=2.2", 
+  "pandas>=2.2", 
+  "scipy>=1.14", 
+  "matplotlib>=3.10", 
+  "seaborn>=0.13"
 ]
 
 [project.optional-dependencies]
 extras = [
-  'munch>=2.5.0',
-  'pyvista>=0.44.2'
+  'munch>=2.5',
+  'pyvista>=0.44'
 ]
 test = [
 	"pytest>=6.0", 
