### 2025-08-06 Lindt8/PHITS-Tools#1 @Lindt8
1. See my [reply](https://github.com/Lindt8/PHITS-Tools/pull/10#discussion_r2256692636) in #10.
2. Fair enough, I've removed requirements.txt in d22cfe42d611322d224469ac97944b4600d3a55f and relaxed `pyproject.toml` dependency version numbers to just specify the minimum minor release in 11f682fb3f79fa6066fefc72aaf461447cc13a97.
