### 2025-08-06 abc9772
make CLI default, regardless of num args, for script entry points, fixes #21
