### 2025-08-06 fc13dc3
make CLI default, regardless of num args, for script entry points, fixes #21
