### 2025-08-06 6544618
clarify script entry point behavior as being CLI-focused
