### 2025-08-06 63e4f59
navigation links now just use anchors, fixes #19
