### 2025-08-06 2cb78ff
fix hover text on badges
