### 2025-08-06 a0e6660
fix hover text on coverage badges everywhere
