### 2025-08-10 Lindt8/PHITS-Tools#22 @MicahGale
Make it easier to work with 10-D tally results
The results from `parse_tally_dump_file` are in a 10-D numpy array. I understand that tally results are hard to work with. In general though working with this many dimensions is error prone, as you can work with the wrong dimension on accident.

**Describe the solution you'd like**

Long-term an alternative to numpy in this case could be Xarray, though that is its own beast.

One possibility is: create indexing generating helpers. I double-checked: numpy is expecting `tuple`s of `slice`s for indexing:

``` python
In [9]: import numpy as np

In [10]: test2 = np.array([[1,2],[3,4]])

In [12]: test2[1,0]
Out[12]: np.int64(3)

In [13]: tu = (1,0)

In [14]: test2[tu]
Out[14]: np.int64(3)
```

So I'm thinking you could make a helper function that takes all keyword arguments, defaults to 0 (flattening), and treats `None` as `:`

``` python

def gen_tally_index(*,ir = 0, iy=0, ....):
    pass
```

And would behave like:

``` python
>>> gen_tally_index(ir=None)
(slice(), 0, 0, 0, 0, 0, ...)
>>> gen_tally_index(ir=2, ie=None, ip=None, ierr=None)
(2,0,0,slice(),0,0,0,slice(),0,slice())

Other options might be to provide some commonly used indices as constants:

``` python
IR = (slice(), 0, 0, 0, 0, 0, ...)
```
