### 2025-08-10 Lindt8/PHITS-Tools#5 @MicahGale
Looking through the test suite closer I have a few thoughts.

1. You don't need to include the files in this repo. You could have the see download the output files, and then work on them. But the need for running `PHITS` for some of the tests makes this a bit excessive for now. I think you have a found a good enough solution for now for splitting unit and integration testing and running tests locally. Eventually it would be nice to shift as much of the testing to CI, but that's a "nice to have".

2. I think it would still be good to have an example output file and gold parsed version included in the test suite. I understand the licensing limitations concern. So what I think you should do is:

    1. Include a "synthetic" output file
    2. Include a parsed version of the output stored in JSON (gold standard)
    3. Compare the parsed version to the gold standard.

3. How to make the "synthetic" output file. The `recommended` output files are not feasible to license. Could you create a simple PHITS problem (like planar source into a non-multiplying slab, etc.) and then store the output from that run? You would need to check the PHITS license to see if this is allowed. Otherwise you could create a new output file by hand based on the example output files, but with made-up values.
