### 2025-08-10 Lindt8/PHITS-Tools#31 @MicahGale
Feedback on JOSS draft
## Paper general review
* [ ] [One does not tally DPA](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.md?plain=1#L47), as DPA is a unit, and not a quantity. I generally use the term: radiation damage
* In the summary great attention is paid to MC tallies, but the rest of the MC modeling process is ignored.
    * [ ] Is more context needed?
* [ ] "Dump files" Sound a lot like MCPL. Use that terminology for clarity?
* [ ] [Use of tally filter and score terminology](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.md?plain=1#L47) will make language clearer.
* [ ] Should NumPy and Pandas be cited?
* [ ] The only table lacks a caption, making it hard to parse without reading the preceding paragraph. A reference to "Table 1" rather than "below" would be clearer as well.
* [ ] Table Null has a plethora of shorthands and acronyms that are never defined, which are hard to understand
* [ ] Seems to be trying to do it all: It has: an API, a CLI, and GUI.
* [ ] DCHAIN tools does 1-group collapse. Should the math for this be included?
* [ ] SHould you mention MCNP, OpenMC, Serpent, etc. (and the various automation ecosystem for these)?
* [ ] How tally normalization is handled should be documented. Though it seems like no normalization is done by PHITS, and this should be explained as well.
* [ ] You should discuss more about the math behind tally rebinning
* [ ] You probably don't want to store `paper.md` in the main branch long term. I would suggest doing `git rm` after the review is complete, and before the next deployment.

## General editorial comments

* [ ] [This sentence has odd phrasing.](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.md?plain=1#L47) It is hard to parse with the subject at the very end of the sentence
* [ ] [Is ASCII actually used?](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.md?plain=1#L47) I am a bit skeptical of this fact, given the code is developed by JAEA. "Text file" would probably be more appropriate and avoids getting into encoding nuances.
* I appreciate a good use of em dashes.
* [ ] [I think ", very convenient"](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.md?plain=1#L47) should use a semi-colon, or be another sentence. This seems to be missing a transition like: "which is very..."
* [ ] [This list  of use cases](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.md?plain=1#L47) could be a free [standing list](https://joss.readthedocs.io/en/latest/paper.html#lists) for clarity, and I don't see why not, since there is no page count surcharge.
* [ ] [Was Material meant to be hyperlinked here?](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.md?plain=1#L47)

# Note:

These notes are only from reading the first draft of the paper.
