### 2025-08-11 371b779
all doc strings are now raw strings, fixes #30
