+++ PHITS_tools/PHITS_tools.py
@@ -307,7 +307,7 @@ if in_debug_mode:   # pragma: no cover
 
 
 def _deprecated_alias(new_func_name):
-    '''@private
+    r'''@private
     Decorator for backward-compatible aliasing of renamed functions.
     '''
     def decorator(func):
@@ -333,7 +333,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
                             save_output_pickle=True, include_phitsout_in_metadata=False, 
                             prefer_reading_existing_pickle=False, compress_pickle_with_lzma=False, 
                             autoplot_tally_output=False):
-    '''
+    r'''
     Description:
         Parse any PHITS tally output file, returning tally metadata and an array of its values (and optionally
         this data inside of a Pandas dataframe too).  Note the separate `parse_tally_dump_file` function for
@@ -765,7 +765,7 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
         phitsout_dict = parse_phitsout_file(include_phitsout_in_metadata, save_phitsout_pickle=save_output_pickle, compress_pickle_with_lzma=compress_pickle_with_lzma)
     
     def handle_dtrk_dyld_file(dtrk_dyld_filepath, tdchain_metadata={}):
-        '''
+        r'''
         Parse an axis=dchain [T-Yield] or [T-Track] file generated by [T-Dchain]
         This will also save its pickle file and generate its plot, if set to do so.
         '''
@@ -1061,7 +1061,7 @@ def parse_tally_dump_file(path_to_dump_file, dump_data_number=None , dump_data_s
                           return_Pandas_dataframe=True, save_namedtuple_list=False, save_Pandas_dataframe=False,
                           compress_pickles_with_lzma=True, prefer_reading_existing_pickle=False,
                           split_binary_dumps_over_X_GB=20, merge_split_dump_handling=0):
-    '''
+    r'''
     Description:
         Parses the dump file of a [T-Cross], [T-Product], or [T-Time] tally generated by PHITS, in ASCII or binary format.
 
@@ -1495,7 +1495,7 @@ def parse_all_tally_output_in_dir(tally_output_dirpath, output_file_suffix = Non
                                   split_binary_dumps_over_X_GB=20, merge_split_dump_handling=0,
                                   autoplot_tally_output=False, autoplot_all_tally_output_in_dir=False
                                   ):
-    '''
+    r'''
     Description:
         Parse all standard PHITS tally output files in a directory *[DIRECTORY mode]*, returning either a list of dictionaries containing
         tally metadata and an array of values from each tally output (and optionally this data inside of a Pandas dataframe too)
@@ -2004,7 +2004,7 @@ def parse_all_tally_output_in_dir(tally_output_dirpath, output_file_suffix = Non
 
 
 def parse_phitsout_file(phitsout_filepath, include_input_echo=True, save_phitsout_pickle=False, compress_pickle_with_lzma=False):
-    '''
+    r'''
     Description:
         Extracts useful information from the "phits.out" file (`file(6)` in the PHITS [Parameters] section) 
         into a dictionary object, including the PHITS version number, time data (start/stop time, CPU time, MPI info, etc.), 
@@ -2286,7 +2286,7 @@ def parse_phitsout_file(phitsout_filepath, include_input_echo=True, save_phitsou
 
 
[…]
