### 2025-08-12 Lindt8/PHITS-Tools#22 @Lindt8
Ah, the indexing generating helper function is a nice idea.  The indices are indeed a bit unwieldy with the 10-D array.  At the moment, I just try to ease this as much as possible via the documentation for the function returning this array, [`parse_tally_output_file()`](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.parse_tally_output_file), providing for easy copy/paste `tally_data[ ir, iy, iz, ie, it, ia, il, ip, ic, ierr ]` for use in scripting, where it'd be on the user to just set these index variables as needed.  A dedicated helper function for this in addition is a good idea for future development.

In the meantime, this same function also returns an equivalent Pandas DataFrame containing all the same numerical information as the array (it's built from the array after all) with well-labeled columns, and this is the more user-friendly option at present, at least for those fond of Pandas.  FWIW, one of my research collaborators using PHITS Tools has informed me that this has been his preferred approach for working with output.  (I, on the other hand, am a bit more old-fashioned with my preference for arrays but absolutely acknowledge Pandas' utility and widespread adoption, hence its support for both standard and dump outputs from PHITS.)
