### 2025-08-12 3418613
added detailed method info for rebinner function, resolves #23
