### 2025-08-14 b310d1e
updated example with custom tally over dump output
