### 2025-08-14 6f0fdc7
add elemental mass IUPAC source and reformat data to dict, resolves #26
