+++ PHITS_tools/PHITS_tools.py
@@ -4499,30 +4499,151 @@ def element_Z_or_symbol_to_name(Z):
 def element_Z_or_symbol_to_mass(Z):
     r'''
     Description:
-        Returns an element's average atomic mass provided its atomic number Z or elemental symbol
+        Returns an element's average atomic mass (standard atomic weight) provided its atomic number Z or elemental symbol
 
     Inputs:
         - `Z` = string of elemental symbol or atomic number Z
 
     Outputs:
-        - `A_avg` = average atomic mass
+        - `A_avg` = average atomic mass (standard atomic weight)
+    
+    Source:
+        [ATOMIC WEIGHTS OF THE ELEMENTS 2023, IUPAC Commission on Isotopic Abundances and Atomic Weights](https://iupac.qmul.ac.uk/AtWt/), Table 2, accessed on 14-Aug-2025.
+    
+    Notes:
+        This function is really just for quick access to average elemental mass/weight information without adding any
+        extra external dependencies to PHITS Tools.  If you wish to work with elemental/isotopic data in more detail, 
+        there are dedicated Python packages for this such as [`mendeleev`](https://github.com/lmmentel/mendeleev). 
+        
     '''
+    
+    atomic_weights_dict = {
+        0: 1.008664, 
+        1: 1.008,
+        2: 4.002602,
+        3: 6.94,
+        4: 9.0121831,
+        5: 10.81,
+        6: 12.011,
+        7: 14.007,
+        8: 15.999,
+        9: 18.998403162,
+        10: 20.1797,
+        11: 22.98976928,
+        12: 24.305,
+        13: 26.9815384,
+        14: 28.085,
+        15: 30.973761998,
+        16: 32.06,
+        17: 35.45,
+        18: 39.95,
+        19: 39.0983,
+        20: 40.078,
+        21: 44.955907,
+        22: 47.867,
+        23: 50.9415,
+        24: 51.9961,
+        25: 54.938043,
+        26: 55.845,
+        27: 58.933194,
+        28: 58.6934,
+        29: 63.546,
+        30: 65.38,
+        31: 69.723,
+        32: 72.63,
+        33: 74.921595,
+        34: 78.971,
+        35: 79.904,
+        36: 83.798,
+        37: 85.4678,
+        38: 87.62,
+        39: 88.905838,
+        40: 91.222,
+        41: 92.90637,
+        42: 95.95,
+        43: 97.0,
+        44: 101.07,
+        45: 102.90549,
+        46: 106.42,
+        47: 107.8682,
+        48: 112.414,
+        49: 114.818,
+        50: 118.71,
+        51: 121.76,
+        52: 127.6,
+        53: 126.90447,
+        54: 131.293,
+        55: 132.90545196,
+        56: 137.327,
+        57: 138.90547,
+        58: 140.116,
+        59: 140.90766,
+        60: 144.242,
+        61: 145.0,
+        62: 150.36,
+        63: 151.964,
+        64: 157.249,
+        65: 158.925354,
+        66: 162.5,
+        67: 164.930329,
+        68: 167.259,
+        69: 168.934219,
+        70: 173.045,
+        71: 174.96669,
+        72: 178.486,
+        73: 180.94788,
+        74: 183.84,
+        75: 186.207,
+        76: 190.23,
+        77: 192.217,
+        78: 195.084,
+        79: 196.96657,
+        80: 200.592,
+        81: 204.38,
+        82: 207.2,
+        83: 208.9804,
+        84: 209.0,
+        85: 210.0,
+        86: 222.0,
+        87: 223.0,
+        88: 226.0,
+        89: 227.0,
+        90: 232.0377,
+        91: 231.03588,
[…]
