### 2025-08-16 Lindt8/PHITS-Tools#22 @MicahGale
(I would use Pandas as well).
