### 2025-08-18 4752b30
address JOSS review, add refs
