+++ paper.bib
@@ -64,3 +64,53 @@
   place       = {United States},
   year        = {2011},
 }
+
+@Article{numpy_ref,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@software{pandas_ref,
+    author       = {The pandas development team},
+    title        = {pandas-dev/pandas: Pandas},
+    month        = feb,
+    year         = 2020,
+    publisher    = {Zenodo},
+    version      = {latest},
+    doi          = {10.5281/zenodo.3509134},
+    url          = {https://doi.org/10.5281/zenodo.3509134}
+}
+
+@article{MCPL_ref,
+title = {Monte Carlo Particle Lists: MCPL},
+journal = {Computer Physics Communications},
+volume = {218},
+pages = {17-42},
+year = {2017},
+issn = {0010-4655},
+doi = {https://doi.org/10.1016/j.cpc.2017.04.012},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465517301261},
+author = {T. Kittelmann and E. Klinkby and E.B. Knudsen and P. Willendrup and X.X. Cai and K. Kanaki},
+keywords = {MCPL, Monte Carlo simulations, Particle storage, File format, Geant4, MCNP, McStas, McXtrace},
+abstract = {A binary format with lists of particle state information, for interchanging particles between various Monte Carlo simulation applications, is presented. Portable C code for file manipulation is made available to the scientific community, along with converters and plugins for several popular simulation packages.
+Program summary
+Program Title: MCPL Program Files doi: http://dx.doi.org/10.17632/cby92vsv5g.1 Licensing provisions: CC0 for core MCPL, see LICENSE file for details. Programming language: C and C++ External routines/libraries: Geant4, MCNP, McStas, McXtrace Nature of problem: Saving particle states in Monte Carlo simulations, for interchange between simulation packages or for reuse within a single package. Solution method: Binary interchange format with associated code written in portable C along with tools and interfaces for relevant simulation packages.}
+}
