+++ paper.md
@@ -11,14 +11,11 @@ tags:
 authors:
   - name: Hunter N. Ratliff
     orcid: 0000-0003-3761-5415
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+    affiliation: '1'
 affiliations:
   - name: Western Norway University of Applied Sciences, Inndalsveien 28, 5063 Bergen, Norway
     index: 1
     ror: 05phns765
-  - name: Japan Atomic Energy Agency, 2-4 Shirakata, Tokai, Naka, Ibaraki 319-1195, Japan
-    index: 2
-    ror: 05nf86y53
 date: 2 January 2025
 bibliography: paper.bib
 
@@ -49,17 +46,17 @@ distributed with and coupled to PHITS is the DCHAIN [@DCHAIN_ref] code[^2].
 [^1]: For current PHITS userbase statistics, see: [https://phits.jaea.go.jp/usermap/PHITS_map_userbase.html](https://phits.jaea.go.jp/usermap/PHITS_map_userbase.html)
 [^2]: PHITS and DCHAIN are distributed by the Japan Atomic Energy Agency and the OECD/NEA Data Bank. For more information, see: [https://phits.jaea.go.jp/howtoget.html](https://phits.jaea.go.jp/howtoget.html).  
 
-Within PHITS are "tallies" which score various physical quantities
+Within PHITS are "tallies" which filter and score various physical quantities
 such as the number of particles passing through a region in space or crossing a surface, 
 the frequency and products of nuclear interactions of various types, deposition of energy/dose, 
-timing of interactions, displacements per atom (DPA), and more.  Users provide the desired
+timing of interactions, radiation damage (in DPA), and more.  Users provide the desired
 binning for the tally histograms to be created (such as specifying a range of energies of interest and 
 how many bins the tally should have within that energy range), and the code will simulate the 
 histories, or "lives", of many particles, outputting the aggregate distributions for the 
 quantities being tallied, which should be converged to the "true" or "real" distributions 
 provided a statistically sufficient number of histories were simulated (often on the order of millions or more).
 For a few tallies, PHITS provides the option to output "dump" files where, in addition to the histograms, 
-detailed raw history-by-history event data are recorded to an ASCII or binary file for every history 
+detailed raw history-by-history event data are recorded to a text or binary file for every history 
 satisfying the tally's criteria and being scored by it, allowing users to, in post, create 
 even more complex tallies and analyses than possible with the stock tallies in PHITS.
 
@@ -77,33 +74,40 @@ may wish to perform on simulation outputs.
 
 # Statement of need
 
-`PHITS Tools` and its `DCHAIN Tools` submodule serve as an interface between the plaintext (and binary) outputs
+PHITS Tools and its DCHAIN Tools submodule serve as an interface between the plaintext (and binary) outputs
 of the PHITS and DCHAIN codes and Python&mdash;greatly expediting further programmatic analyses, 
 comparisons, and visualization&mdash;and provide some extra analysis tools. 
 The outputs of the PHITS code are, aside from the special binary "dump" files, plaintext files formatted 
 for processing by a custom visualization code (generating Encapsulated PostScript files) 
 shipped with and automatically ran by PHITS, and those of the DCHAIN 
-code are formatted in a variety of tabular, human-readable structures.  Historically, programmatic
-extraction and organization of numerical results and metadata from both codes often required 
+code are formatted in a variety of tabular, human-readable structures.  
+
+Historically, programmatic extraction and organization of numerical results and 
+metadata from both codes often required 
 writing a bespoke processing script for most individual simulations, 
 possibly preceded by manual data extraction/isolation too. 
[…]
