### 2025-08-18 ca63b87
fix table and bib formatting
