### 2025-08-18 9a1fb38
fix inline citation syntax
