+++ paper.md
@@ -105,7 +105,7 @@ containing all numerical results (structure illustrated in Table \ref{tally_outp
 a Pandas [@pandas_ref] DataFrame containing the same numerical information for users 
 preferring working with Pandas.
 
-: Structure of parsed tally output from PHITS \label{tally_output_struct}
+: Structure of parsed tally output from PHITS []{label="tally_output_struct"}
 
 +------------+---------------------------------------------------------------------------------------+
 | axis       | description (using PHITS nomenclature, input syntax in `monospace` font)              |
