### 2025-08-18 02a5e54
test manual table citation
