### 2025-08-18 65adccd
undo table changes for debugging
