### 2025-08-18 4f82234
reintroduce table change, fix Pandas ref
