+++ paper.bib
@@ -79,8 +79,8 @@ keywords = {MCPL, Monte Carlo simulations, Particle storage, File format, Geant4
 }
 
 @software{pandas_ref,
-    author       = {The pandas development team},
-    title        = {pandas-dev/pandas: Pandas},
+    author       = {{The pandas development team}},
+    title        = {{pandas-dev/pandas: Pandas}},
     month        = feb,
     year         = 2020,
     publisher    = {Zenodo},
