### 2025-08-18 de215f7
fix hyperlink with escapes
