### 2025-08-18 c959845
attempt to fix table hyperlink to docs
