### 2025-08-18 4f1f289
table url debugging
