### 2025-08-18 87afd6b
more table url debugging
