### 2025-08-18 143794d
more table footer debugging
