### 2025-08-18 dcdbd6f
table footer debugging
