### 2025-08-18 b093c84
attempt table fix with asterisk escape
