### 2025-08-18 a1752fe
try just reformatting table width
