+++ paper.md
@@ -112,31 +112,31 @@ preferring working with Pandas.
 -->
 Table 1: Structure of parsed tally output from PHITS
 
-+------------+---------------------------------------------------------------------------------------+
-| axis       | description (using PHITS nomenclature, input syntax in `monospace` font)              |
-+:===========+:======================================================================================+
-| 0 / `ir`   | Geometry mesh: `reg` / `x` / `r` / `tet` *                                            |
-+------------+---------------------------------------------------------------------------------------+
-| 1 / `iy`   | Geometry mesh: `1` / `y` / `1`                                                        |
-+------------+---------------------------------------------------------------------------------------+
-| 2 / `iz`   | Geometry mesh: `1` / `z` / `z` *                                                      |
-+------------+---------------------------------------------------------------------------------------+
-| 3 / `ie`   | Energy mesh: `eng` ([T-Deposit2] `eng1`)                                              |
-+------------+---------------------------------------------------------------------------------------+
-| 4 / `it`   | Time mesh                                                                             |
-+------------+---------------------------------------------------------------------------------------+
-| 5 / `ia`   | Angle mesh                                                                            |
-+------------+---------------------------------------------------------------------------------------+
-| 6 / `il`   | LET mesh                                                                              |
-+------------+---------------------------------------------------------------------------------------+
-| 7 / `ip`   | Particle type / group (`part =`)                                                      |
-+------------+---------------------------------------------------------------------------------------+
-| 8 / `ic`   | Special: [T-Deposit2] `eng2`, [T-Yield] `mass`/`charge`/`chart`, [T-Interact] `act`   |
-+------------+---------------------------------------------------------------------------------------+
-| 9 / `ierr` | `= 0/1/2`, Value / relative uncertainty / absolute uncertainty \*                     |
-+============+=======================================================================================+
-| \*exceptional behavior with [T-Cross] tally when `enclos = 0` is set; see [full documentation](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.parse_tally_output_file)        |
-+============+=======================================================================================+
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| axis       | description (using PHITS nomenclature, input syntax in `monospace` font)                                                                                     |
++:===========+:=============================================================================================================================================================+
+| 0 / `ir`   | Geometry mesh: `reg` / `x` / `r` / `tet` *                                                                                                                   |
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| 1 / `iy`   | Geometry mesh: `1` / `y` / `1`                                                                                                                               |
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| 2 / `iz`   | Geometry mesh: `1` / `z` / `z` *                                                                                                                             |
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| 3 / `ie`   | Energy mesh: `eng` ([T-Deposit2] `eng1`)                                                                                                                     |
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| 4 / `it`   | Time mesh                                                                                                                                                    |
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| 5 / `ia`   | Angle mesh                                                                                                                                                   |
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| 6 / `il`   | LET mesh                                                                                                                                                     |
++------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| 7 / `ip`   | Particle type / group (`part =`)                                                                                                                             |
[…]
