### 2025-08-18 ad46cf1
reintroduce table caption
