+++ paper.md
@@ -101,7 +101,7 @@ particle species, and fair amount of exceptions/edge cases related to specific t
 highlight the utility of such a universal processing code for PHITS. 
 When parsing standard PHITS tally output, PHITS Tools returns a metadata dictionary, 
 a 10-dimensional NumPy [@numpy_ref] array universally accommodating of all possible PHITS tally output
-containing all numerical results (structure illustrated in Table 1), and 
+containing all numerical results (structure illustrated in Table \ref{tally_output_struct}), and 
 a Pandas [@pandas_ref] DataFrame containing the same numerical information for users 
 preferring working with Pandas.
 
@@ -110,33 +110,34 @@ preferring working with Pandas.
 [full documentation](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.parse_tally_output_file)
 
 -->
-Table 1: Structure of parsed tally output from PHITS
 
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| axis       | description (using PHITS nomenclature, input syntax in `monospace` font)                                                                                     |
-+:===========+:=============================================================================================================================================================+
-| 0 / `ir`   | Geometry mesh: `reg` / `x` / `r` / `tet` *                                                                                                                   |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 1 / `iy`   | Geometry mesh: `1` / `y` / `1`                                                                                                                               |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 2 / `iz`   | Geometry mesh: `1` / `z` / `z` *                                                                                                                             |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 3 / `ie`   | Energy mesh: `eng` ([T-Deposit2] `eng1`)                                                                                                                     |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 4 / `it`   | Time mesh                                                                                                                                                    |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 5 / `ia`   | Angle mesh                                                                                                                                                   |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 6 / `il`   | LET mesh                                                                                                                                                     |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 7 / `ip`   | Particle type / group (`part =`)                                                                                                                             |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 8 / `ic`   | Special: [T-Deposit2] `eng2`, [T-Yield] `mass`/`charge`/`chart`, [T-Interact] `act`                                                                          |
-+------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
-| 9 / `ierr` | `= 0/1/2`, Value / relative uncertainty / absolute uncertainty *                                                                                             |
-+============+==============================================================================================================================================================+
-| *exceptional behavior with [T-Cross] tally when `enclos = 0` is set; see [full documentation](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.parse_tally_output_file)  |
-+============+==============================================================================================================================================================+
+: Structure of parsed tally output from PHITS []{label="tally_output_struct"}
+
++--------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
+| axis         | description (using PHITS nomenclature, input syntax in `monospace` font)                                                                                     |
++:=============+:=============================================================================================================================================================+
+| 0 / `ir`     | Geometry mesh: `reg` / `x` / `r` / `tet` *                                                                                                                   |
++--------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
[…]
