### 2025-08-18 94c79a2
debug table caption ref
