### 2025-08-18 6694be3
try table ref in tex syntax
