### 2025-08-18 130b1b6
remove redundant "Table" word
