+++ paper.md
@@ -101,7 +101,7 @@ particle species, and fair amount of exceptions/edge cases related to specific t
 highlight the utility of such a universal processing code for PHITS. 
 When parsing standard PHITS tally output, PHITS Tools returns a metadata dictionary, 
 a 10-dimensional NumPy [@numpy_ref] array universally accommodating of all possible PHITS tally output
-containing all numerical results (structure illustrated in Table \autoref{tally_output_struct}), and 
+containing all numerical results (structure illustrated in \autoref{tally_output_struct}), and 
 a Pandas [@pandas_ref] DataFrame containing the same numerical information for users 
 preferring working with Pandas.
 
@@ -112,7 +112,6 @@ preferring working with Pandas.
 -->
 
 : Structure of parsed tally output from PHITS \label{tally_output_struct}
-
 +-------------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------+
 | axis              | description (using PHITS nomenclature, input syntax in `monospace` font)                                                                                     |
 +:==================+:=============================================================================================================================================================+
