### 2025-08-18 c17afab
Update paper.md
