### 2025-08-18 16c906e
expand acronyms
