### 2025-08-18 0b2965c
fix enumerated list formatting
