### 2025-08-18 34a0286
add short summary of radiation transport
