### 2025-08-18 6a57906
Update paper.md
