### 2025-08-18 7cb1581
Update paper.md
