### 2025-08-18 c20b10c
add FLAIR PHITS integration
