+++ paper.bib
@@ -103,4 +103,24 @@ keywords = {MCPL, Monte Carlo simulations, Particle storage, File format, Geant4
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
+@article{Flair3_ref,
+	author = {{Donadon, André} and {Hugo, Gabrielle} and {Theis, Christian} and {Vlachoudis, Vasilis}},
+	title = {FLAIR3 – recasting simulation experiences with the Advanced Interface for FLUKA and other Monte Carlo codes},
+	DOI= "10.1051/epjconf/202430211005",
+	url= "https://doi.org/10.1051/epjconf/202430211005",
+	journal = {EPJ Web Conf.},
+	year = 2024,
+	volume = 302,
+	pages = "11005",
+}
 
+@article{Fluka2024_ref,
+	author = {{The FLUKA Collaboration} and {Ballarini, Francesca} and {Batkov, Konstantin} and {Battistoni, Giuseppe} and {Bisogni, Maria Giuseppina} and {Böhlen, Till T.} and {Campanella, Mauro} and {Carante, Mario P.} and {Chen, Daiyuan} and {De Gregorio, Angelica} and {Degtiarenko, Pavel V.} and {De la Torre Luque, Pedro} and {dos Santos Augusto, Ricardo} and {Engel, Ralph} and {Fassò, Alberto} and {Fedynitch, Anatoli} and {Ferrari, Alfredo} and {Ferrari, Anna} and {Franciosini, Gaia} and {Kraan, Aafke Christine} and {Lascaud, Julie} and {Li, Wenxin} and {Liu, Juntao} and {Liu, Zhiyi} and {Magro, Giuseppe} and {Mairani, Andrea} and {Mattei, Ilaria} and {Mazziotta, Mario N.} and {Morone, Maria C.} and {Müller, Stefan E.} and {Muraro, Silvia} and {Ortega, Pablo G.} and {Parodi, Katia} and {Patera, Vincenzo} and {Pinsky, Lawrence S.} and {Ramos, Ricardo L.} and {Ranft, Johannes} and {Rosso, Valeria} and {Sala, Paola R.} and {Santana Leitner, Mario} and {Sportelli, Giancarlo} and {Tessonnier, Thomas} and {Ytre-Hauge, Kristian S.} and {Zana, Lorenzo}},
+	title = {The FLUKA code: Overview and new developments},
+	DOI= "10.1051/epjn/2024015",
+	url= "https://doi.org/10.1051/epjn/2024015",
+	journal = {EPJ Nuclear Sci. Technol.},
+	year = 2024,
+	volume = 10,
+	pages = "16",
+}
