### 2025-08-18 0f2c369
attempt to shorten some sentences
