### 2025-08-18 85da870
further sentence shortening efforts
