### 2025-08-18 0aa6b95
Update paper.md
