### 2025-08-18 fb78ef9
Update paper.md
