### 2025-08-18 ec659f5
Update paper.md
