### 2025-08-18 7e3c369
Update paper.md
