### 2025-08-18 a602b9e
Update paper.md
