### 2025-08-18 4705f1e
add GUI details
