### 2025-08-18 e1f5c65
Update paper.md
