### 2025-08-18 5c1a35c
Update paper.md
