### 2025-08-20 #8311 @Lindt8
@junghans Could you provide some clarification on how this word count limit is evaluated / what number is used?

Earlier on in this thread, the bot quoted the paper at 1910 words.  After addressing review feedback in Lindt8/PHITS-Tools/issues/31 with a careful effort on not adding much length, the paper has gotten just a bit longer, then calculated as being **1969 words**.  However, I realized I had a bunch of HTML commented lines `` from the JOSS paper.md template still in there.  After deleting them, now the bot assesses it at **1621 words**.

From what I can tell, **the bot is doing an extremely lazy word count check**.

If I just paste the raw markdown content of the current paper.md file into Microsoft Word, Word tells me the paper's length is 1651 words, close enough to the bot's estimate to see this does seem to be the case.  Only deleting the YAML-ish formatted header (with title, tags, author, affiliation, date, bibliography filename, etc.) takes this down to 1575 words, already well below the bot's current estimate.  Iterating on this pruning exercise:

- raw `paper.md` contents: 1651 words
- without YAML header: 1575 words (-76)
- without Acknowledgements: 1535 words (-40)
- without the table (which could be a figure): 1331 words (-204!)
- without section headers: 1323 words (-8)
- without footnotes: 1293 words (-30)
- without in-text citations: 1285 words (-8)

The table seems to be a _huge_ source of word count inflation here; I don't know how it's arriving at that figure. (I could make the table into a SVG/PDF to confuse it less perhaps.)

In any case, the main content of the paper, without the YAML header and table inflation is still in the ballpark of 1300 words, which is above the 250–1000 word range.  I can work on making it shorter, but if it is to be under 1000 words by however the bot appears to be scoring, this would be a major rewrite of the paper.

I suppose this is up to your interpretation of the [JOSS paper guidelines](https://joss.readthedocs.io/en/latest/paper.html) stating **_"Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper."_**, particularly with your judgement of the bot's counting approach and "significantly longer than 1000 words" here.

In any case, I am happy to make any adjustments necessary to comply with length criteria; however, I would request information on how this is evaluated so I know what target to aim for in my word count reduction efforts.
