### 2025-08-20 #8311 @junghans
I think ~1.2k is fine, but let's wait for @ftessier review.
