### 2025-09-11 56063e0
add tally_data_indices helper function
