+++ PHITS_tools/PHITS_tools.py
@@ -69,6 +69,7 @@ functions return the data objects they produce for your own further analyses.
 
 ### General Purpose Functions
 
+- `tally_data_indices`              : helper function for generating indexing tuple for use with the `tally_data` 10-D NumPy array
 - `tally`                           : tally/histogram values (and their indices) falling within a desired binning structure (useful with "dump" files)
 - `rebinner`                        : rebin a set of y-data to a new x-binning structure (edges need not necessarily be preserved)
 - `autoplot_tally_results`          : make plot(s), saved as PDFs, of tally results from tally output Pandas DataFrame(s)
@@ -448,6 +449,16 @@ def parse_tally_output_file(tally_output_filepath, make_PandasDF = True, calcula
        automatically when parsing the tally output file.  Thus, for very simple tallies, most of these indices will be
        set to 0 when accessing tally results, e.g. `tally_data[2,0,0,:,0,0,0,:,0,:]` to access the full energy spectrum
        in the third region for all scored particles / particle groups with the values and uncertainties.
+       
+       The `tally_data_indices()` function is also available to help with array access since it can be a bit cumbersome.
+       With this function, you could instead simply use `tally_data[tally_data_indices(ir=2)]`, presuming a 
+       `reg` geometry mesh and no time or angle meshes and not a tally with special axes. 
+       To be completely explicit in matching the nominal syntax, one could instead use to the same end: 
+        
+       `tally_data[tally_data_indices(default_to_all=False, ir=2, ie="all", ip="all", ierr="all")]` 
+       
+       Also note that `tally_data_indices()` allows specification of regions and particles by value/name rather than
+       indices alone, which can be quite handy.
        
        -----
        
@@ -2283,6 +2294,227 @@ def parse_phitsout_file(phitsout_filepath, include_input_echo=True, save_phitsou
                 pickle.dump(phitsout_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)
         print('\t\tPickle file written:', pickle_path)
     return phitsout_dict
+
+def tally_data_indices(*, default_to_all=True, tally_metadata=None, **axes):
+    r'''
+    Description:
+        This function is used for generating indexing tuples for the `tally_data` 10-D Numpy array outputted from 
+        `parse_tally_output_file()`, allowing selection of axes and slices more easily than keeping track of 10 
+        different indices yourself.  For reference, the `tally_data` array is structured and nominally accessed as 
+            
+       `tally_data[ ir, iy, iz, ie, it, ia, il, ip, ic, ierr ]`, with indices explained below:
+
+       Tally data indices and corresponding mesh/axis:
+
+        - `0` | `ir`, Geometry mesh: `reg` / `x` / `r` / `tet` ([T-Cross] `ir surf` if `mesh=r-z` with `enclos=0`)
+        - `1` | `iy`, Geometry mesh:  `1` / `y` / `1`
+        - `2` | `iz`, Geometry mesh:  `1` / `z` / `z` ([T-Cross] `iz surf` if `mesh=xyz` or `mesh=r-z` with `enclos=0`)
+        - `3` | `ie`, Energy mesh: `eng` ([T-Deposit2] `eng1`)
+        - `4` | `it`, Time mesh
+        - `5` | `ia`, Angle mesh
+        - `6` | `il`, LET mesh
+        - `7` | `ip`, Particle type (`part = `)
+        - `8` | `ic`, Special: [T-Deposit2] `eng2`; [T-Yield] `mass`, `charge`, `chart`; [T-Interact] `act`
+        - `9` | `ierr = 0/1/2`, Value / relative uncertainty / absolute uncertainty (expanded to `3/4/5`, or `2/3` if
[…]
