### 2025-09-12 8a436ae
add index helper tests, refine docstring
