+++ tests/test_tally_and_rebinner.py
@@ -7,7 +7,91 @@ def test_tally_data_indices():
     tally_metadata = {'reg_groups': ['1', '2', '16', '50', '51', '99'], 
                       'part_groups': ['all', 'neutron', 'proton', '-(neutron proton)']
                       }
-    assert tally_data_indices(default_to_all=False, ir=2, ie=":", ip=":", ierr=":") == (2,0,0,slice(None),0,0,0,slice(None),0,slice(None))
+    assert tally_data_indices(default_to_all=False, ir=2, ie=":", ip="all", ierr=None) == (2,0,0,slice(None),0,0,0,slice(None),0,slice(None))
+    assert tally_data_indices(default_to_all=False, tally_metadata=tally_metadata, reg=16, ie=":", ip=":", ierr=":") == (2,0,0,slice(None),0,0,0,slice(None),0,slice(None))
+    assert tally_data_indices(ireg=2, ie=":", ip=":", ierr=":") == (2,slice(None),slice(None),slice(None),slice(None),slice(None),slice(None),slice(None),slice(None),slice(None))
+    assert tally_data_indices(default_to_all=False, tally_metadata=tally_metadata, reg=[1,2,16], ie=":", part="proton", ierr=":") == ([0,1,2],0,0,slice(None),0,0,0,2,0,slice(None))
+
+    # Test different formats for specifying slices
+    # np.s_ slice form
+    idx = tally_data_indices(ie=np.s_[:10:2])
+    assert isinstance(idx[3], slice) and (idx[3].start, idx[3].stop, idx[3].step) == (None, 10, 2)
+    # np.s_ integer-sequence form
+    idx = tally_data_indices(ip=np.s_[[0, 2, 5]])
+    assert idx[7] == [0, 2, 5]
+    # boolean mask passthrough
+    mask = [True, False, True, False]
+    idx = tally_data_indices(ip=mask)
+    assert idx[7] is mask
+    # negative index and reverse slice
+    idx = tally_data_indices(ir=-1, ie=slice(None, None, -1))
+    assert idx[0] == -1 and isinstance(idx[3], slice) and idx[3].step == -1
+
+    # Test with synthetic tally_data array to ensure equivalence with old approach
+    shape = (6, 1, 1, 5, 1, 1, 1, 4, 1, 3)
+    idx = np.indices(shape) # Unique per-index value for easy equality checks
+    flat = np.ravel_multi_index(tuple(idx), shape).astype(np.int64)
+    tally_data = flat  # deterministic, unique per coordinate
+    # full spectrum in region index 2 for all particles, with errors
+    new_idx = tally_data_indices(default_to_all=False, ir=2, ie=":", ip=":", ierr=":")
+    assert np.array_equal(
+        tally_data[2, 0, 0, :, 0, 0, 0, :, 0, :],
+        tally_data[new_idx]
+    )
+    # region/particle by VALUE using metadata (region 16 -> ir=2, 'neutron' -> ip=1), single bin checks
+    new_idx = tally_data_indices(default_to_all=False, tally_metadata=tally_metadata, reg=16, part='neutron', ie=2, ierr=0)
+    assert np.array_equal(
+        tally_data[2, 0, 0, 2, 0, 0, 0, 1, 0, 0],
+        tally_data[new_idx]
+    )
+    # np.s_ single-axis: particles [0,2]
+    new_idx = tally_data_indices(ip=np.s_[[0, 2]])
+    assert np.array_equal(
+        tally_data[:, :, :, :, :, :, :, [0, 2], :, :],
+        tally_data[new_idx]
+    )
+    # reverse energy slice
+    new_idx = tally_data_indices(ie=slice(None, None, -1))
+    assert np.array_equal(
+        tally_data[:, :, :, ::-1, :, :, :, :, :, :],
[…]
