### 2025-09-16 354bd00
fix refs, minor grammar
