+++ paper.bib
@@ -9,7 +9,7 @@
   url       = {https://doi.org/10.1080/00223131.2023.2275736},
   journal   = {Journal of Nuclear Science and Technology},
   publisher = {Taylor \& Francis},
-  year      = {2024},
+  year      = {2023},
 }
 
 @Article{DCHAIN_ref,
@@ -56,7 +56,7 @@
 @TechReport{PNNL_materials_compendium,
   author      = {McConn, Ronald J and Gesh, Christopher J and Pagh, Richard T and Rucker, Robert A and Williams, III, Robert},
   institution = {Pacific Northwest National Lab. (PNNL), Richland, WA (United States)},
-  title       = {Compendium of Material Composition Data for Radiation Transport Modeling},
+  title       = {{Compendium of Material Composition Data for Radiation Transport Modeling}},
   doi         = {10.2172/1023125},
   url         = {https://www.osti.gov/biblio/1023125},
   annote      = {Introduction Meaningful simulations of radiation transport applications require realistic definitions of material composition and densities. When seeking that information for applications in fields such as homeland security, radiation shielding and protection, and criticality safety, researchers usually encounter a variety of materials for which elemental compositions are not readily available or densities are not defined. Publication of the Compendium of Material Composition Data for Radiation Transport Modeling, Revision 0, in 2006 was the first step toward mitigating this problem. Revision 0 of this document listed 121 materials, selected mostly from the combined personal libraries of staff at the Pacific Northwest National Laboratory (PNNL), and thus had a scope that was recognized at the time to be limited. Nevertheless, its creation did provide a well-referenced source of some unique or hard-to-define material data in a format that could be used directly in radiation transport calculations being performed at PNNL. Moreover, having a single common set of material definitions also helped to standardize at least one aspect of the various modeling efforts across the laboratory by providing separate researchers the ability to compare different model results using a common basis of materials. The authors of the 2006 compendium understood that, depending on its use and feedback, the compendium would need to be revised to correct errors or inconsistencies in the data for the original 121 materials, as well as to increase (per users suggestions) the number of materials listed. This 2010 revision of the compendium has accomplished both of those objectives. The most obvious change is the increased number of materials from 121 to 372. The not-so-obvious change is the mechanism used to produce the data listed here. The data listed in the 2006 document were compiled, evaluated, entered, and error-checked by a group of individuals essentially by hand, providing no library file or mechanism for revising the data in a consistent and traceable manner. The authors of this revision have addressed that problem by first compiling all of the information (i.e., numbers and references) for all the materials into a single database, maintained at PNNL, that was then used as the basis for this document.},
@@ -66,13 +66,13 @@
 }
 
 @article{MCPL_ref,
-title = {Monte Carlo Particle Lists: MCPL},
+title = {{Monte Carlo Particle Lists: MCPL}},
 journal = {Computer Physics Communications},
 volume = {218},
 pages = {17-42},
 year = {2017},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2017.04.012},
[…]
