### 2025-09-16 #8311 @Lindt8
Thanks @ftessier ! Good catches on the issues in the references in the PDF. I've fixed them along with your grammar fixes and phrasing improvements in https://github.com/Lindt8/PHITS-Tools/pull/36. The now-updated paper PDF on `main` can be viewed [here](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.pdf).

With these small changes, the word count has not appreciably changed. If @junghans is fine with it, I can summon the bot again to regenerate the PDF on the JOSS end.

I will very soon (hopefully tomorrow) pull all of the changes made to PHITS Tools in connection with this review, currently in the `develop` branch, into the `main` branch and publish a new release 1.6.0 (on GH and PyPI).  I've got a nearly-complete feature implementation for @MicahGale 's great suggestion in https://github.com/Lindt8/PHITS-Tools/issues/22 that I want to incorporate into the example too and pull into this release as well.  I'll add another comment here once I've published the new release version.

Thank you all for the review and feedback!
