### 2025-09-16 #8311 @ftessier
Fantastic, I think PHITS-Tools will be very useful for the PHITS community, and moreover serves as a potent example of how to approach raw data processing in other Monte Carlo simulation software!
