### 2025-09-17 4dad05d
update version number, polish docs
