+++ PHITS_tools/PHITS_tools.py
@@ -55,15 +55,16 @@ to create a shortcut to this executable and place the shortcut somewhere easily 
 The executable will be located in your Python installation's "Scripts" folder; you can find it easily with `where PHITS-Tools-GUI` (Windows), 
 `which PHITS-Tools-GUI` (macOS/Linux), or `python -c "import sysconfig; print(sysconfig.get_paths()['scripts'])"` (all platforms).
 
-The CLI and GUI options result in the parsed file's contents being saved to a pickle file, which can be reopened
-and used later in a Python script.  When using the main functions below within a Python script which has imported the PHITS_tools
+The CLI and GUI options result in the parsed file's contents being saved to a [pickle](https://docs.python.org/3/library/pickle.html) 
+file, which can be reopened and used later in a Python script. 
+When using the main functions below within a Python script which has imported the PHITS_tools
 module, you can optionally choose not to save the pickle files (if desired) and only have the tally output/dump parsing
 functions return the data objects they produce for your own further analyses.
 
 ### **Main PHITS Output Parsing Functions**
 
 - `parse_tally_output_file`         : general parser for standard output files for all PHITS tallies
-- `parse_tally_dump_file`           : parser for dump files from "dump" flag in PHITS [T-Cross], [T-Time], and [T-Track] tallies
+- `parse_tally_dump_file`           : parser for dump files from "dump" flag in PHITS [T-Cross], [T-Time], [T-Track], etc. tallies
 - `parse_all_tally_output_in_dir`   : run `parse_tally_output_file()` over all standard output files in a directory (and, optionally, `parse_tally_dump_file()` over all dump files too)
 - `parse_phitsout_file`             : creates a metadata dictionary of a PHITS run from its "phits.out" file
 
@@ -168,7 +169,7 @@ On Windows, using "phits/bin/phits.bat":
 - After the if statement (right before the `rem - Your file processing ends here` line), insert a new line with the following command:
 - `python "C:\path\locating\PHITS_Tools\PHITS_tools.py" "%%~nxF" -po -m -d -ddir -ddeg -lzma -p -pa`
 - Or, if PHITS Tools was installed with `pip`, `python "C:\path\locating\PHITS_Tools\PHITS_tools.py"` can be replaced with `phits-tools` as:
-    - `phits-tools "%%~nxF" -po -m -d -ddir -ddeg -lzma -p -pa`
+    - **`phits-tools "%%~nxF" -po -m -d -ddir -ddeg -lzma -p -pa`**
 
 On Linux/Mac, using "phits/bin/phits.sh":
 
@@ -176,7 +177,7 @@ On Linux/Mac, using "phits/bin/phits.sh":
 - On the line after the end of the if statement `fi`, add the following command:
 - `python "/path/locating/PHITS_Tools/PHITS_tools.py" $1 -po -m -d -ddir -ddeg -lzma -p -pa`
 - Or, if PHITS Tools was installed with `pip`, `python "/path/locating/PHITS_Tools/PHITS_tools.py"` can be replaced with `phits-tools` as:
-    - `phits-tools $1 -po -m -d -ddir -ddeg -lzma -p -pa`
+    - **`phits-tools $1 -po -m -d -ddir -ddeg -lzma -p -pa`**
 
 
 (Of course, if necessary, replace "`python`" with however you typically call python in your environment, e.g. `py`, `python3`, etc.)
@@ -228,12 +229,14 @@ On Windows, using "phits/dchain-sp/bin/dchain.bat":
 - Scroll down toward the bottom of the script, to the section with the line `rem - Your file processing ends here.`
[…]
