### 2025-09-17 0128654
fix manage_mc_materials docs URL
