### 2025-09-17 Lindt8/PHITS-Tools#22 @Lindt8
I've implemented a new helper function called [`tally_data_indices`](https://lindt8.github.io/PHITS-Tools/#PHITS_tools.tally_data_indices) that does exactly this, generating indexing tuples for accessing the 10-D array in an easier and much less error-prone way.  Users can specify axes by their baseline "canonical" names or with relevant aliases that may make more sense for specific tallies, and indices along an axis can be specified in a variety of ways.

At present just for the region and particle axes, users can also specify actual region numbers / particle names to the `reg`/`part` arguments instead of needing to know/look up what indices correspond to the different regions and particles scored in the tally.  For example, one could specify `tally_data[tally_data_indices(reg=16, part='neutron')]` instead of `tally_data[tally_data_indices(ireg=2, ipart=1)]` after having looked up the indices for region 16 and neutrons in their respective metadata dictionary lists. This alleviates one of the larger pain points with using the 10-D array.

It might be interesting to implement something in the future that would allow easy selection of various sub ranges along the numerical axes (e.g., specifying only returning results with energy between 2 and 5 MeV), but this would be a lot more work and is generally something less difficult to manage at present given the numerical axes are already in a natural ascending order, unlike the categorical region number and particle axes.

Thanks again for the great idea on this!
