### 2025-09-17 #8311 @Lindt8
@junghans Yes, the paper proof (both in the [repo](https://github.com/Lindt8/PHITS-Tools/blob/main/paper.pdf) and just [generated by the bot](https://github.com/openjournals/joss-reviews/issues/8311#issuecomment-3302874327)) is finalized and ready for any final pass.

By updated, I just meant that I'd like the final [JOSS page](https://joss.theoj.org/papers/ef67acccadb883867ba60dc9e018ff70) for the publication to reflect the updated title rather than the title of this review thread and the initial submission. (The title changed during the review process to more accurately reflect PHITS Tools being a single package as opposed to just being two python modules.) By code version number, I had just noted that the first post in the thread included a version number, which is now quite outdated.

I don't really know how code version and title stuff gets handled on the JOSS backend, but I just wanted to point out that the version number and the title changed since initial submission.

Looking at other JOSS publications, I don't see version numbers explicitly listed, but a Zenodo software archive seems to be linked.  Such [Zenodo archive for the specific 1.6.0 release of PHITS Tools](https://zenodo.org/records/17143210) has already been generated automatically.
