### 2025-09-19 Lindt8/PHITS-Tools#22 @MicahGale
Also I just came across the fact that numpy supports ellipses for indexing: https://numpy.org/doc/stable/user/basics.indexing.html#dimensional-indexing-tools, which could be another good tool. Thanks for implementing this.
