### 2025-09-22 #8311 @MicahGale
> [@MicahGale](https://github.com/MicahGale) Any final comments on the proof?

Looks good. I have no further comments.
