+++ paper.md
@@ -26,21 +26,14 @@ bibliography: paper.bib
 
 Various areas within the nuclear sciences&mdash;such as 
 nuclear facility design, medical physics, experimental nuclear physics, and radiation protection&mdash;rely 
-on complex codes employing nuclear data and a large variety of physics models to simulate the 
+on complex codes employing nuclear data and physics models to simulate the
 transport (and interactions) of radiation through matter to answer important questions, 
 in both research and applied contexts. 
-For example: How should a shielding wall be designed to comply with radiation safety regulations? 
-What dose will an individual receive in a particular exposure scenario?
-After how long will an irradiated radioactive sample decay enough to become safe to handle?
-How should an experiment be designed to make the most of limited time at an accelerator facility?
-
-While simplified "rule of thumb" calculations can provide crude answers in some basic scenarios,
-fully modeling a radiation scenario is often necessary to obtain a much more precise answer. 
-PHITS [@PHITS333_ref] (Particle and Heavy Ion Transport code System) 
-is one such general purpose Monte Carlo particle transport simulation code, presently with over 7400 users[^1].
+PHITS [@PHITS333_ref] (Particle and Heavy Ion Transport code System)
+is one such general purpose Monte Carlo particle transport simulation code, presently with over 7400 users worldwide[^1].
 Though PHITS can simulate a large variety of complex physics, it can only do so on the extremely short
 time scales of nuclear reactions. To calculate the creation and destruction of nuclides
-with time (activation, buildup, burnup, and decay) on any scale (seconds to centuries), 
+with time (activation, buildup, burnup, and decay) on any timescale (seconds to centuries),
 distributed with and coupled to PHITS is the DCHAIN [@DCHAIN_ref] code[^2].
 
 [^1]: For current PHITS userbase statistics, see: [https://phits.jaea.go.jp/usermap/PHITS_map_userbase.html](https://phits.jaea.go.jp/usermap/PHITS_map_userbase.html)
@@ -48,24 +41,22 @@ distributed with and coupled to PHITS is the DCHAIN [@DCHAIN_ref] code[^2].
 
 Radiation transport simulations minimally require specifying the involved 
 geometry (defined shapes, regions, and materials), radiation source terms, and "tallies" that 
-filter and score the various physical quantities of interest to be outputted. 
+filter and score the physical quantities of interest to be outputted.
 PHITS's tallies can score 
-the number of particles passing through a region in space or crossing a surface, 
-the frequency and products of nuclear interactions of various types, deposition of energy/dose, 
-timing of interactions, radiation damage (in displacements per atom, DPA), and more. 
-Users provide the desired 
-binning for the tally histograms to be created (e.g., specifying a range of energies of interest and 
-how many bins the tally will have in that energy range), and the code will simulate the 
-histories, or "lives", of many particles, outputting the aggregate distributions for the 
-quantities being tallied, which should be converged to the "true/real" distributions 
+the number of particles passing through a region or surface,
+the frequency/products/timing of nuclear interactions,
+deposition of energy/dose, radiation damage, and more.
+Users provide the desired criteria and
+binning for the tally histograms (e.g., specifying a range of energies and the number of bins spanning it),
+and the code simulates the histories, or "lives", of many particles, outputting aggregate distributions for the
+tallied quantities, which should be converged to the "true/real" distributions
 provided a statistically sufficient number of histories were simulated (often on the order of millions or more).
-For a few tallies, PHITS provides the option to output "dump" files where, in addition to the histograms, 
-detailed raw history-by-history event data are recorded to a text or binary file for every history 
-satisfying the tally's criteria and being scored by it, allowing users to, in post-processing, create
[…]
