### 2025-09-24 #8311 @Lindt8
@junghans I have shortened the paper by a bit over 200 words ([diff here](https://github.com/Lindt8/PHITS-Tools/pull/42/files)), bringing it much closer to 1000 words relative to my [previous estimate](https://github.com/openjournals/joss-reviews/issues/8311#issuecomment-3207791301), and this also ended up fixing the orphaned line situation and has it looking visually nicer too.  The largest cuts were made to the paragraphs you highlighted. Nothing substantial was cut/changed, and all reviewer requested changes are still contained in this version. Thanks for looking over the paper!

In the end, I've created a new [JOSS-specific-tagged release](https://github.com/Lindt8/PHITS-Tools/releases/tag/JOSS_final_v1.6.0) for this update, such that there is now a new Zenodo archive DOI / release that is 100% consistent with these more recent changes to just the paper. The release/Zenodo archive's title matches the paper's title, and I've added my name, ORCID, and affiliation there (now updated to include my university department too, there and in the paper).

The new Zenodo DOI is [10.5281/zenodo.17193688](https://doi.org/10.5281/zenodo.17193688).  The version number remains v1.6.0.
