### 2025-09-24 #8311 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
